% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{ab}
\alias{ab}
\title{Initial range for primary parameter values}
\usage{
ab(min, max)
}
\arguments{
\item{min}{Minimum value.}

\item{max}{Maximum value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Define primary model parameter initial values as range. For nonparametric,
this range will be absolutely respected. For parametric, the range serves
to define the mean (midpoint) and standard deviation (1/6 of the range) of the
initial parameter value distribution.
}
