% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{all_bmi}
\alias{all_bmi}
\title{CDC Pediatric and Adolescent BMI Table}
\format{
A data frame with the following 9 columns: Sex (1 = male), Agemos;
L, M, S (coefficients for calculating z-scores), P3, P5, P10, P25, P50, P75,
P85, P90, P95, P97: age and sex specific BMI percentiles
}
\usage{
all_bmi
}
\description{
Centers for Disease Control Pediatric and Adolescent BMI Table
}
\details{
Coefficients to calculate sex-specific BMI z-scores and percentiles.
Downloaded from https://www.cdc.gov/nccdphp/dnpa/growthcharts/resources/biv-cutoffs.pdf.
}
\author{
Michael Neely
}
\keyword{datasets}
