% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{badData}
\alias{badData}
\title{Pmetrics data file with errors}
\format{
\link{PM_data}
}
\usage{
badData
}
\description{
Example data set for an NPAG/IT2B run, which has been corrupted with errors.
}
\details{
Errors include missing covariate on first line for subject 1,
alphanumeric covariate for
gender, and trailing dose for subject 1.
}
\author{
Michael Neely
}
\keyword{datasets}
