% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bestdose}
\alias{bestdose}
\title{Run BestDose optimization to find optimal doses}
\usage{
bestdose(
  model_path,
  prior_path,
  past_data_path,
  target_data_path,
  time_offset,
  dose_min,
  dose_max,
  bias_weight,
  target_type,
  params,
  kind
)
}
\description{
Run BestDose optimization to find optimal doses
}
