% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Launch Model Builder app}
\usage{
build_model(data, ...)
}
\arguments{
\item{data}{Optional \link{PM_data} object which will be used to read
covariates.}

\item{...}{Currently unused}
}
\value{
Launches the shiny app.
}
\description{
Open the shiny model builder app.
}
\details{
The app will open in a separate window.
}
\author{
Michael Neely
}
