% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plot.R
\name{build_plot}
\alias{build_plot}
\title{Launch Plot Builder app}
\usage{
build_plot(x, ...)
}
\arguments{
\item{x}{Optional object to plot}

\item{\dots}{Not currently used}
}
\value{
Launches the shiny app.
}
\description{
Open the shiny plot builder app.
}
\details{
The app will open in a separate window.
}
\author{
Michael Neely
}
