% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{clear_build}
\alias{clear_build}
\title{Clear build files}
\usage{
clear_build(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

There are no build files to clear; models are compiled just-in-time at run time.
}
\author{
Michael Neely
}
\keyword{internal}
