% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{close_live_session}
\alias{close_live_session}
\title{Close a local live reporting session.}
\usage{
close_live_session(session_id)
}
\arguments{
\item{session_id}{Live session identifier.}
}
\value{
NULL, invisibly.
}
\description{
Close a local live reporting session.
}
\keyword{internal}
