% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{combination}
\alias{combination}
\title{Combination error model}
\usage{
combination(add, prop, constant = FALSE)
}
\arguments{
\item{add}{Initial value for lambda}

\item{prop}{Initial value for gamma}

\item{constant}{Estimate if \code{FALSE} (default).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a combination additive (lambda) and proportional error model
}
\details{
This function is not yet implemented.
}
