% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeFunc_methods.R
\name{compute.ploq}
\alias{compute.ploq}
\title{Compute P(y_ij<LOQ) for all observations}
\usage{
compute.ploq(npdeObject)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}
}
\value{
an object of class NpdeObject with a component ploq in the ["results"] slot
}
\description{
For each observation in the dataset, computes the probability to be below LOQ
using the simulations under the model. This is generally performed automatically.
}
\references{
K. Brendel, E. Comets, C. Laffont, C. Laveille, and F.
Mentre. Metrics for external model evaluation with an application to the
population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research},
23:2036--49, 2006.
}
\seealso{
\code{\link{npde}}, \code{\link{autonpde}}
}
\author{
Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\keyword{internal}
