% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert correlation matrix to covariance matrix}
\usage{
cor2cov(cor, sd)
}
\arguments{
\item{cor}{A correlation matrix.}

\item{sd}{A vector of standard deviations corresponding to the variables in the correlation matrix.}
}
\value{
A covariance matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Converts a correlation matrix to a covariance matrix using standard deviations.
}
\details{
Uses matrix multiplication to convert a correlation matrix to a covariance matrix.
}
\author{
Michael Neely
}
