% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{covariate}
\alias{covariate}
\title{Model covariate declaration}
\usage{
covariate(name, constant = FALSE)
}
\arguments{
\item{name}{Name of the covariate in quotation marks.}

\item{constant}{If \code{FALSE} (default), allow the covariate value to be
linearly interpolated between values. \strong{NOTE} that covariate values
are only applied at the times of doses. Values on observation rows are
ignored because a covariate value is an input, not an output. See the
\href{https://lapkb.github.io/Pmetrics/articles/data.html}{Data Objects} article
for details on this.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Declare covariates in the model that are in the data. Order in the model
should be the same as the order in the data.
}
