% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{cycle.2}
\alias{cycle.2}
\title{Example IT2B Output}
\format{
An R data frame made by \code{\link{makeCycle}} at the end of a successful IT2B run.
}
\usage{
cycle.2
}
\description{
Example output from an IT2B run.
}
\details{
The run consisted of a model with an absorptive compartment and a central compartment.
There were 4 parameters in the model: lag time of absorption (Tlag1),
rate constant of absorption (Ka), volume (V) and rate constant of elmination (Ke).
There were 20 subjects in the dataset.  The run was
100 cycles long and did not converge.
}
\author{
Michael Neely
}
\keyword{datasets}
