% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{dataEx}
\alias{dataEx}
\title{Pmetrics data file}
\format{
\link{PM_data}
}
\usage{
dataEx
}
\description{
Example data set for an NPAG/IT2B run.
}
\details{
Data are kindly supplied by Chuck Peloquin, PharmD. They consist of multiple
rifapentine oral doses followed by 6-7 concentrations in 20 adult subjects.
Covariates include weight (kg), africa (origin), age (years),
gender (1 = male), and height (cm).
}
\author{
Michael Neely
}
\keyword{datasets}
