% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeFunc_plots.R
\name{default.npde.plots}
\alias{default.npde.plots}
\title{Default plots for a NpdeObject object}
\usage{
default.npde.plots(npdeObject, ...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}
}
\description{
Default plots for a NpdeObject object
}
\keyword{internal}
\keyword{plot}
