% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dummy_compile}
\alias{dummy_compile}
\title{Dummy function to cache compilation artifacts.}
\usage{
dummy_compile(template_path)
}
\arguments{
\item{template_path}{Path to the template directory.}
}
\value{
Path to the build directory.
}
\description{
Dummy function to cache compilation artifacts.
}
