% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{errorPoly}
\alias{errorPoly}
\title{Assay error coefficients}
\usage{
errorPoly(coeffs, constant = FALSE)
}
\arguments{
\item{coeffs}{Vector of up to four values for C0, C1, C2, C3,
e.g. \code{c(0.15, 0.1, 0, 0)}}

\item{constant}{If \code{FALSE} (default), use values in data first, but if
missing, use values in model. If \code{TRUE}, use values in model regardless.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Specify the coefficients for the assay error polynomial.
}
