% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERRreport.R
\name{ERRreport}
\alias{ERRreport}
\title{Summarize ERR Run}
\usage{
ERRreport(wd, icen, type)
}
\arguments{
\item{wd}{The working directory containing the ASS0001 file}

\item{icen}{Not used, but included for compatibility with other report functions}

\item{type}{Not used, but included for compatibility with other report functions}
}
\value{
Two files are placed in the \code{wd}
\item{ASS0001 }{A text file of the results}
\item{errlog }{A text file with a log of the session},
}
\description{
Generates a summary of an ERR run
}
\details{
Creates an HTML page summarizing an ERR run.  This report is generated
automatically at the end of a successful run.
}
\author{
Michael Neely
}
