% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{file_exists}
\alias{file_exists}
\title{Receives a path to a CSV file and returns true if the file exists.}
\usage{
file_exists(path)
}
\description{
Receives a path to a CSV file and returns true if the file exists.
}
