% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{fixed}
\alias{fixed}
\title{Fixed primary parameter values}
\usage{
fixed(fixed, constant = FALSE, gtz = FALSE)
}
\arguments{
\item{fixed}{The starting value for the fixed parameter.}

\item{constant}{If \code{FALSE} (default), the value for \code{fixed} will serve
as the initial estimate for a parameter with unknown mean and zero variance.
The parameter value will be updated to a final value at convergence
or when the maximum number of cycles is reached. If \code{TRUE}, the value for
\code{fixed} will remain unchanged, creating a parameter with known mean and zero
variance, i.e. a constant value in the population.}

\item{gtz}{Greater than zero. If \code{FALSE} (default), ensure parameter values
remain positive by discarding negative values. Only relevant for parametric
analyses, since lower limit of parameter values for nonparametric are strictly
controlled by the range.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Fix parameter values to be the same in the population.
}
