% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{getCov}
\alias{getCov}
\title{Extract covariate information}
\usage{
getCov(mdata)
}
\arguments{
\item{mdata}{A \link{PM_data} or \emph{PMmatrix} object}
}
\value{
A list with named items: \emph{ncov, covnames, covstart, covend}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts covariate information from a Pmetrics data objects.
}
\details{
When given a \link{PM_data} or \emph{PMmatrix} object, will return a list
with the number of covariates, their names, and the starting and
ending column numbers
}
\examples{
library(PmetricsData)
getCov(dataEx)
}
\author{
Michael Neely
}
