% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMconfig.R
\name{getFixedColNames}
\alias{getFixedColNames}
\title{Names of fixed columns}
\usage{
getFixedColNames()
}
\value{
A vector of fixed column names:
\code{c("id", "evid", "time", "dur", "dose", "addl", "ii", "input", "out", "outeq", "c0", "c1", "c2", "c3")}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Returns the names of fixed columns (non-covariate) in Pmetrics data objects.
}
\examples{
getFixedColNames()
}
\author{
Michael Neely
}
