% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMconfig.R
\name{getFixedColNum}
\alias{getFixedColNum}
\title{Number of fixed columns}
\usage{
getFixedColNum()
}
\value{
An integer with the number of fixed columns.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Returns the number of fixed columns (non-covariate) in Pmetrics data objects.
}
\examples{
\dontrun{
getFixedColNum()
}
}
\author{
Michael Neely
}
