% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{getPMdata}
\alias{getPMdata}
\title{Get Pmetrics package example data}
\usage{
getPMdata()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get the data and examples for the Pmetrics package.
}
\details{
This function installs the \strong{PmetricsData} package available on github.
The repository URL is \link{https://github.com/LAPKB/PmetricsData}. These data
are used in all Pmetrics examples.
}
