% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{growth}
\alias{growth}
\title{CDC Pediatric and Adolescent Growth Data Table}
\format{
A data frame with the following 9 columns: KNOT (integer age in months); A, B1, B2, B3 (coefficients for calculating
percentiles), SEX, AGE, PERCENTILE, and CHART (length x age, wt x age, wt x length, hc x age, or ht x age).
}
\usage{
growth
}
\description{
Centers for Disease Control Pediatric and Adolescent Growth Data Table
}
\details{
Coefficients to calculate sex-specific percentiles of length, weight and head cicumference data
in children from 0 to 18 years.  Downloaded and combined from \href{https://www.cdc.gov/growthcharts/data_tables.htm}{}.
Used with the \code{\link[=qgrowth]{qgrowth()}} function to generate height and weight percentiles for the purposes of simulation.
}
\author{
Michael Neely
}
\keyword{datasets}
