% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_cargo_installed}
\alias{is_cargo_installed}
\title{Checks if Cargo is installed on the system.}
\usage{
is_cargo_installed()
}
\value{
TRUE if Cargo is installed, FALSE otherwise.
}
\description{
Checks if Cargo is installed on the system.
}
