% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{ITex}
\alias{ITex}
\title{Example IT2B Output}
\format{
R6 \link{PM_result}
}
\usage{
ITex
}
\description{
Example output from an IT2B run.
}
\details{
This is an R6 Pmetrics \link{PM_result} object created with \link{PM_load} after an NPAG run.
The run consisted of a model with an absorptive compartment and a central compartment.
There were 4 parameters in the model: lag time of absorption (Tlag1),
rate constant of absorption (Ka), volume (V) and rate constant of elmination (Ke).
There were 20 subjects in the dataset.  The run was
100 cycles long and did not converge.
}
\author{
Michael Neely
}
\keyword{datasets}
