% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{ITload}
\alias{ITload}
\title{Load IT2B run}
\usage{
ITload(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use PM_load() instead.
}
\author{
Michael Neely
}
\keyword{internal}
