% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{locales}
\alias{locales}
\title{Pmetrics locales}
\format{
Dataframe with languages and iso693 two- and three-letter codes.
}
\usage{
locales
}
\description{
Pmetrics locales
}
\author{
Michael Neely
}
\keyword{datasets}
