% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{make_valid}
\alias{make_valid}
\title{Create a Pmetrics validation object}
\usage{
make_valid(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is replaced by \link{PM_valid}, which is
typically called with the \verb{$validate} method for a \link{PM_result} object,
but can be called directly on a \link{PM_result} object.
}
\seealso{
\link{make_valid}
}
\author{
Michael Neely
}
\keyword{internal}
