% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFinal.R
\name{makeFinal}
\alias{makeFinal}
\title{Summarize NPAG or IT2B Final Cycle Population Values}
\usage{
makeFinal(data)
}
\arguments{
\item{data}{A suitable data object of the \emph{NPAG} or \emph{IT2B} class (see \code{\link{NPparse}} or \code{\link{ITparse}}).}
}
\value{
The output of \code{makeFinal} is a list of class \emph{PMfinal}, which contains the following:
\item{popPoints }{(NPAG only) Dataframe of the final cycle joint population density of grid points
with column names equal to the name of each random parameter plus \emph{prob} for the
associated probability of that point}
\item{popMean }{The final cycle mean for each random parameter distribution}
\item{popSD }{The final cycle standard deviation for each random parameter distribution}
\item{popCV }{The final cycle coefficient of variation (SD/Mean) for each random parameter distribution}
\item{popVar }{The final cycle variance for each random parameter distribution}
\item{popCov }{The final cycle random parameter covariance matrix}
\item{popCor }{The final cycle random parameter correlation matrix}
\item{popMedian }{The final cycle median values for each random parameter}
\item{postPoints}{(NPAG only) Dataframe of posterior population points for each of the first 100 subject,
with columns id, point, parameters and probability.  The first column is the subject, the second column has the population
point number, followed by the values for the parameters in that point and the probability.}
\item{postMean }{A \emph{nsub} x \emph{npar} data frame containing
the means of the posterior distributions for each parameter.}
\item{postSD }{A \emph{nsub} x \emph{npar} data frame containing
the SDs of the posterior distributions for each parameter.}
\item{postVar }{A \emph{nsub} x \emph{npar} data frame containing
the variances of the posterior distributions for each parameter.}
\item{postCov }{NPAG only: An array of dimensions \emph{npar} x \emph{npar} x \emph{nsub} that
contains the covariances of the posterior distributions for each parameter and subject.}
\item{postCor }{NPAG only: An array of dimensions \emph{npar} x \emph{npar} x \emph{nsub} that
contains the correlations of the posterior distributions for each parameter and subject.}
\item{postMed }{A \emph{nsub} x \emph{npar} data frame containing
the medians of the posterior distributions for each parameter.}
\item{shrinkage }{A data frame with the shrinkage for each parameter.
The total population variance for a parameter
is comprised of variance(EBE) plus average variance(EBD),
where each subject's EBE is the Empirical Bayes Estimate or mean posterior value
for the parameter. EBD is the Empirical Bayes Distribution, or
the full Bayesian posterior parameter value distribution for each subject.

The typical definition of \eqn{\eta} shrinkage is
\eqn{[1 - \frac{SD(\eta)}{\omega}]} or \eqn{[1 - \frac{var(\eta)}{\omega^2}]},
where \eqn{\eta} is the EBE and \eqn{\omega^2} is the population variance of \eqn{\eta}.

In parametric modeling approaches \eqn{\eta} is the interindividual variability around
the typical (mean) value of the parameter in the population, usually referred to
as \eqn{\theta}. In nonparametric approaches, there is no assumption of normality, so
\eqn{\eta} simply becomes each subject's mean parameter value estimate.

Here is how Pmetrics derives and then calculates shrinkage for a given parameter.
\deqn{popVar = var(EBE) + mean(var(EBD))}
\deqn{1 = \frac{var(EBE)}{popVar} + \frac{mean(var(EBD)}{popVar}}
\deqn{1 - \frac{var(EBE)}{popVar} = \frac{mean(var(EBD))}{popVar}}
\deqn{shrinkage = \frac{mean(var(EBD))}{popVar}}
Shrinkage is therefore a fraction between 0 and 1. If Bayesian posterior distributions are wide
for a given parameter and \eqn{mean(var(EBD))} is high due to sparse or uninformative sampling,
then most of the population variance is due
to this variance and shrinkage is high, i.e., individual posterior estimates (EBE)
shrink towards the population mean. Be aware, however, that a Bayesian posterior
parameter value distribution for a given subject who is sparsely sampled may also
be a single support point with no variance. Therefore EBD under nonparametric
assumptions is not always large with uninformative sampling. This means that shrinkage
is not as readily interpretable in nonparametric population modeling.

An alternative is to consider the number of support points relative to the number
of subjects. Highly informed, distinct subjects will result in the maximum possible
number of support points, \emph{N}, which is the same as the number of subjects.
In contrast, badly undersampled subjects can result in only one support point.
There is no formal criterion for this statistic, but it can be used in combination
with shrinkage to assess the information content of the data.
}
\item{gridpts }{(NPAG only) Initial number of support points}
\item{nsub }{Number of subjects}
\item{ab }{Matrix of boundaries for random parameter values}
A plot method exists in \code{\link{plot}} for \emph{PMfinal} objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts final cycle information from NPAG or IT2B run.
}
\details{
This function will parse the output of \code{\link{NPparse}} or \code{\link{ITparse}} to generate a
list suitable for analysis and plotting of NPAG  or IT2B final cycle population values.
}
\examples{
final <- makeFinal(NPex$NPdata)
final
names(final)
final2 <- makeFinal(ITex$ITdata)
final2
names(final2)
}
\seealso{
\code{\link{NPparse}}, \code{\link{ITparse}},  \code{\link{plot.PMfinal}}
}
\author{
Michael Neely
}
