% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNPDE.R
\name{makeNPDE}
\alias{makeNPDE}
\title{Simulation-based model diagnostics}
\usage{
makeNPDE(run, data, outeq, nsim = 1000, ...)
}
\arguments{
\item{run}{When the current working directory is the Runs folder, the folder name of a previous run that you wish to use for the npde,
which will typically be a number, e.g. 1.}

\item{data}{An optional data file in the Runs folder to use as an external dataset for NPDE.  The population parameters
from \code{run} will be used with the subjects in \code{data}.  If missing, the original population
will be used.}

\item{outeq}{The number of the output equation to simulate/test.  Default is missing, which will test all output equations.}

\item{nsim}{The number of simulations per subject in the data file.  We recommend 1000 (the default)
to return valid npde results.  More may result in excessive simulation times.}

\item{\dots}{Other parameters to be passed to \code{\link{SIMrun}}.}
}
\value{
The output of \code{makeNPDE} is a list of class \code{PMnpde} with objects
of \code{NpdeObject} class.  Additionally, two objects with run numbers appended will be saved to the output directory of the run for subsequent
loading with \code{\link{PMload}}: npde and sim.  \emph{npde} is the PMnpde object, and \emph{sim} is a 
PMsim object of all simulations combined which can be used for visual predictive checks
(see \code{\link{plot.PMsim}}).
}
\description{
Use simulations to run model diagnostic tests.
}
\details{
This function is a Pmetrics wrapper to the \code{autonpde} function in the \code{npde}
package of Comets et al that will generate normalized prediction distribution errors.
Output from a loaded NPAG or IT2B run will be used as the population model supplied to the
simulator. The function will iterate through the .csv file, using each subject as a template to simulate \code{nsim} new individuals. 
It is HIGHLY recommended to use the default value of 1000 for \code{nsim} for the most valid  
calculation of npde.  More than this could take a long time to execute.
The mean population values will be used for each parameter and the covariance matrix.  
Errors may arise if extreme or negative concentrations are simulated
from excessively large covariance matrices.
}
\references{
Brendel K, Comets E, Laffont CM, Laveille C, Mentre F (2006) Metrics for external model evaluation with an application to the population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research}, 23:2036-49

Mentre F, Escolano S (2006) Prediction discrepancies for the evaluation of nonlinear mixed-effects models. \emph{J Pharmacokinet Pharmacodyn}, 33:345-67
}
\seealso{
\code{\link{SIMrun}}, \code{\link{autonpde}}, \code{\link{plot.PMnpde}}
}
\author{
Michael Neely
}
