% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOP.R
\name{makeOP}
\alias{makeOP}
\title{Generated observed vs. predicted data}
\usage{
makeOP(data)
}
\arguments{
\item{data}{A suitable data object of the \emph{NPAG} or \emph{IT2B} class (see \code{\link{NPparse}} or \code{\link{ITparse}}).}
}
\value{
The output of \code{makeOP} is a data frame of class \emph{PMop}, which has a population and posterior
prediction object (also class \emph{PMop}) for each output equation.  Each of these has 13 columns:
\item{id }{subject identification}
\item{time }{observation time in relative hours}
\item{obs }{observation}
\item{pred }{prediction}
\item{pred.type }{Population predictions based on Bayesian prior parameter value distribution,
or individual predictions based on Bayesian posterior parameter value distributions}
\item{icen }{Predictions based on mean or median of Bayesian \code{pred.type} parameter values}
\item{outeq }{output equation number}
\item{block }{dosing block number for each subject, as defined by dose resets (evid=4).}
\item{obsSD }{standard deviation of the observation based on the assay error polynomial}
\item{d }{prediction error, \code{pred}-\code{obs}}
\item{ds }{squared prediction error}
\item{wd }{weighted prediction error, which is the prediction error divided by the \code{obsSD}}
\item{wds }{weighted squared prediction error}
A plot method exists in \code{\link{plot}} for \emph{PMop} objects.
}
\description{
Generates an observed vs. predicted data.frame from an \emph{NPAG} or \emph{IT2B} object
}
\details{
\code{makeOP} will parse the output of \code{\link{NPparse}} or \code{\link{ITparse}} to generate a
data.frame suitable for analysis and plotting of observed vs. population or
or individual predicted outputs.
}
\examples{
data(PMex1)
op <- makeOP(NPdata.1)
op
names(op)
summary(op)
}
\seealso{
\code{\link{NPparse}}, \code{\link{ITparse}}, \code{\link{plot.PMop}}, \code{\link{summary.PMop}}
}
\author{
Michael Neely
}
