% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePop.R
\name{makePop}
\alias{makePop}
\title{Individual Bayesian population prior predictions at short intervals}
\usage{
makePop(run, NPdata)
}
\arguments{
\item{run}{The number of the folder that contains the relevant run.  If missing, \code{NPdata} will be used.}

\item{NPdata}{Optional name of NPdata object if run is missing.}
}
\value{
A dataframe of class \emph{PMpop} with columns:
\item{id}{ Subject id}
\item{time}{ Time of predictions in decimal hours}
\item{icen}{ Prediction based on mean or median of Bayesian posterior parameter distribution}
\item{pred}{ Predicted output for each outeq}
\item{outeq}{ Output equation number}
\item{block}{ Observation blocks within subjects as defined by EVID=4 dosing events}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Returns the Bayesian population prior predictions at short intervals specified during the NPAG run,
up to 12 minutes.
}
\details{
Predictions are calculated using the median (default) or mean of the
Bayesian prior parameter values and each subject's dosing, sampling, and covariate
(if relevant) information.
}
\examples{
pop <- makePop(NPdata = NPex$NPdata)
}
\author{
Michael Neely
}
