% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePTAtarg.R
\name{makePTAtarget}
\alias{makePTAtarget}
\title{Make PTA target object}
\usage{
makePTAtarget(x)
}
\arguments{
\item{x}{A data.frame or name of .csv file in working directory whose first two
columns are targets and the number of samples for each target.  An example can be
seen for Staphylococcus aureus susceptibility to vancomycin at the EUCAST website
at \link{http://mic.eucast.org/Eucast2/regShow.jsp?Id=1214}.}
}
\value{
A data frame with two columns named targets and n, of class \emph{PMpta.targ}.
}
\description{
Make a Percent Target Attainment (PTA) Target
}
\details{
\code{makePTAtarget} generates an object of class \emph{PMpta.targ} which can
be used in the \code{\link{makePTA command}} for targets sampled from a distribution.
}
\seealso{
\code{\link{makePTA}}
}
