% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValid.R
\name{makeValid}
\alias{makeValid}
\title{Create a Pmetrics validation object}
\usage{
makeValid(run, input = 1, outeq = 1, tad = F, binCov, doseC, timeC, tadC, ...)
}
\arguments{
\item{run}{When the current working directory is the Runs folder, the folder name of a previous run that you wish to use for the npde,
which will typically be a number, e.g. 1.}

\item{input}{The input number.  Default is 1.}

\item{outeq}{The number of the output equation to simulate/test.  Default is 1.}

\item{\dots}{Other parameters to be passed to \code{\link{SIMrun}}.}

\item{nsim}{The number of simulations per subject in the data file.  We recommend 1000 (the default)
to return valid npde results.  More may result in excessive simulation times.}
}
\value{
The output of \code{makeNPDE} is a list of class \code{PMnpde} with objects
of \code{NpdeObject} class.  Additionally, two objects with run numbers appended will be saved to the output directory of the run for subsequent
loading with \code{\link{PMload}}: npde and sim.  \emph{npde} is the PMnpde object, and \emph{sim} is a 
PMsim object of all simulations combined which can be used for visual predictive checks
(see \code{\link{plot.PMsim}}).
}
\description{
Creates a Pmetrics validation object
}
\details{
\code{makeValid} will create an object suitable for plotting visual predictive checks (VPCs), prediction-corrected visual
predictive checks (pcVPCs) and normalized prediction distribution errors (NPDEs). The function will guide the user
through appropriate clustering of doses, covariates and sample times for prediction correction using the methods of
}
\references{

}
\seealso{
\code{\link{SIMrun}}, \code{\link{autonpde}}, \code{\link{plot.PMnpde}}
}
\author{
Michael Neely
}
