% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{mdata.1}
\alias{mdata.1}
\title{Example NPAG Output}
\format{
The original data file as read by \code{\link{PMreadMatrix}} at the end of a successful NPAG run.
}
\usage{
mdata.1
}
\description{
Example output from an NPAG run.
}
\details{
The run consisted of a model with an absorptive compartment and a central compartment.
There were 4 parameters in the model: lag time of absorption (Tlag1),
rate constant of absorption (Ka), volume (V) and rate constant of elmination (Ke).
There were 20 subjects in the dataset.  The run was
100 cycles long and did not converge.
}
\author{
Michael Neely
}
\keyword{datasets}
