% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{mic1}
\alias{mic1}
\title{Example MIC data}
\format{
An R data frame containing example MIC distribution data in two columns:
\itemize{
\item mic Minimum inhibitory concentration
\item n Number of organisms with the given MIC
}
}
\usage{
mic1
}
\description{
Example MIC data
}
\details{
This data frame contains MIC data for vancomycin against S. aureus.  It was obtained
from the EUCAST website at \href{https://mic.eucast.org}{}.  Select the organism
or drug, and then select the desired row of the resulting table to see
a histogram (top) and table (bottom) of MIC distributions.

Copy the table into excel, save as a .csv file, and read into R using a function like
\code{\link[=read.csv]{read.csv()}} or \code{\link[readr:read_delim]{readr::read_csv()}}.  Then use \code{\link[=makePTAtarget]{makePTAtarget()}}.
}
\author{
Michael Neely
}
\keyword{datasets}
