% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{mod_lib_names}
\alias{mod_lib_names}
\title{List of Pmetrics model library model names}
\usage{
mod_lib_names()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns a list of names of all Pmetrics model library objects in the global environment.
}
