% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{model_parameters}
\alias{model_parameters}
\title{Retrieves the model parameters from the given model.}
\usage{
model_parameters(model_source)
}
\arguments{
\item{model_source}{Model definition written in the pharmsol DSL.}
}
\value{
List of model parameters.
}
\description{
Retrieves the model parameters from the given model.
}
