% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{modelLibrary}
\alias{modelLibrary}
\title{Pmetrics Model Library}
\format{
A tibble with the following columns
\itemize{
\item \strong{ncomp} The number of compartments in the model, including any bolus
depo compartment(s).
\item \strong{par} Either "K" for parameterization with rate constants, or "CL" for
clearances.
\item \strong{bolus} Does the model contain a bolus input, e.g. oral?
\item \strong{elim} List of compartment numbers with elimination
\item \strong{ode} Does the model contain differential equations?
\item \strong{mod} the \link{PM_model}
\item \strong{name} Model name in plain English
}
}
\usage{
modelLibrary
}
\description{
Model Library
}
\details{
This is a tibble containing models in the Pmetrics library.
}
\author{
Michael Neely
}
\keyword{datasets}
