% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{modEx}
\alias{modEx}
\title{Pmetrics model object}
\format{
R6 \link{PM_model}
}
\usage{
modEx
}
\description{
Example model for an NPAG/IT2B run. There are 4 parameters in the model: lag time of absorption (Tlag1),
rate constant of absorption (Ka), volume (V) and rate constant of elmination (Ke).
There are 5 covariates: weight in kg (WT), whether from Africa or not (AFRICA), age in years (AGE),
1 for male (GENDER), and height in cm (HEIGHT).  There is one output equation, and the model uses
gamma plus an error polynomial derived from the assay.
}
\author{
Michael Neely
}
\keyword{datasets}
