% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{modifyList2}
\alias{modifyList2}
\title{Modify a list with another list, allowing NULL values}
\usage{
modifyList2(x, val, keep.null = FALSE)
}
\arguments{
\item{x}{A list to be modified.}

\item{val}{A list of values to modify \code{x}.}

\item{keep.null}{A logical value indicating whether to keep NULL values in \code{val}.
Default is FALSE.}
}
\value{
A modified list, as in \code{\link[utils:modifyList]{utils::modifyList()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Version of \code{\link[utils:modifyList]{utils::modifyList()}} that works with lists which have unnamed elements.
}
