% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMoptions.R
\name{movePMoptions}
\alias{movePMoptions}
\title{Move Pmetrics User Options}
\usage{
movePMoptions()
}
\value{
NULL, invisibly.
}
\description{
Move user options file for Pmetrics
}
\details{
This function will move user options file for Pmetrics from internal to external
location so that options can persist when the Pmetrics package is updated.
}
\author{
Michael Neely
}
