% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{msd}
\alias{msd}
\title{Initial mean/SD for primary parameter values}
\usage{
msd(mean, sd, gtz = FALSE)
}
\arguments{
\item{mean}{Initial mean.}

\item{sd}{Initial standard deviation.}

\item{gtz}{Greater than zero. If \code{FALSE} (default), ensure parameter values
remain positive by discarding negative values. Only relevant for parametric
analyses, since lower limit of parameter values for nonparametric are strictly
controlled by the range.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Define primary model parameter initial values as mean and standard
deviation, which translate to a range. The mean serves as the midpoint
of the range, with 3 standard deviations above and below the mean to define
the min and max of the range. For nonparametric,
this range will be absolutely respected. For parametric,
values can be estimated beyond the range.
}
