% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeObject.R
\name{npde.main.NpdeObject}
\alias{npde.main.NpdeObject}
\title{Main npde function}
\usage{
\method{npde.main}{NpdeObject}(object)
}
\arguments{
\item{object}{A NpdeObject object}
}
\value{
A NpdeObject object updated with the results
}
\description{
Main npde function, used to compute pd and npde
}
\keyword{internal}
\keyword{model}
