% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeFunc_plots.R
\name{npde.plot.covariates}
\alias{npde.plot.covariates}
\title{Covariate plots for a NpdeObject object}
\usage{
npde.plot.covariates(npdeObject, which="x", ...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{which}{one of "x" (scatterplots of the metric versus X), "pred" (scatterplots of the metric versus predictions) or "ecdf" (empirical distribution function)}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}
}
\description{
Covariate plots for a NpdeObject object
}
\keyword{internal}
\keyword{plot}
