% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeFunc_plots.R
\name{npde.plot.data}
\alias{npde.plot.data}
\alias{npde.plot.dist}
\alias{npde.plot.scatter}
\alias{npde.plot.vpc}
\alias{npde.plot.loq}
\alias{aux.scatter.box}
\alias{aux.scatter}
\alias{compute.bands.true}
\alias{compute.bands}
\title{Plot functions for NpdeObjects}
\usage{
npde.plot.data(npdeObject, ...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}
}
\description{
Workhorse functions for plots (internal)
}
\keyword{internal}
\keyword{plot}
