% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{NPex}
\alias{NPex}
\title{Example NPAG Output with validation}
\format{
R6 \link{PM_result}
}
\usage{
NPex
}
\description{
Example output from an NPAG run with validation.
}
\details{
This is an R6 Pmetrics \link{PM_result} object created with \code{\link[=PM_load]{PM_load()}} after an NPAG run.
The run consisted of a model with an absorptive compartment and a central compartment.
There were 4 parameters in the model: lag time of absorption (Tlag1),
rate constant of absorption (Ka), volume (V) with weight as a covariate, and
rate constant of elmination (Ke).
There were 20 subjects in the dataset.  The run was
100 cycles long and did not converge. It was then validated with the \verb{$validate}
method for \link{PM_result} objects.
}
\author{
Michael Neely
}
\keyword{datasets}
