% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPparse.R
\name{NPparse}
\alias{NPparse}
\title{Parse Pmetrics NPAG Output}
\usage{
NPparse(outfile = "NP_RF0001.TXT")
}
\arguments{
\item{outfile}{This is the filename of the output from NPAG. Typically,
the file will be called NP_RF0001.txt, and this is the default.}
}
\value{
The output of \code{NPparse} is a list with the following objects and
 of the class \emph{NPAG}.
 \item{nsub }{Number of subjects}
 \item{nactve }{Number of active grid points at the final cycle}
 \item{nvar }{Number of random variables or parameters in the model}
 \item{nofix }{Number of fixed variables or parameters in the model}
 \item{par }{Names of random parameters}
 \item{parfix }{Names of fixed parameters}
 \item{covnames }{Names of covariates}
 \item{ab }{Initial boundaries for each random parameter}
 \item{valfix }{Values for fixed parameters}
 \item{ndim }{Number of differential equations in model, or 0 for only output equation, or -1 for analytic
solution (algabraic)}
 \item{indpts }{Index for the initial number of gridpoints in the model}
 \item{icycst }{Starting cycle number}
 \item{icycmax }{Maximum number of cycles specified by the user}
 \item{icyctot }{Number of cycles run.  If less than \code{icycmax}, convergence occurred.}
 \item{converge }{Boolean value if convergence occurred.}
 \item{ODEtol }{Ordindary Differential Equation solver tolerance.}
 \item{prior }{Prior density for the run, either \dQuote{UNIFORM} or the name of the user-specified density file, typically \dQuote{DEN0001}.}
 \item{ERRmod }{Assay error model: 1 for SD; 2 for SD*gamma; 3 for additive lambda model; and 4 for gamma only}
 \item{numeqt }{Number of output equations}
 \item{ndrug }{Number of drug inputs}
 \item{salt }{Vector of values of the salt fraction for each \code{ndrug}}
 \item{ndose }{Vector of the number of doses for each subject in the population}
 \item{ncov }{Number of covariates in the model}
 \item{nobs }{Vector of the number of observations for each subject in the population}
 \item{nobsmax }{Maximum number of observation in any individual subject}
 \item{numt }{Vector of the number of time points for each subject at which a prediction is generated for each \emph{numeqt} output equation}
 \item{corden }{Final cycle joint population density of parameter estimates}
 \item{postden }{Array of posterior parameter value distributions for the first 100  subjects at each observation time point.
 \emph{postden[nsub,nactvepost,density]} where \emph{nactvepost} is the posterior grid point}
 \item{pyjgx }{Matrix of posterior probability of each \emph{nactve} point for each subject, given that subject's data}
 \item{ypredpop }{Array of population model predictions for each subject at each observation time point.
 \emph{ypredpop[nsub,numeqt,time,type]} where \emph{type} is 1=mean, 2=median, 3=mode of the population prior used to calculate ypredpop}
 \item{ypredbay }{Array of Bayesian posterior model predictions for each subject at each observation time point.
 \emph{ypredbay[nsub,numeqt,time,type]} where \emph{type} is 1=mean, 2=median, 3=mode of the population prior used to calculate ypredbay}
 \item{ttpred }{Matrix of the prediction time points for each subject, with \emph{nsub} rows and max(\emph{numt}) columns}
 \item{exx }{Array of the mean, median, and mode of the posterior marginal distribution for each parameter in each subject, of the form \emph{exx[nvar,type,nsub]}}
 \item{ypredpopt }{Array of population model predictions for each subject at each
 \emph{ttpred} time point,  of the form \emph{ypredpopt[nsub,  numeqt,  time,  type]},
 where type is 1=mean,  2=median,  3=mode of the population prior used to calculate
 \emph{ypredpopt}}
 \item{ilog }{Matrix of cycle number and associated log-likelihood}
 \item{iic }{Matrix with cycle number and Akaike Information Criterion (AIC) and Bayesian Information Criterion (BIC) for each cycle}
 \item{imean }{Matrix of cycle numbers and associated means for each random parameter}
 \item{isd }{Matrix of cycle numbers and associated standard deviations for each random parameter}
 \item{iaddl }{Array of additional information for each random parameter in each cycle,
 of the form \emph{iaddl[info, nvar, cycle]},  where info is a value from 1 to 12:
 1= mode; 2= skewness; 3= kurtosis; 4-8 give percentiles of the distribution (4=2.5\%; 5=25\%;
 6=50\% [median],  7=75\%; 8=97.5\%); 9= the standard deviation of a normal distribution
 with the same interquartile range; 10=the standard deviation of a normal distribution
 with the same 95\% range; 11=the average of 9 and 10; 12=the \% scaled information}
 \item{igamlam }{Matrix of cycle number and associated gamma or lambda}
 \item{blog }{Vector of each subject's Bayesian posterior log-likelihood}
 \item{bmean }{Matrix of subject numbers and associated Bayesian posterior means for each random parameter}
 \item{bsd }{Matrix of subject numbers and associated Bayesian posterior standard deviations for each random parameter}
 \item{baddl }{Array of Bayesian posterior additional information for each random parameter
 for each subject, of the form \emph{baddl[info, nvar, nsub]},  where info is the same as for \emph{iaddl}.}
 \item{bauc }{Matrix of AUC blocks for each subject with 5 columns:
 [nsub,  numeqt,  nblock,  tau,  auc]; \emph{nsub} and \emph{numeqt} are as previously defined;
 \emph{nblock} is the AUC block as defined by successive dose reset (evid=4) events;
 \emph{tau} is the time interval for that block; \emph{auc} is the AUC for that block}
 \item{sdata }{Subject data consisting of 5 columns: [id,  nsub,  age,  sex,  ht],
 \emph{id} is the original identification number in the .csv matrix file;
 \emph{nsub} is the sequential subject number in the NPAG run; \emph{age},
 \emph{sex} and \emph{ht} will be missing for .csv input and present if included in .wrk input files}
 \item{dosecov }{Matrix with all dosing information for each subject,  including times,  routes,  amounts,  and associated covariate values}
 \item{outputs }{Matrix with measured outputs for each subject and associated assay error polynomials.
  The order of the columns is nsub, time, numeqt, observation, c0, c1, c2, c3, where the last
  four columns are the coefficients of the assay error polynomial for that observation, such that
  SD[obs] = c0 + c1*[obs] + c2*[obs]**2 + c3*[obs]**3}
 \item{negflag }{A flag indicating that some negative predictions were changed to missing.
 This means that the model may be misspecified.}
 \item{mdata }{The filename of the data used in the run.}
}
\description{
\code{NPparse} processes the output from an NPAG run into a list.
}
\details{
This function can take some time to process the RFILE, depending on the number of subjects,
 doses, observations, etc.  Typical wait times are a few seconds up to 5 minutes.
 When processing is complete a summary of the extracted data will be reported on the console.
}
\author{
Michael Neely
}
