% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPrun.R
\name{NPrun}
\alias{NPrun}
\title{Run a nonparmetric population model}
\usage{
NPrun(data, model, ...)
}
\arguments{
\item{data}{A \code{\link[=PM_data]{PM_data()}} object, the character name of a data .csv file, or an
appropriate data frame that can be coerced into a \code{\link[=PM_data]{PM_data()}} object.}

\item{model}{A \code{\link[=PM_model]{PM_model()}} object, or a list that can be coerced into one.}

\item{...}{Additional arguments passed to the \verb{$fit} method in \code{\link[=PM_model]{PM_model()}}.}
}
\value{
Unlike older versions of Pmetrics, this wrapper function will return
a \code{\link[=PM_result]{PM_result()}} directly.
}
\description{
This function runs a nonparametric population model using the provided data and model object.
}
\details{
This is a wrapper around the current method of fitting models to data in Pmetrics,
available to maintain backwards compatibility with versions of Pmetrics prior to 3.0.0.
}
\examples{
\dontrun{
run1 <- NPrun(data = dataEx, model = modEx, cycles = 10)
run1$op$plot()
}
}
\author{
Michael Neely
}
\keyword{internal}
