% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{one_comp_bolus}
\alias{one_comp_bolus}
\title{One-compartment bolus model template}
\usage{
one_comp_bolus()
}
\value{
A \code{PM_model} object.
}
\description{
Create a \code{PM_model} object for the \code{one_comp_bolus} model template.
}
