% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{one_comp_bolus_cl}
\alias{one_comp_bolus_cl}
\title{One-compartment bolus clearance model template}
\usage{
one_comp_bolus_cl()
}
\value{
A \code{PM_lib} object.
}
\description{
Create a \code{PM_lib} object for the \code{one_comp_bolus_cl} model template.
}
