% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{one_comp_iv}
\alias{one_comp_iv}
\title{One-compartment IV model template}
\usage{
one_comp_iv()
}
\value{
A \code{PM_lib} object.
}
\description{
Create a \code{PM_lib} object for the \code{one_comp_iv} model template.
}
