% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeMain.R
\name{pdemenu}
\alias{pdemenu}
\title{Interactive menu to set the options for the npde() function}
\usage{
pdemenu()
}
\value{
A list with the information needed to compute the pd/npde
}
\description{
Interactive menu to set the options for the npde() function
}
\references{
K. Brendel, E. Comets, C. Laffont, C. Laveille, and F.
Mentre. Metrics for external model evaluation with an application to the
population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research},
23:2036--49, 2006.
}
\seealso{
\code{\link{npde.graphs}}, \code{\link{gof.test}}
}
\author{
Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\keyword{internal}
