% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_bestdose.R
\name{plot.bd}
\alias{plot.bd}
\title{Plot BestDose predictions}
\usage{
\method{plot}{bd}(
  x,
  mult = 1,
  outeq = 1,
  quiet = FALSE,
  legend = TRUE,
  log = FALSE,
  grid = FALSE,
  future_region = FALSE,
  xlab = "Time",
  ylab = "Concentration",
  title = NULL,
  xlim = NULL,
  ylim = NULL,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{bd} object containing the best dose predictions.}

\item{mult}{Numeric multiplier to apply to the concentrations (default: 1, no scaling).}

\item{outeq}{Numeric value of outeq to filter observations for plotting (default: 1).}

\item{quiet}{Logical indicating whether to suppress messages (default: FALSE).}

\item{legend}{Logical indicating whether to display a legend (default: TRUE).}

\item{log}{Logical indicating whether to use a logarithmic scale for the y-axis (default: FALSE).}

\item{grid}{Logical indicating whether to display a grid (default: FALSE).}

\item{future_region}{Logical indicating whether to display the divider line and shaded region for future data (default: FALSE).}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{title}{Title of the plot.}

\item{xlim}{Limits for the x-axis.}

\item{ylim}{Limits for the y-axis.}

\item{print}{Logical indicating whether to print the plot (default: TRUE).}

\item{...}{Additional arguments passed to the plotting function.}
}
\value{
Invisibly returns a plotly object.
}
\description{
Plot observed and predicted concentrations over time for both past and future data,
with options to apply a multiplier to the concentrations, and customize the plot appearance. The top 5 posterior support
points are highlighted with distinct colors.
}
