% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMMopt.R
\name{plot.MMopt}
\alias{plot.MMopt}
\title{Plot Pmetrics Multiple-Model Optimal Sampling Objects}
\usage{
\method{plot}{MMopt}(x, mm.col = "red", mm.lty = 2, mm.lwd = 2, ...)
}
\arguments{
\item{x}{The name of an \emph{MMopt} data object generated by \code{\link{MMopt}}}

\item{mm.col}{Color of the optimal sample time reference lines.  Default is red.}

\item{mm.lty}{Type of the optimal sample time reference lines.  Default is dashed.}

\item{mm.lwd}{Width of the optimal sample time reference lines.  Default is 2.}

\item{\dots}{Other parameters to pass to \code{\link{plot.PMsim}}.}
}
\value{
Plots the simulation profiles with MMoptimal times indicated as vertical lines.
}
\description{
Plots \emph{MMopt} objects
}
\details{
Simulated observations are plotted on the y-axis vs. time on the x.axis.  
Optimal sampling times are indicated as vertical lines.
}
\seealso{
\code{\link{plot.PMsim}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{axis}}
}
\author{
Michael Neely
}
