% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_cycle.R
\name{plot.PM_cycle_data}
\alias{plot.PM_cycle_data}
\title{Plot PM_cycle_data objects}
\usage{
\method{plot}{PM_cycle_data}(x, ...)
}
\arguments{
\item{x}{A \code{PM_cycle_data} object}

\item{...}{Additional arguments passed to \link{plot.PM_cycle}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Plots the raw data (\code{class: PM_cycle_data}) from a \link{PM_cycle} object in the same way as plotting a \link{PM_cycle} object.
Both use \link{plot.PM_cycle}.
}
\examples{
# There is no example we can think of to filter or otherwise process a PM_cycle object,
# but we provide this function for completeness.
NPex$cycle$data \%>\% plot()
}
