% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_data.R
\name{plot.PM_data_data}
\alias{plot.PM_data_data}
\title{Plot method for PM_data data frames}
\usage{
\method{plot}{PM_data_data}(x, ...)
}
\arguments{
\item{x}{A data frame in the format of the \code{standard_data} field of a \link{PM_data} object}

\item{...}{Additional arguments passed to \link{plot.PM_data}}
}
\value{
A plot of the data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Allows plotting of altered \link{PM_data} objects.
}
\details{
This is useful if you want to modify the data in a \link{PM_data} object, e.g. to filter the data, but still want to use the plotting capabilities of \link{plot.PM_data}.
See \link{plot.PM_data} for details on how to format the plot.
}
\examples{
\dontrun{
# filter then plot the standard_data data frame from a PM_data object
dataEx$standard_data |>
  filter(gender == 0) |>
  plot()
}

}
\seealso{
\link{PM_data}, \link{plot.PM_data}
}
\author{
Michael Neely
}
