% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_final.R
\name{plot.PM_final}
\alias{plot.PM_final}
\title{Plot Pmetrics Final Cycle Parameter Value Distributions}
\usage{
\method{plot}{PM_final}(
  x,
  formula = NULL,
  line,
  marker = TRUE,
  standardize,
  legend,
  log,
  grid = TRUE,
  xlab,
  ylab,
  zlab,
  title,
  xlim,
  ylim,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{The name of an \link{PM_final} data object as a field in a \link{PM_result} R6
object, e.g \code{PM_result$final}.}

\item{formula}{An optional formula of the form \code{y ~ x}, where \code{y} and \code{x}
are two model parameters to plot in a 3-dimensional bivariate plot.  See details.}

\item{line}{\verb{r template("line")} The \code{line} argument is used to format:
\itemize{
\item the density line drawn over an NPAG \link{PM_final} object. Default is \code{FALSE},
which means no density line will be drawn. Use \code{TRUE} to draw the default line,
which is black, solid, and of width 1, or specify as a list to control these
elements, e.g. \code{line = list(color = "red", width = 2, dash = "dash")}
\item the drop lines from point to lower surface when a \code{formula} is specified
to generate a bivariate plot from an NPAG \link{PM_final} object.
In this case, default is \code{line = TRUE}. The default
format is black, dashed, and of width 1.
\item the lines drawing the normal distribution of parameter values from an IT2B \link[=PM_final]{PM_Final} object.
Again, here the default is \code{line = TRUE}, and the format is black, solid, of width 1.
See \link{density}.  Ignored for IT2B output.
}}

\item{marker}{See details for which objects the \code{marker} argument controls.
This argument maps to the plotly marker object.
It can be boolean or a list.
\code{TRUE} will plot the marker with default characteristics.
\code{FALSE} will suppress marker plotting.
If a list, can control many marker characteristics, including overriding defaults.
Use the plotly \code{plotly::schema()} command in the console and navigate
to traces > scatter > attributes > marker to see all the ways the marker
can be formatted. Most common will be:
\itemize{
\item \code{color} Fill color for NPAG bars, marker color for bivariate NPAG plots.
Ignored for IT2B plots.
\item \code{symbol} Plotting character. See \code{plotly::schema()}, traces > scatter >
attributes > marker > symbol > values. Only relevant for bivariate NPAG plots.
\item \code{size} Character size in points. Only relevant for bivariate NPAG plots.
\item \code{opacity} Bar fill color for univariate NPAG plots or marker opacity for
bivariate NPAG plots. Ignored for IT2B plots.
Ranges between 0 (fully transparent) to 1 (fully opaque).
\item \code{line} A list of  additional attributes governing the outline for bars in
univariate NPAG plots or markers in bivariate NPAG plots. Ignored for IT2B plots.
\itemize{
\item \code{color} Outline color. Default is "black".
\item \code{width} Outline width. Default is 1.
Example: \code{marker = list(color = "red", symbol = "triangle", opacity = 0.8, line = list(color = "black", width = 2))}
}
}}

\item{standardize}{Standardize the normal parameter distribution plots from IT2B to the same
scale x-axis.  Ignored for NPAG output.}

\item{legend}{Ignored for this plot.}

\item{log}{\verb{r template("log")}}

\item{grid}{\verb{r template("grid")}}

\item{xlab}{\verb{r template("xlab")} Default is the name of the plotted x-variable.
Formatting can be controlled, but the text is recommended to not be changed.}

\item{ylab}{\verb{r template("ylab")} Default is "Probability" for univariate plots,
and the name of the plotted y-variable for bivariate plots.
Formatting can be controlled, but the text is recommended to not be changed.}

\item{zlab}{Only for bivariate plots. Default is "Probability". Can be a list
to control formatting or default text, as for \code{xlab} and \code{zlab}.}

\item{title}{\verb{r template("title")} Default is to have no title on plots.}

\item{xlim}{\verb{r template("xlim")}}

\item{ylim}{\verb{r template("ylim")}}

\item{print}{If \code{TRUE}, will print the plotly object and return it. If \code{FALSE}, will only return the plotly object.}

\item{...}{\verb{r template("dotsPlotly")}}
}
\value{
Plots the object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Plot R6 \link{PM_final} objects loaded as a field in the
\link{PM_result} object, e.g. \code{PM_result$final}.
}
\details{
This is a function usually called by the \verb{$plot()} method for \link{PM_final} objects
within a \link{PM_result} to generate a plot the parameter value probability densities
after completion of a model fitting.
The function can be called directly on a \link{PM_final} object.

If \code{formula} is omitted, this will generate a marginal plot for each parameter.
For NPAG data, this will be a histogram of marginal values for each parameter and the associated probability
of that value.  For IT2B, this will be a series of normal distributions with mean and standard deviation
equal to the mean and standard deviation of each parameter marginal distribution.

On the other hand, if \code{formula} is specified as two parameters, e.g. CL~V, this will generate a bivariate plot.
For NPAG data, it will be support point with size proportional to the probability
of each point.  For IT2B, it will be an elliptical distribution of a bivariate normal distribution centered at the mean
of each plotted variable and surrounding quantiles of the bivariate distribution plotted in decreasing shades of grey.
Mulitvariate normal density code is borrowed from the mvtnorm package.
}
\examples{
\dontrun{
# NPAG
NPex$final$plot()
NPex$final$plot(line = TRUE)
NPex$final$plot(Ke ~ V)
# IT2B
ITex$final$plot()
ITex$final$plot(Ke ~ V)
}
}
\seealso{
\link{PM_final}, \link[plotly:schema]{schema}

Other PMplots:
\code{\link[=plot.PM_cov]{plot.PM_cov()}},
\code{\link[=plot.PM_cycle]{plot.PM_cycle()}},
\code{\link[=plot.PM_data]{plot.PM_data()}},
\code{\link[=plot.PM_model]{plot.PM_model()}},
\code{\link[=plot.PM_op]{plot.PM_op()}},
\code{\link[=plot.PM_opt]{plot.PM_opt()}},
\code{\link[=plot.PM_pop]{plot.PM_pop()}},
\code{\link[=plot.PM_post]{plot.PM_post()}},
\code{\link[=plot.PM_pta]{plot.PM_pta()}},
\code{\link[=plot.PM_sim]{plot.PM_sim()}},
\code{\link[=plot.PM_valid]{plot.PM_valid()}}
}
\author{
Michael Neely
}
\concept{PMplots}
