% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_final.R
\name{plot.PM_final_data}
\alias{plot.PM_final_data}
\title{Plot PM_final_data objects}
\usage{
\method{plot}{PM_final_data}(x, ...)
}
\arguments{
\item{x}{A `PM_final_data`` object}

\item{...}{Additional arguments passed to \link{plot.PM_final}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Plots the raw data (\code{class: PM_final_data}) from a \link{PM_final} object in the same way as plotting a \link{PM_final} object.
Both use \link{plot.PM_final}.
}
\examples{
NPex$final$data \%>\% plot()
}
