% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PM_model.R
\name{plot.PM_model}
\alias{plot.PM_model}
\title{Plot PM_model objects}
\usage{
\method{plot}{PM_model}(model)
}
\arguments{
\item{model}{The name of an \link{PM_model} object.}
}
\value{
Plots the object.
}
\description{
Plots a \link{PM_model} based on differential equations using network plots from tidygraph and ggraph packages.
}
\details{
This accepts a \link{PM_model} object with minimum of \verb{$eqn} field and optionally an \verb{$out} field within the
\verb{$model_list} field.
}
\examples{
#to be implemented: NPex$model$plot()
}
\seealso{
\link{PM_model}, \code{\link[ggraph:ggraph]{ggraph::ggraph()}}, \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}

Other PMplots: 
\code{\link{plot.MMopt}()},
\code{\link{plot.PM_cycle}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_pta}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()},
\code{\link{plot.PMcov}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMdopt}()}
}
\author{
Markus Hovd, Julian Otalvaro, Michael Neely
}
\concept{PMplots}
