% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_op.R
\name{plot.PM_op_data}
\alias{plot.PM_op_data}
\title{Plot PM_op_data objects}
\usage{
\method{plot}{PM_op_data}(x, ...)
}
\arguments{
\item{x}{A `PM_op_data`` object}

\item{...}{Additional arguments passed to \link{plot.PM_op}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Plots the raw data (\code{class: PM_op_data}) from a \link{PM_op} object in the same way as plotting a \link{PM_op} object.
Both use \link{plot.PM_op}.
}
\examples{
\dontrun{
NPex$op$data \%>\%
dplyr::filter(pred > 5) \%>\%
dplyr::filter(pred < 10) \%>\%
plot()
}
}
