% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_opt.R
\name{plot.PM_opt}
\alias{plot.PM_opt}
\title{Plot Pmetrics Multiple-Model Optimal Sampling Objects}
\usage{
\method{plot}{PM_opt}(x, line = list(probs = NA), times = T, print = TRUE, ...)
}
\arguments{
\item{x}{A \link{PM_opt} object}

\item{line}{Passed to \link{plot.PM_sim} with default as \code{list(probs = NA)}.}

\item{times}{Format the vertical lines for optimal times. Default is
dashed red line. r template("line")`}

\item{print}{If \code{TRUE}, will print the plotly object and return it. If \code{FALSE}, will only return the plotly object.}

\item{...}{Other parameters to pass to \link{plot.PM_sim}.}

\item{probs}{Default is NA. See \link{plot.PM_sim} for details.}
}
\value{
Plots the simulation profiles with MM optimal times indicated as vertical lines.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Plots \link{PM_opt} objects
}
\details{
Simulated observations are plotted on the y-axis vs. time on the x.axis.
Optimal sampling times are indicated as vertical lines. Defaults for optimal
sample times are red, dash, width 2. Defaults for the \code{line} format are as
for \link{plot.PM_sim}.
}
\seealso{
\link{plot.PM_sim}

Other PMplots: 
\code{\link{plot.PM_cov}()},
\code{\link{plot.PM_cycle}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_model}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_pop}()},
\code{\link{plot.PM_post}()},
\code{\link{plot.PM_pta}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()}
}
\author{
Michael Neely
}
\concept{PMplots}
