% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMsim.R
\name{plot.PM_simlist}
\alias{plot.PM_simlist}
\title{Plot list of Pmetrics Simulation Objects}
\usage{
\method{plot}{PM_simlist}(x, ...)
}
\arguments{
\item{x}{The name of a \emph{PM_simlist} object created when \code{combine = F}
is used when parsing the results of a simulation with multiple subjects
in the data template.}

\item{...}{Parameters passsed to \link{plot.PM_sim}.}
}
\description{
Plots a list of \emph{PM_sim} objects created when \code{combine = F}.
}
\details{
When simulating from a data template with multiple subjects, the
default is to create a list of \link{PM_sim} objects of class \emph{PM_simlist}.
This function passes that list to \link{plot.PM_sim}. If the 'simnum' argument
is not chosen, the first object will be plotted.
}
\examples{
library(PmetricsData)
simEx$plot(at = 2)
}
