% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMcov.R
\name{plot.PMcov}
\alias{plot.PMcov}
\title{Plot Pmetrics Covariate objects}
\usage{
\method{plot}{PMcov}(
  x,
  formula,
  icen = "median",
  include,
  exclude,
  mult = 1,
  log = F,
  square = F,
  ref = F,
  lowess = F,
  grid = F,
  ident = F,
  reg = F,
  ci = 0.95,
  cex = 1,
  cex.lab = 1.2,
  x.stat = 0.6,
  y.stat = 0.1,
  col.stat = "black",
  cex.stat = 0.8,
  lwd = 2,
  col = "red",
  xlim,
  ylim,
  xlab,
  ylab,
  out = NA,
  ...
)
}
\arguments{
\item{x}{The name of an \emph{PMcov} data object generated by \code{\link{makeCov}}}

\item{formula}{This is a mandatory formula of the form \code{y ~ x}, where \code{y} and \code{x}
are the two \code{data} parameters to plot.}

\item{icen}{A character vector to summarize covariate and parameter values.  Default is \dQuote{median}, but can also be one of \dQuote{none},
\dQuote{mean}.  If \code{time} is a variable in \code{formula}, the value will be set to \dQuote{none} and the \code{y} values
will be aggregated by subject ID vs. time.}

\item{include}{A vector of subject IDs to include in the plot, e.g. c(1:3,5,15)}

\item{exclude}{A vector of subject IDs to exclude in the plot, e.g. c(4,6:14,16:20)}

\item{mult}{Multiplication factor for y axis, e.g. to convert mg/L to ng/mL}

\item{log}{Boolean operator to plot in log-log space; the default is \code{False}}

\item{square}{Boolean operator to force a square plot with equal x and y limits; the default is \code{True}}

\item{ref}{Boolean operator to draw a unity line; the default is \code{True} unless \dQuote{time} is the x value in \code{formula} in which case this is ignored}

\item{lowess}{Boolean operator to draw a lowess regression line; the default is \code{False} and this is ignored if \dQuote{time} is the x value in \code{formula}}

\item{grid}{Either a boolean operator to plot a reference grid, or a list with elements x and y,
each of which is a vector specifying the native coordinates to plot grid lines; the default is \code{False}.
For example, grid=list(x=seq(0,24,2),y=1:10).  Defaults for missing x or y will be calculated by \code{\link{axTicks}}.}

\item{ident}{Boolean operator to plot points as ID numbers; the default is \code{False}.
This option is useful to identify outliers.}

\item{reg}{Boolean operator to draw a linear regression line; the default is \code{True} unless \dQuote{time} is the x value in \code{formula} in which case this is ignored.
If this option is selected, regression statistics will be printed on the plot if at least 3 subjects are included.}

\item{ci}{The confidence interval for the linear regression parameter estimates; the default is 0.95.}

\item{cex}{Size of the plot symbols.}

\item{cex.lab}{Size of the plot labels.}

\item{x.stat}{Horizontal position to plot the linear regression statistics;
the units are relative to the origin, i.e. extreme left is 0 and extreme right is 1.}

\item{y.stat}{Vertical position to plot the linear regression statistics;
the units are relative to the origin, i.e. extreme bottom is 0 and extreme top is 1.}

\item{col.stat}{Color of the text for the regression statistics.}

\item{cex.stat}{Size of the text for the regression statistics.}

\item{lwd}{Width of the various regression or reference lines (unity, linear regression, or
lowess regression)}

\item{col}{This parameter will be applied to the plotting symbol and is \dQuote{red} by default.}

\item{xlim}{Limits of the x-axis as a vector, e.g. \code{c(0,1)}.  It does not need to be specified, but can be.}

\item{ylim}{Analogous to \code{xlim}}

\item{xlab}{Label for the x-axis.  If missing, will default to the name of the x-variable.}

\item{ylab}{Label for the y-axis.  If missing, will default to the name of the y-variable.}

\item{out}{Direct output to a PDF, EPS or image file.  Format is a named list whose first argument,
\code{type} is one of the following character vectors: \dQuote{pdf}, \dQuote{eps} (maps to \code{postscript}),
\dQuote{\code{png}}, \dQuote{\code{tiff}}, \dQuote{\code{jpeg}}, or \dQuote{\code{bmp}}.  Other named items in the list
are the arguments to each graphic device. PDF and EPS are vector images acceptable to most journals
in a very small file size, with scalable (i.e. infinite) resolution.  The others are raster images which may be very
large files at publication quality dots per inch (DPI), e.g. 800 or 1200. Default value is \code{NA} which means the
output will go to the current graphic device (usually the monitor). For example, to output an eps file,
out=list(\dQuote{eps}) will generate a 7x7 inch (default) graphic.}

\item{\dots}{Other parameters as found in \code{\link{plot.default}}.}
}
\value{
Plots the object.
}
\description{
Plot PMcov objects
}
\details{
This method will plot any two columns, specified using a formula, of a PMcov object, which contains covariate and Bayesian posterior parameter information
for each subject.  Specifiying any two variables that do not include time will result in a scatter plot with optional regression and reference lines.  If
time is included as the x variable, the y variable will be plotted vs. time, aggregated by subject.  This can be useful to see time varying parameters,
although a formula within formula approach may be required, e.g. plot(cov.1,I(cl_0*wt**0.75)~time) in order to see the change in cl over time according to
the change in wt over time, even though cl_0 is constant for a given subject.
}
\examples{
data(cov.1)
plot(cov.1,V~wt)
}
\seealso{
\code{\link{makeCov}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{axis}}

Other PMplots: 
\code{\link{plot.MMopt}()},
\code{\link{plot.PM_cov}()},
\code{\link{plot.PM_cycle}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_model}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_pta}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMdopt}()}
}
\author{
Michael Neely
}
\concept{PMplots}
