% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMdopt.R
\name{plot.PMdopt}
\alias{plot.PMdopt}
\title{Plot Pmetrics D-optimal Times}
\usage{
\method{plot}{PMdopt}(
  x,
  col.mean = "red",
  lwd.mean = 4,
  ticksize.mean = 0.1,
  xlab = "Time",
  ylab = "Probability",
  layout = c(1, 1),
  ...
)
}
\arguments{
\item{x}{The name of an \emph{PMdopt} data object generated by \code{\link{Dopt}}}

\item{col.mean}{This parameter will be applied to the tick mark indicating the weighted mean
optimal time}

\item{lwd.mean}{This parameter will be applied to the tick mark indicating the weighted mean
optimal time}

\item{ticksize.mean}{This parameter will be applied to the tick mark indicating the weighted mean
optimal time}

\item{xlab}{Define x-axis label.  Default is \dQuote{Time}.}

\item{ylab}{Define y-axis label.  Default \dQuote{Probability}.}

\item{layout}{This parameter specifies the number of rows and columns per page, e.g. layout=c(2,2).}

\item{\dots}{Other parameters as found in \code{\link{plot.default}}.}
}
\value{
Plots the object.
}
\description{
Plot PMdopt objects
}
\details{
This function will plot the output of the \code{\link{Dopt}} function.  A histogram is generated
with the probability distribution of each optimal time for each support point in the model,
and the weighted mean for that time.
}
\seealso{
\code{\link{Dopt}}, \code{\link{summary.PMdopt}}

Other PMplots: 
\code{\link{plot.MMopt}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PMcov}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMpta}()},
\code{\link{plot.PMsim}()},
\code{\link{plot.PMvalid}()}
}
\author{
Michael Neely
}
\concept{PMplots}
