% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMop.R
\name{plot.PMop}
\alias{plot.PMop}
\title{Plot Pmetrics Observed vs. Predicted Objects}
\usage{
\method{plot}{PMop}(
  x,
  include,
  exclude,
  pred.type = "post",
  icen = "median",
  outeq = 1,
  mult = 1,
  resid = F,
  log = F,
  square = T,
  ref = T,
  lowess = F,
  reg = T,
  grid,
  ident = F,
  ci = 0.95,
  cex = 1,
  cex.lab = 1.2,
  x.stat = 0.4,
  y.stat = 0.1,
  col.stat = "black",
  cex.stat = 1.2,
  lwd = 2,
  col = "red",
  xlim,
  ylim,
  xlab,
  ylab,
  out = NA,
  ...
)
}
\arguments{
\item{x}{The name of an \emph{PMop} data object generated by \code{\link{makeOP}}.}

\item{include}{A vector of subject IDs to include in the plot, e.g. c(1:3,5,15)}

\item{exclude}{A vector of subject IDs to exclude in the plot, e.g. c(4,6:14,16:20)}

\item{pred.type}{Either 'post' for a posterior object or 'pop' for a population object.  Default is 'post'.}

\item{icen}{Can be either "median" for the predictions based on medians of \code{pred.type} parameter value
distributions, or "mean".  Default is "median".}

\item{outeq}{Output equation number.  Default is 1.}

\item{mult}{Multiplication factor for x and y axes, e.g. to convert mg/L to ng/mL.  Ignored for residual plots.}

\item{resid}{Boolean operator to generate a plot of weighted prediction error vs. prediction,
a plot of weighted prediction error vs. time, and histogram plot of the weighted prediction errors,
with overlying normal distribution of the same mean and variance if \code{ref} is true, and
a P-value for the Kolmogorov-Smirnov test for non-normality if \code{reg} is true.  The default is \code{False}.}

\item{log}{Boolean operator to plot in log-log space.  This parameter is ignored for residual plots.  The default is \code{False}}

\item{square}{Boolean operator to force a observed vs. predicted plots to be square with equal x and y limits.
This parameter is ignored for residual plots.  The default is \code{True}}

\item{ref}{Boolean operator to draw a reference line of slope 1 in observed vs. predicted plots and
slope 0 in residual plots, or a reference normal distribution in residual histogram; the default is \code{True}}

\item{lowess}{Boolean operator to draw a lowess regression line in observed vs. predicted or
residual plots; the default is \code{False}}

\item{reg}{Boolean operator to draw a linear regression line and print regression statistics on the plot.
For weighted residual plots, it will print the mean weighted prediction error with P value for difference from 0,
and the standard deviation of the weighted prediction errors, as well as the probability that the distribution of
weighted residuals is not different from normal by the Kolmogorov-Smirnov test.  The default is \code{True}.}

\item{grid}{Either a boolean operator to plot a reference grid, or a list with elements x and y,
each of which is a vector specifying the native coordinates to plot grid lines; the default is \code{False}.
For example, grid=list(x=seq(0,24,2),y=1:10).  Defaults for missing x or y will be calculated by \code{\link{axTicks}}.
For residual plots, list values for \code{grid} will be interpreted as \code{True}, i.e. custom grid lines are not allowed.}

\item{ident}{Boolean operator to plot points as ID numbers; the default is \code{False}.
This option is useful to identify outliers.}

\item{ci}{The confidence interval for the linear regression parameter estimates; the default is 0.95.}

\item{cex}{Size of the plot symbols.}

\item{cex.lab}{Size of the plot labels.}

\item{x.stat}{Horizontal position to plot the regression or residual statistics;
the units are relative to the origin, i.e. extreme left is 0 and extreme right is 1.}

\item{y.stat}{Vertical position to plot the regression or residual statistics; 
the units are relative to the origin, i.e. extreme bottom is 0 and extreme top is 1.}

\item{col.stat}{Color of the text for the regression or residual statistics.}

\item{cex.stat}{Size of the text for the regression or residual statistics}

\item{lwd}{Width of the various regression or reference lines (reference, linear regression, or
lowess regression)}

\item{col}{This parameter will be applied to the plotting symbol and is \dQuote{red} by default.}

\item{xlim}{Limits of the x-axis as a vector, e.g. \code{c(0,1)}.  It does not need to be specified, but can be.}

\item{ylim}{Analogous to \code{xlim}}

\item{xlab}{Label for the x-axis.  If missing, will default to \dQuote{Observed}.}

\item{ylab}{Label for the y-axis.  If missing, will default to \dQuote{Predicted}.}

\item{out}{Direct output to a PDF, EPS or image file.  Format is a named list whose first argument, 
\code{type} is one of the following character vectors: \dQuote{pdf}, \dQuote{eps} (maps to \code{postscript}),
\dQuote{\code{png}}, \dQuote{\code{tiff}}, \dQuote{\code{jpeg}}, or \dQuote{\code{bmp}}.  Other named items in the list
are the arguments to each graphic device. PDF and EPS are vector images acceptable to most journals
in a very small file size, with scalable (i.e. infinite) resolution.  The others are raster images which may be very
large files at publication quality dots per inch (DPI), e.g. 800 or 1200. Default value is \code{NA} which means the 
output will go to the current graphic device (usually the monitor). For example, to output an eps file,
out=list(\dQuote{eps}) will generate a 7x7 inch (default) graphic.}

\item{\dots}{Other parameters as found in \code{\link{plot.default}}.}
}
\value{
Plots the object.
}
\description{
Plot PMop objects
}
\examples{
data(PMex1)
plot(op.1)
}
\seealso{
\code{\link{makeOP}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{axis}}
}
\author{
Michael Neely
}
