% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMpta.R
\name{plot.PMpta}
\alias{plot.PMpta}
\title{Plot PMpta Percent Target Attainment objects}
\usage{
\method{plot}{PMpta}(
  x,
  include,
  exclude,
  plot.type = "pta",
  log = T,
  pch,
  grid,
  xlab,
  ylab,
  col,
  lty,
  lwd = 4,
  legend = T,
  ci = 0.9,
  out = NA,
  ...
)
}
\arguments{
\item{x}{The name of an \emph{PMpta} data object read by \code{\link{makePTA}}}

\item{include}{A vector of simulations (regimens) to include in the plot, e.g. c(1,3)}

\item{exclude}{A vector of simulations (regimens) in the plot, e.g. c(2,4:6)}

\item{plot.type}{Character vector controlling type of plot.
Default is \dQuote{pta}, which plots proportion with success on the y-axis and target on the x-axis.
The other choice is \dQuote{pdi}, which plots the median pdi (pharmacodynamic index), e.g. AUC/MIC, on the
y-axis, and target on the x-axis.}

\item{log}{Boolean operator to plot x-axis in logarithmic scale; the default is \code{True}}

\item{pch}{Vector of integers which control the plotting symbol for each regimen curve; the default is 1:nsim.  NA results in no symbol.
Use 0 for open square, 1 for open circle, 2 for open triangle, 3 for cross, 4 for X, or 5 for a diamond.
Other alternatives are \dQuote{*} for asterisks, \dQuote{.} for tiny dots, or \dQuote{+} for a smaller,
bolder cross.  These plotting symbols are standard for R (see \code{\link{par}}).}

\item{grid}{Either a boolean operator to plot a reference grid, or a list with elements x and y,
each of which is a vector specifying the native coordinates to plot grid lines; the default is \code{False}.
For example, grid=list(x=seq(0,24,2),y=1:10).  Defaults for missing x or y will be calculated by \code{\link{axTicks}}.}

\item{xlab}{Label for the x axis.  Default is \dQuote{MIC}}

\item{ylab}{Label for the y axis.  Default is \dQuote{Proportion with success}}

\item{col}{A vector of color names to be used for each regimen plotted.  If the
length of \code{col} is too short, values will be recycled.}

\item{lty}{A vector of line types to be used for each regimen plotted.  If the
length of \code{lty} is too short, values will be recycled.}

\item{lwd}{Line width, with default of 4.}

\item{legend}{Either a boolean operator or a list of parameters to be supplied to the \code{\link{legend}}
function (see its documentation).  If \code{False}, a legend will not be plotted.
If \code{True} (the default), the default legend parameters will be used, as documented in that function, with exceptions
as noted in \emph{Details}.}

\item{ci}{Confidence interval around curves on \code{pdi} plot, on scale of 0 to 1. Default is 0.9.}

\item{out}{Direct output to a PDF, EPS or image file.  Format is a named list whose first argument,
\code{type} is one of the following character vectors: \dQuote{pdf}, \dQuote{eps} (maps to \code{postscript}),
\dQuote{\code{png}}, \dQuote{\code{tiff}}, \dQuote{\code{jpeg}}, or \dQuote{\code{bmp}}.  Other named items in the list
are the arguments to each graphic device. PDF and EPS are vector images acceptable to most journals
in a very small file size, with scalable (i.e. infinite) resolution.  The others are raster images which may be very
large files at publication quality dots per inch (DPI), e.g. 800 or 1200. Default value is \code{NA} which means the
output will go to the current graphic device (usually the monitor). For example, to output an eps file,
out=list(\dQuote{eps}) will generate a 7x7 inch (default) graphic.}

\item{\dots}{Other parameters as found in \code{\link{plot.default}}.}
}
\value{
Plots the object.
}
\description{
Plots PMpta objects
}
\details{
This function will plot the percent target attainment for objects made with the \code{\link{makePTA}} function.
For the legend, defaults that are different that the standard are:
\itemize{
  \item x Default \dQuote{topright}
  \item legend Default will be the labeled regimen names supplied during \code{\link{makePTA}},
  or if missing, \dQuote{Regimen 1, Regimen 2,...Regimen n}, where \emph{n} is the number of
  regimens in the PMpta object.
  This default can be overridden by a supplied character vector of regimen names.
  \item col The color of each Regimen plot as specified by the default color scheme or \code{col}
  \item pch The plotting character for each Regimen plot as specified by the default plotting characters or \code{pch}
  \item lty The line type of each Regimen plot as specified by the default line types or \code{lty}
  \item bg Default \dQuote{white}
}
}
\seealso{
\code{\link{makePTA}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{axis}}
}
\author{
Michael Neely
}
