% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMsim.R
\name{plot.PMsim}
\alias{plot.PMsim}
\title{Plot Pmetrics Simulation Objects}
\usage{
\method{plot}{PMsim}(
  x,
  mult = 1,
  log = T,
  probs = list(0.05, 0.25, 0.5, 0.75, 0.95),
  ci = 0.95,
  binSize = 0,
  outeq = 1,
  marker = F,
  grid = F,
  xlab = "Time",
  ylab = "Output",
  xlim,
  ylim,
  legend = F,
  obs,
  ...
)
}
\arguments{
\item{x}{The name of an \emph{PMsim} data object generated by \link{PM_sim}}

\item{mult}{Multiplication factor for y axis, e.g. to convert mg/L to ng/mL}

\item{log}{Boolean operator to plot in log-log space; the default is \code{False}}

\item{probs}{List of quantiles to plot, e.g. \code{probs = list(0.05, 0.5, 0.95)}. If set to \code{NA}, all simulated profiles will be plotted,
and numerical predictive checking will be suppressed. To format individual quantile lines, expand the list object for that quantile
For example, \code{probs = list(0.05, list(value = 0.5, color = "red", width = 2, dash = "dash"), 0.95)}.
Use the plotly \code{schema()} command
in the console and navigate to traces -> scatter -> attributes -> line to see all the ways the marker
can be formatted. Most common will be:
\itemize{
\item color Fill color and outline color if not otherwise specified
\item symbol Plotting character. See \code{schema()}, traces -> scatter -> attributes -> marker -> symbol -> values.
\item size Character size in points
\item opacity Ranging between 0 (fully transparent) to 1 (fully opaque).
}}

\item{ci}{Width of confidence interval bands around simulated quantiles, from 0 to 1.  If 0, or \emph{nsim}<100, will not plot.
Default is 0.95, i.e. 95th percentile with tails of 2.5 percent above and below excluded.}

\item{binSize}{Width of binning interval for simulated concentrations, in time units, e.g. hours.
A \code{binSize} of 0.5 will pull all simulated concentrations +/- 0.5 hours into the same time.  This is useful
for plotting PMsim objects made during \link{makeValid}. The default is 0, i.e. no binning.}

\item{outeq}{Which output equation to plot if more than 1}

\item{marker}{Controls the plotting symbol for observations; default is \code{FALSE} which results in no symbol.
Will be set to \code{TRUE} if \code{obs} is supplied. In this case, the default plotly marker will be an open circle, black in color,
point size 8. To change these, specify arguments as a list to \code{marker}. For example,
\verb{marker = list(color = "red", symbol = "triangle", size = 10}). Use the plotly \code{schema()} command
in the console and navigate to traces -> scatter -> attributes -> marker to see all the ways the marker
can be formatted. Most common will be:
\itemize{
\item color Marker color, default black.
\item symbol Plotting character. See \code{schema()}, traces -> scatter -> attributes -> marker -> symbol -> values.
Default is open circle.
\item size Character size in points, default 8.
\item opacity Ranging between 0 (fully transparent) to 1 (fully opaque, the default).
\item line A list of  additional attributes governing the outline for filled shapes, most commonly
color and width, e.g \code{marker = list(color = "red", line = list(color = "black", width = 2))}
If missing,  defaults to marker color and width 1.
}}

\item{grid}{If \code{FALSE}, the default, no grid plotted. If \code{TRUE}, the default
color \emph{grey50} and width 1 will be plotted at major tick marks. If a list,
color and width can be customized, e.g. \code{grid = list(gridcolor = "black", gridwidth = 2)}.
The grid parameter modifies the xaxis and yaxis layout object in plotly.
See \code{schema()} layout -> xaxis/yaxis -> gridcolor/gridwidth.}

\item{xlab}{Label for x-axis; default is "Time". The xlab parameter modifies
the xaxis title in the layout object in plotly.
See \code{schema()} layout -> xaxis -> title.}

\item{ylab}{Label for y-axis; default is "Output". The ylab parameter modifies
the yaxis title in the layout object in plotly.
See \code{schema()} layout -> yaxis -> title.}

\item{xlim}{Limits of the x-axis as a vector, e.g. \code{c(0,1)}.
The xlim parameter modifies
the xaxis range in the layout object in plotly.
See \code{schema()} layout -> xaxis -> range}

\item{ylim}{Analogous to \code{xlim} for the yaxis.}

\item{legend}{Either a boolean operator (most common) or a list of parameters to be supplied to plotly.
If \code{FALSE} (the default), a legend will not be plotted. Type \code{schema()}
into the console (with the plotly package loaded), select layout > layoutAttributes > legend
for more details on the available options for formatting.}

\item{obs}{The name of a \link{PM_result} data object or the PM_op field in the
PM_result object, all generated by \link{PM_load}.  For example, if
\code{run1 <- PM_load(1)} and \code{sim1} is a PM_sim object, then
\code{sim1$plot(obs = run1)} or \code{sim1$plot(obs = run1$op)}.
If specified,
the observations will be overlaid upon the simulation plot
enabling a visual predictive check.  In this case,
a list object will be returned with two items: $npc containing the quantiles
and probability that the observations
are below each quantile (binomial test); and $simsum, the times of each
observation and the
value of the simulated quantile with upper and lower confidence intervals at that time.
Additionally, the number of observations beyond the 5th and 95th percentiles will be reported
and the binomial test P-value if this number is different than the expected 10\% value.}

\item{\dots}{Other parameters which can be passed to control axis formatting. Specify as lists
named either xaxis or yaxis. See \code{schema()} layout -> xaxis/yaxis for options.}
}
\value{
Plots the simulation object.  If \code{obs} is included, a list will be returned with
the folowing items:
\item{npc}{A dataframe with three columns: quantile, prop_less, pval.  \emph{quantile} are those specified
by the \code{probs} argument to the plot call; \emph{prop_less} are the proportion of simulated
observations at all times less than the quantile; \emph{pval} is the P-value of the difference in the
prop.less and quantile by the beta-binomial test.}
\item{simsum}{A dataframe with the quantile concentration at each simulated time,
with lower and upper confidence intervals}
\item{obs}{A dataframe similar to an PMop object made by \code{\link{makeOP}}
with the addition of the quantile for each observation}
}
\description{
Plots \emph{PMsim} objects with the option to perform a visual and numerical predictive check
}
\details{
Simulated observations are plotted as quantiles on the y-axis vs. time on the x.axis.  If measured
observations are included, a visual and numerical predictive check will be performed.
}
\seealso{
\link{PM_sim}, \link{plot_ly}, \link{schema}
}
\author{
Michael Neely
}
