% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{plot.tidy_cov}
\alias{plot.tidy_cov}
\title{Wrapper function for plot.PMcov}
\usage{
\method{plot}{tidy_cov}(x, ...)
}
\arguments{
\item{x}{A \emph{PMcov} object with the \emph{tidy_cov} class to be plotted}

\item{...}{Arguments passed to \link{plot.PMcov}}
}
\value{
A \link{plot.PMcov} object
}
\description{
This redirects to plot.PMcov for tidy_cov objects
}
\details{
See \link{plot.PMcov}. Alternative way to plot is
\code{PM_result$cov$plot()}.
}
