% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{plot.tidy_cycle}
\alias{plot.tidy_cycle}
\title{Wrapper function for plot.PM_cycle}
\usage{
\method{plot}{tidy_cycle}(x, ...)
}
\arguments{
\item{x}{A \emph{PMcycle} object with the \emph{tidy_cycle} class to be plotted}

\item{...}{Arguments passed to \link{plot.PMcycle}}
}
\value{
A \link{plot.PM_cycle} object
}
\description{
This redirects to plot.PM_cycle for tidy_cycle objects
}
\details{
See \link{plot.PMcycle}. Alternative way to plot is
\code{PM_result$cycle$plot()}.
}
