% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{plot.tidy_op}
\alias{plot.tidy_op}
\title{Wrapper function for plot.PM_op}
\usage{
\method{plot}{tidy_op}(x, ...)
}
\arguments{
\item{x}{A \emph{PMop} object with the \emph{tidy_op} class to be plotted}

\item{...}{Arguments passed to \link{plot.PM_op}}
}
\value{
A \link{plot.PM_op} object
}
\description{
This redirects to plot.PM_op for tidy_op objects
}
\details{
See \link{plot.PM_op}. Alternative way to plot is
\code{PM_result$op$plot()}.
}
