% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMbatch.R
\name{PM_batch}
\alias{PM_batch}
\title{Execute a Batch of PM_fit objects}
\usage{
PM_batch(batch, batchname = "batch", ...)
}
\arguments{
\item{batch}{list of PM_fit objects}

\item{batchname}{name of the folder in which the batch will be executed}

\item{...}{other parameters passed to PM_fit$run(...)}
}
\description{
Runs a set of PM_fit objects in order
}
\details{
This function receives a batch (list) of PM_fit objects and executes them in order
}
\keyword{internal}
