% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_bestdose.R
\name{PM_bestdose}
\alias{PM_bestdose}
\title{Object to contain BestDose optimization results}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This object is created after a successful BestDose optimization run.
BestDose finds optimal dosing regimens to achieve target drug concentrations
or AUC values using Bayesian optimization.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_bestdose-new}{\code{PM_bestdose$new()}}
\item \href{#method-PM_bestdose-print}{\code{PM_bestdose$print()}}
\item \href{#method-PM_bestdose-get_doses}{\code{PM_bestdose$get_doses()}}
\item \href{#method-PM_bestdose-get_predictions}{\code{PM_bestdose$get_predictions()}}
\item \href{#method-PM_bestdose-get_auc_predictions}{\code{PM_bestdose$get_auc_predictions()}}
\item \href{#method-PM_bestdose-get_objf}{\code{PM_bestdose$get_objf()}}
\item \href{#method-PM_bestdose-get_status}{\code{PM_bestdose$get_status()}}
\item \href{#method-PM_bestdose-get_method}{\code{PM_bestdose$get_method()}}
\item \href{#method-PM_bestdose-save}{\code{PM_bestdose$save()}}
\item \href{#method-PM_bestdose-clone}{\code{PM_bestdose$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$new(
  prior = NULL,
  model = NULL,
  past_data = NULL,
  target = NULL,
  dose_range = list(min = 0, max = 1000),
  bias_weight = 0.5,
  target_type = "concentration",
  time_offset = NULL,
  settings = NULL,
  result = NULL,
  problem_obj = NULL,
  bias_override = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-print"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-print}{}}}
\subsection{Method \code{print()}}{
Print summary of BestDose results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-get_doses"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-get_doses}{}}}
\subsection{Method \code{get_doses()}}{
Get optimal dose values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$get_doses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric vector of optimal doses
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-get_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-get_predictions}{}}}
\subsection{Method \code{get_predictions()}}{
Get concentration-time predictions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$get_predictions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with predictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-get_auc_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-get_auc_predictions}{}}}
\subsection{Method \code{get_auc_predictions()}}{
Get AUC predictions (if available)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$get_auc_predictions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with AUC predictions or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-get_objf"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-get_objf}{}}}
\subsection{Method \code{get_objf()}}{
Get objective function value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$get_objf()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric objective function value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-get_status}{}}}
\subsection{Method \code{get_status()}}{
Get optimization status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$get_status()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string with status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-get_method"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-get_method}{}}}
\subsection{Method \code{get_method()}}{
Get optimization method used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$get_method()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string: "posterior" or "uniform"
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-save"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-save}{}}}
\subsection{Method \code{save()}}{
Save results to RDS file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$save(filename = "bestdose_result.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Path to save file. Default: "bestdose_result.rds"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
