% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_bestdose.R
\name{PM_bestdose_problem}
\alias{PM_bestdose_problem}
\title{Prepare a reusable BestDose optimization problem}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Use \code{PM_bestdose_problem} to mirror the Rust workflow: compute posterior
support points once, inspect them in R, and solve for multiple bias weights
without repeating the expensive initialization step.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_bestdose_problem-new}{\code{PM_bestdose_problem$new()}}
\item \href{#method-PM_bestdose_problem-finalize}{\code{PM_bestdose_problem$finalize()}}
\item \href{#method-PM_bestdose_problem-optimize_raw}{\code{PM_bestdose_problem$optimize_raw()}}
\item \href{#method-PM_bestdose_problem-optimize}{\code{PM_bestdose_problem$optimize()}}
\item \href{#method-PM_bestdose_problem-clone}{\code{PM_bestdose_problem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose_problem-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose_problem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose_problem$new(
  prior,
  model,
  past_data = NULL,
  target,
  dose_range = list(min = 0, max = 1000),
  bias_weight = 0.5,
  target_type = "concentration",
  time_offset = NULL,
  settings = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose_problem-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose_problem-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose_problem$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose_problem-optimize_raw"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose_problem-optimize_raw}{}}}
\subsection{Method \code{optimize_raw()}}{
Run optimization and return raw list (doses, objf, predictions)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose_problem$optimize_raw(bias_weight = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose_problem-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose_problem-optimize}{}}}
\subsection{Method \code{optimize()}}{
Run optimization and return a \code{PM_bestdose} result object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose_problem$optimize(bias_weight = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_bestdose_problem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_bestdose_problem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_bestdose_problem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
