% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_cov}
\alias{PM_cov}
\title{Contains covariate data}
\description{
Contains a data frame with subject-specific covariate data output
from \link{makeCov}
}
\details{
For each subject, \link{makeCov} extracts covariate information and
Bayesian posterior parameter estimates.
This output of this function is suitable for exploration of covariate-
parameter, covariate-time, or parameter-time relationships.
}
\author{
Michael Neely, Julian Otalvaro
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data frame with the following columns
\itemize{
\item id Subject identification
\item time Times of covariate observations
\item covnames... Columns with each covariate observations in the dataset for each subject and \code{time}
\item parnames... Columns with each parameter in the model and the \code{icen} summary
for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian
parameter means and medians
\item icen The type of summarized Bayesian posterior individual parameter values: mean or median
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PM_cov$new()}}
\item \href{#method-summary}{\code{PM_cov$summary()}}
\item \href{#method-plot}{\code{PM_cov$plot()}}
\item \href{#method-print}{\code{PM_cov$print()}}
\item \href{#method-clone}{\code{PM_cov$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with covariate-parameter information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$new(cov)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cov}}{The parsed output from \link{makeCov}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creation of new \code{PM_cov} object is automatic and not generally necessary
for the user to do.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMcov}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{summary.PMcov}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PMcov}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{plot.PMcov}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{print}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Print method for \emph{PM_cov}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
