% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_final}
\alias{PM_final}
\title{Final Cycle Population Values}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Contains final cycle information from run.
}
\details{
Contains the results of \link{makeFinal}, which is a
list suitable for analysis and plotting of final cycle population values.

However, if you wish to manipulate the entire data frame,
use the \code{data} field, e.g. \code{probs <- run1$final$data$popPoints \%>\% select(prob)}.
This will select the probabilities of the support points. If
you are unfamiliar with the \verb{\%>\%} pipe function, please type \code{help("\%>\%", "magrittr")}
into the R console and look online for instructions/tutorials in tidyverse, a
powerful approach to data manipulation upon which Pmetrics is built.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{popPoints}}{(NPAG only) Data frame of the final cycle joint population density of grid points
with column names equal to the name of each random parameter plus \emph{prob} for the
associated probability of that point}

\item{\code{popMean}}{The final cycle mean for each random parameter distribution}

\item{\code{popSD}}{The final cycle standard deviation for each random parameter distribution}

\item{\code{popCV}}{The final cycle coefficient of variation (SD/Mean) for each random parameter distribution}

\item{\code{popVar}}{The final cycle variance for each random parameter distribution}

\item{\code{popCov}}{The final cycle random parameter covariance matrix}

\item{\code{popCor}}{The final cycle random parameter correlation matrix}

\item{\code{popMedian}}{The final cycle median values for each random parameter,
i.e. those that have unknown mean and unknown variance, both of which are
fitted during the run}

\item{\code{popRanFix}}{The final cycle median values for each parameter that is
random but fixed to be the same for all subjects, i.e. unknown mean, zero
variance, with only mean fitted in the run}

\item{\code{postPoints}}{(NPAG only) Data frame of posterior population points for each of the first 100 subject,
with columns id, point, parameters and probability.  The first column is the subject, the second column has the population
point number, followed by the values for the parameters in that point and the probability.}

\item{\code{postMean}}{A \emph{nsub} x \emph{npar} data frame containing
the means of the posterior distributions for each parameter.}

\item{\code{postSD}}{A \emph{nsub} x \emph{npar} data frame containing
the SDs of the posterior distributions for each parameter.}

\item{\code{postVar}}{A \emph{nsub} x \emph{npar} data frame containing
the variances of the posterior distributions for each parameter.}

\item{\code{postCov}}{NPAG only: An array of dimensions \emph{npar} x \emph{npar} x \emph{nsub} that
contains the covariances of the posterior distributions for each parameter and subject.*}

\item{\code{postCor}}{NPAG only: An array of dimensions \emph{npar} x \emph{npar} x \emph{nsub} that
contains the correlations of the posterior distributions for each parameter and subject.}

\item{\code{postMed}}{A \emph{nsub} x \emph{npar} data frame containing
the medians of the posterior distributions for each parameter.*}

\item{\code{shrinkage}}{A data frame with the shrinkage for each parameter.  \code{popVar}
is comprised of variance(EBE) + variance(EBD), where EBE is the Empirical
Bayes Estimate or mean of the posterior
distribution for the parameter. EBD is the Empirical Bayes Distribution, or
the full Bayesian posterior distribution. In other words, if Bayesian
posterior distributions are wide
for a given parameter due to sparse or uninformative sampling,
then most of the population variance is due
to this variance and shrinkage of the EBE variance is high because
individual posterior estimates
shrink towards the population mean.}

\item{\code{gridpts}}{(NPAG only) Initial number of support points}

\item{\code{nsub}}{Number of subjects}

\item{\code{ab}}{Matrix of boundaries for random parameter values}

\item{\code{data}}{A data frame combining all the above fields as its columns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_final-new}{\code{PM_final$new()}}
\item \href{#method-PM_final-summary}{\code{PM_final$summary()}}
\item \href{#method-PM_final-plot}{\code{PM_final$plot()}}
\item \href{#method-PM_final-clone}{\code{PM_final$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_final-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_final-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with final cycle information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_final$new(final)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{final}}{The parsed output from \link{makeFinal}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creation of new \code{PM_final} object is automatic and not generally necessary
for the user to do.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_final-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_final-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_final$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMfinal}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{summary.PMfinal}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_final-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_final-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_final$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PMfinal}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{plot.PMfinal}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_final-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_final-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_final$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
