% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_fit.R
\name{PM_fit}
\alias{PM_fit}
\title{Object to define and run a model and data in Pmetrics}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{PM_fit} objects comprise a \code{PM_data} and \code{PM_model} object ready for analysis
}
\details{
Data and model objects can be previously created as \link{PM_data} or \link{PM_model} objects,
or created on the fly when making a new PM_fit object. PM_fit objects contain
methods to cross-check data and model objects for compatibility, as well as to
run the analysis.
}
\author{
Michael Neely
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{\link{PM_data} object}

\item{\code{model}}{\link{PM_model} object}

\item{\code{backend}}{Backend used for calculations; default is value in PMoptions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_fit-new}{\code{PM_fit$new()}}
\item \href{#method-PM_fit-run}{\code{PM_fit$run()}}
\item \href{#method-PM_fit-save}{\code{PM_fit$save()}}
\item \href{#method-PM_fit-load}{\code{PM_fit$load()}}
\item \href{#method-PM_fit-check}{\code{PM_fit$check()}}
\item \href{#method-PM_fit-clone}{\code{PM_fit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_fit-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_fit-new}{}}}
\subsection{Method \code{new()}}{
Create a new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_fit$new(data = data, model = model, backend = getPMoptions()$backend, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Either the name of a  \link{PM_data}
object in memory or the quoted name of a Pmetrics
data file in the current working directory, which will crate a \link{PM_data}
object on the fly. However, if created on the fly, this object
will not be available to other
methods or other instances of \link{PM_fit}.}

\item{\code{model}}{Similarly, this is either the name of a \link{PM_model}
object in memory or the quoted name of a Pmetrics text model file
in the current working directory. Again, if created on the fly,
the object will not be available to other
methods or other instances of \link{PM_fit}.}

\item{\code{backend}}{Backend used for calculations; default is value in PMoptions.}

\item{\code{...}}{Other parameters passed to \link{PM_data} or \link{PM_model} if created
from a filename}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_fit-run"></a>}}
\if{latex}{\out{\hypertarget{method-PM_fit-run}{}}}
\subsection{Method \code{run()}}{
Fit the model to the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_fit$run(
  run = NULL,
  include = NULL,
  exclude = NULL,
  cycles = 100,
  prior = "sobol",
  density0 = 0.01,
  seed = 23,
  overwrite = FALSE,
  algorithm = "NPAG",
  report = getPMoptions("report_template"),
  artifacts = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run}}{Specify the run number of the output folder.  Default if missing is the next available number.}

\item{\code{include}}{Vector of subject id values in the data file to include in the analysis.
The default (missing) is all.}

\item{\code{exclude}}{A vector of subject IDs to exclude in the analysis, e.g. \code{c(4,6:14,16:20)}}

\item{\code{cycles}}{Number of cycles to run. Default is 100.}

\item{\code{prior}}{The distribution for the initial support points, which can be
one of several options.
\itemize{
\item The default is "sobol", which is a semi-random distribution. This is the distribution
typically used when fitting a new model to the data.
}

The following all specify non-random, informative prior distributions. They
are useful for either continuing a previous
run which did not converge or for fitting a model to new data, whether to simply
calculate Bayesian posteriors with \code{cycles = 0} or to revise the model to a new
covergence with the new data.
\itemize{
\item The name of a suitable \link{PM_result} object from a prior run loaded with \link{PM_load}.
This starts from the non-uniform, informative distribution obtained at the end of a prior NPAG run.
Example: \verb{run1 <- PM_load(1); fit1$run(prior = run1)}.
\item A character string with the filename of a csv file containing a prior distribution with
format as for 'theta.csv' in the output folder of a prior run: column headers are parameter
names, and rows are the support point values. A final column with probabilities
for each support point is not necessary, but if present will be ignored, as these
probabilities are calculated by the engine. Note that the parameter names must match the
names of the primary variables in the model. Example: \code{fit1$run(prior = "mytheta.csv")}.
\item The number of a previous run with \code{theta.csv} in the output folder which will be read
as for the filename option above. Example: \code{fit1$run(prior = 2)}.
\item A data frame obtained from reading an approriate file, such that the data frame
is in the required format described in the filename option above. Example:
\verb{mytheta <- read_csv("mytheta.csv"); fit1$run(prior = mytheta)}.
}}

\item{\code{density0}}{The proportion of the volume of the model parameter
hyperspace used to calculate the initial number of support points if one of
the semi-random, uniform distributions are selected in the \code{prior} argument
above. The initial points are
spread through that hyperspace and begin the search for the optimal
parameter value distribution (support points) in the population.
The volume of the parameter space is the product of the ranges for all parameters.
For example if using two parameters \code{Ke} and \code{V}, with ranges of [0, 5] and [10, 100],
the volume is (5 - 0) x (100 - 10) = 450 The default value of \code{density0} is 0.01, so the initial
number of support points will be 0.01 x 450 = 4.5, increased to the nearest integer,
which is 5. The greater the initial number of points, the less chance of
missing the globally maximally likely parameter value distribution,
but the slower the run.}

\item{\code{seed}}{Seed used if \code{prior = "sobol"}. Ignored otherwise.}

\item{\code{overwrite}}{Boolean operator to overwrite existing run result folders.  Default is \code{FALSE}.}

\item{\code{algorithm}}{The algorithm to use for the run.  Default is "NPAG". Alternatives: "NPOD".}

\item{\code{report}}{If missing, the default Pmetrics report template as specified in \link{getPMoptions}
is used. Otherwise can be "plotly", "ggplot", or "none".}

\item{\code{artifacts}}{Default is \code{TRUE}.  Set to \code{FALSE} to suppress creating the \code{etc} folder. This folder
will contain all the compilation artifacts created during the compilation and run steps.}

\item{\code{intern}}{Run NPAG in the R console without a batch script.  Default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function will fit the model contained in the \code{model} field to the data
in the \code{data} field, using the engine specified by the \code{engine} argument and
modified by other arguments. If all function arguments are default,
the simplest execution of this method is
\verb{$run()}.
}

\subsection{Returns}{
A successful run will result in creation of a new folder in the working
directory with the results inside the folder.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_fit-save"></a>}}
\if{latex}{\out{\hypertarget{method-PM_fit-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_fit object to a .rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_fit$save(file_name = "PMfit.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be created. The
default is "PMfit.rds".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_fit-load"></a>}}
\if{latex}{\out{\hypertarget{method-PM_fit-load}{}}}
\subsection{Method \code{load()}}{
\code{PM_fit} objects contain a \code{save} method which invokes \link{saveRDS} to write
the object to the hard drive as an .rds file. This is the corresponding load
function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_fit$load(file_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be read, the default is "PMfit.rds".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{PM_fit} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_fit-check"></a>}}
\if{latex}{\out{\hypertarget{method-PM_fit-check}{}}}
\subsection{Method \code{check()}}{
Checks for errors in data and model objects and agreement between them.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_fit$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_fit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_fit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
