% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_manual.R
\name{PM_manual}
\alias{PM_manual}
\title{Open user and function manuals.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
PM_manual()
}
\description{
Opens the Pmetrics User reference online
}
\details{
Help for Pmetrics.
}
