% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{PM_model_list}
\alias{PM_model_list}
\title{Compile PM_model object}
\description{
Compile PM_model object

Compile PM_model object
}
\note{
This function can only be used with the Rust backend. If the backend is not set to "rust",
an error will be thrown.
}
\examples{

## ------------------------------------------------
## Method `PM_model_list$compile`
## ------------------------------------------------

\dontrun{
model$compile()
}

}
\section{Super class}{
\code{\link[Pmetrics:PM_model]{Pmetrics::PM_model}} -> \code{PM_model_list}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_model_list-new}{\code{PM_model_list$new()}}
\item \href{#method-PM_model_list-write}{\code{PM_model_list$write()}}
\item \href{#method-PM_model_list-write_rust}{\code{PM_model_list$write_rust()}}
\item \href{#method-PM_model_list-update}{\code{PM_model_list$update()}}
\item \href{#method-PM_model_list-compile}{\code{PM_model_list$compile()}}
\item \href{#method-PM_model_list-clone}{\code{PM_model_list$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Pmetrics" data-topic="PM_model" data-id="get_primary"><a href='../../Pmetrics/html/PM_model.html#method-PM_model-get_primary'><code>Pmetrics::PM_model$get_primary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Pmetrics" data-topic="PM_model" data-id="plot"><a href='../../Pmetrics/html/PM_model.html#method-PM_model-plot'><code>Pmetrics::PM_model$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Pmetrics" data-topic="PM_model" data-id="print"><a href='../../Pmetrics/html/PM_model.html#method-PM_model-print'><code>Pmetrics::PM_model$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model_list-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model_list-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model_list$new(model_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model_list-write"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model_list-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model_list$write(model_path = "genmodel.txt", engine = "npag")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model_list-write_rust"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model_list-write_rust}{}}}
\subsection{Method \code{write_rust()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model_list$write_rust(file_name = "parsed_model.txt")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model_list-update"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model_list-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model_list$update(changes_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model_list-compile"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model_list-compile}{}}}
\subsection{Method \code{compile()}}{
Compiles a PM_model object using the Rust backend.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model_list$compile()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This function compiles a PM_model object into a binary format using the Rust backend.
It writes the model to a temporary file, compiles it, and stores the path to the compiled binary.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
model$compile()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model_list-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model_list-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model_list$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
