% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{PM_range}
\alias{PM_range}
\title{Range for primary parameter values}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Define a range for primary model parameter values.
This is a private class used internally by the \code{PM_model} class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{min}}{Minimum value of the range.}

\item{\code{max}}{Maximum value of the range.}

\item{\code{mean}}{Mean value of the range, calculated as (min + max) / 2.}

\item{\code{sd}}{Standard deviation of the range, calculated as (max - min) / 6.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_range-new}{\code{PM_range$new()}}
\item \href{#method-PM_range-print}{\code{PM_range$print()}}
\item \href{#method-PM_range-clone}{\code{PM_range$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_range-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_range-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new range object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_range$new(min, max)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min}}{Minimum value of the range.}

\item{\code{max}}{Maximum value of the range.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_range-print"></a>}}
\if{latex}{\out{\hypertarget{method-PM_range-print}{}}}
\subsection{Method \code{print()}}{
Print the range.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_range$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_range-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_range-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_range$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
