% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_result}
\alias{PM_result}
\title{Results of a Pmetrics run}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This object contains all of the results after a Pmetrics runs. It is created
by using the \link{PM_load} function.
}
\details{
After a run completes, results are stored on your hard drive. They are loaded
back into R with \link{PM_load} to create the \link{PM_result} object, which contains both
the results and functions to analyze or plot the result.
}
\examples{

## ------------------------------------------------
## Method `PM_result$mmopt`
## ------------------------------------------------

library(PmetricsData)
mmopt1 <- NPex$mmopt(data = exData, limits = NA)
mmopt1$summary()
mmopt1$plot()

}
\author{
Michael Neely, Julian Otalvaro
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{NPdata}}{List with all output from NPAG, made by \link{NPparse}}

\item{\code{ITdata}}{List with all output from IT2B, made by \link{ITparse}}

\item{\code{pop}}{A \link{PM_pop} object}

\item{\code{post}}{A \link{PM_post} object}

\item{\code{final}}{A \link{PM_final} object}

\item{\code{cycle}}{A \link{PM_cycle} object}

\item{\code{op}}{A \link{PM_op} object}

\item{\code{cov}}{A \link{PM_cov} object}

\item{\code{data}}{\link{PM_data} object representing the original .csv data file used in the run. The predictions contained in the \verb{$data} fields from \verb{$pop} and \verb{$post} will be added to this \link{PM_data} object to permit easy addition of such predictions to raw data plots. See \link{plot.PM_data} for more details.}

\item{\code{model}}{text string representing the original model file used in the run}

\item{\code{errfile}}{Name of error file if it exists}

\item{\code{success}}{Boolean if successful run}

\item{\code{valid}}{If the \verb{$validate} method has been executed after a run, this object will be added to
the \code{PM_result} object.  It contains the information required to plot visual predictive checks and normalized prediction
error discrepancies via the npde code developed by Comets et al. Use the
\verb{$save} method on the augmented \code{PM_result} object to save it with the
new validation results.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_result-new}{\code{PM_result$new()}}
\item \href{#method-PM_result-plot}{\code{PM_result$plot()}}
\item \href{#method-PM_result-summary}{\code{PM_result$summary()}}
\item \href{#method-PM_result-auc}{\code{PM_result$auc()}}
\item \href{#method-PM_result-nca}{\code{PM_result$nca()}}
\item \href{#method-PM_result-sim}{\code{PM_result$sim()}}
\item \href{#method-PM_result-save}{\code{PM_result$save()}}
\item \href{#method-PM_result-validate}{\code{PM_result$validate()}}
\item \href{#method-PM_result-step}{\code{PM_result$step()}}
\item \href{#method-PM_result-mmopt}{\code{PM_result$mmopt()}}
\item \href{#method-PM_result-load}{\code{PM_result$load()}}
\item \href{#method-PM_result-clone}{\code{PM_result$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with data from previous run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$new(out, quiet = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out}}{The parsed output from \link{PM_load}, which is
automatically generated. This is not a user-modifiable.}

\item{\code{quiet}}{Quietly validate. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creation of new \code{PM_result} objects is via \link{PM_load}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-plot}{}}}
\subsection{Method \code{plot()}}{
Plot generic function based on type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of plot based on class of object}

\item{\code{...}}{Plot-specific arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-summary}{}}}
\subsection{Method \code{summary()}}{
Summary generic function based on type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$summary(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of summary based on class of object}

\item{\code{...}}{Summary-specific arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-auc"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-auc}{}}}
\subsection{Method \code{auc()}}{
AUC generic function based on type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$auc(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of AUC based on class of object}

\item{\code{...}}{Summary-specific arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-nca"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-nca}{}}}
\subsection{Method \code{nca()}}{
Perform non-compartmental analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$nca(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeNCA}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{makeNCA}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-sim"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-sim}{}}}
\subsection{Method \code{sim()}}{
Simulates using the self$final object.
For parameter information refer to \link{SIMrun}. It will return a \code{PM_sim} object
by running \link{SIMparse} at the end of the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$sim(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to \link{SIMrun}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-save"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_result object to an .Rdata file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$save(run, file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run}}{The run output folder number to save the revised result. If missing,
will save in the current working directory. For example, if folder "1" is in
your current working directory, specify \code{run = 1} to save the result to the "outputs"
subfolder of the "1" folder.}

\item{\code{file}}{Custom file name. Default is "PMout.Rdata".}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This is useful if you have updated the result in some way, for example you
have run the \verb{$make_valid} method on the \code{PM_result} object, which returns
an internal simulation based validation as a new \code{valid} field. To save this
validation, use this \verb{$save} method. Note that unless a \code{file} name is provided,
the changes will overwrite the
previous run results, although unchanged items will be preserved. This is the
usual workflow. However, a custom file name may be useful to share the run
results with someone.

The saved object is an .Rdata file. When loaded, it should be assigned to an R
object, e.g. \code{run2 <- PM_result$new("filename")}. An equivalent statement would
be \code{run2 <- PM_load(file = "filename")}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the result by internal simulation methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$validate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{make_valid}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-step"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-step}{}}}
\subsection{Method \code{step()}}{
Conduct stepwise linear regression of Bayesian posterior parameter values
and covariates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$step(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{PMstep}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-mmopt"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-mmopt}{}}}
\subsection{Method \code{mmopt()}}{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Method to compute 1 or more MM-optimal sampling times.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$mmopt(
  data = NULL,
  nsamp = 1,
  weight = list(none = 1),
  predInt = 0.5,
  mmInt = NULL,
  outeq = 1,
  clean = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{One of two choices which will provide the template dosage and sampling regimens:
\itemize{
\item \link{PM_data} object
\item Character vector with the filename of a Pmetrics data file
that contains template regimens and observation times.
}

If \code{data} is missing, the default value of "data.csv" will be used and Pmetrics will try to open a so-named filed in teh current working directory. In either choice, the value for outputs
can be coded as any number(s) other than -99.  The number(s) will be replaced in the simulator output with the simulated values.}

\item{\code{nsamp}}{The number of MM-optimal sample times to compute; default is 1, but can be any number.  Values >4 will take an exponentially longer time.}

\item{\code{weight}}{List whose names indicate the type of weighting, and values indicate the relative weight. Values should sum to 1.  Names can be any of the following:
\itemize{
\item \strong{none} The default. MMopt times will be chosen to maximally discriminate all responses at all times.
\item \strong{AUC} MMopt times will be chosen to maximally discriminate AUC, regardless of the shape of the response profile.
\item \strong{max} MMopt times will be chosen to maximally discriminate maximum, regardless of the shape of the response profile.
\item \strong{min} MMopt times will be chosen to maximally discriminate minimum, regardless of the shape of the response profile.
}

Any combination of AUC, max, and min can be chosen.  If "none" is specified, other
weight types will be ignored and the relative value will be set to 1.
For example,\code{list(auc = 0.5,max = 0.5)} or \code{list(auc = 0.2, min = 0.8)}.
The default is \code{list(none = 1)}.}

\item{\code{predInt}}{The interval in fractional hours for simulated predicted outputs at times other than those specified in the template \code{data}.
The default is 0.5, which means there will be simulated outputs every 30 minutes from time 0 up
to the maximal time in the template file.  You may also specify \code{predInt}
as a vector of 3 values, e.g. \code{c(1,4,1)}, similar to the R command \link{seq}, where the
first value is the start time, the second is the stop time, and the third is the
step value.  Outputs for times specified in the template file will also be simulated.
To simulate outputs \emph{only} at the output times in the template data (i.e. EVID=0 events), use \code{predInt = 0}.
Note that the maximum number of predictions total is 594, so the interval must be sufficiently large to accommodate this for a given
number of output equations and total time to simulate over.  If \code{predInt} is set so that this cap is exceeded, predictions will be truncated.}

\item{\code{mmInt}}{Specify the time intervals from which MMopt times can be selected.
These should only include simulated times specified by \code{predInt}.}

\item{\code{outeq}}{Output equation to optimize}

\item{\code{clean}}{Boolean parameter to specify whether temporary files made in the
course of the simulation run should be deleted. Defaults to \code{TRUE}.
This is primarily used for debugging.}

\item{\code{...}}{Other parameters to pass to \link{SIMrun}, which are not usually necessary.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Based on the multiple-model optimization algorithm, this method computes \emph{nsamp} times where the concentration time profiles are the most separated, thereby minimizing the risk of choosing the incorrect Bayesian posterior for an individual. The algorithm was published as
Bayard, David S., and Michael Neely. "Experiment Design for Nonparametric
Models Based on Minimizing Bayes Risk: Application to Voriconazole."
Journal of Pharmacokinetics and Pharmacodynamics 44, no. 2 (April 2017):
95–111. https://doi.org/10.1007/s10928-016-9498-5.
}

\subsection{Returns}{
A object of class \emph{MM_opt} with 3 items.
\itemize{
\item \strong{sampleTime} The MM-optimal sample times
\item \strong{bayesRisk} The Bayesian risk of mis-classifying a subject based on the sample times.  This
is more useful for comparisons between sampling strategies, with minimization the goal.
\item \strong{simdata} A \link{PM_sim} object with the simulated profiles
\item \strong{mmInt} A list with the \code{mmInt} values, \code{NULL} if \code{mmInt} is missing.
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(PmetricsData)
mmopt1 <- NPex$mmopt(data = exData, limits = NA)
mmopt1$summary()
mmopt1$plot()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-load"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-load}{}}}
\subsection{Method \code{load()}}{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Deprecated method to load prior results saved with the \verb{$save} method.
Replaced by \link{PM_load}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$load(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_result-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_result-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
