% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_tree.R
\name{PM_tree}
\alias{PM_tree}
\alias{PMtree}
\title{Create a new Pmetrics folder tree}
\usage{
PM_tree(project = "NewProject", path = getwd())

PMtree(...)
}
\arguments{
\item{project}{A character string of a new project name, e.g. "DrugX"}

\item{path}{The full path to the root folder for the new project.  Default is the
current working directory.}

\item{...}{Arguments passed to \link{PM_tree}.}
}
\value{
A new folder named \code{project} with the following subfolders:
\item{Rscript }{The folder for the Rscript containing all run instructions.
Within this folder will be a skeleton R script for the project.}
\item{Runs }{The folder for all Pmetrics runs.  Put run files, i.e. a data file and a
model file in this directory prior to each run.}
\item{Sim }{The folder for all simulations related to the project.}
\item{src }{The folder for source data files in their original format, to preserve
integrity and for audit purposes.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Sets up a directory tree for a new Pmetrics project
}
\details{
This function will create a new project folder tree with appropriate subfolders and a skeleton R script.
}
\examples{
\dontrun{
PM_tree("DrugX")
}
}
\seealso{
\link{PM_manual}
}
\author{
Michael Neely
}
