% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_tutorial.R
\name{PM_tutorial}
\alias{PM_tutorial}
\title{Introductory tutorial to Pmetrics}
\usage{
PM_tutorial(wd = getwd())
}
\arguments{
\item{wd}{The working directory in which to create the Examples folder.
Default is current working directory.}
}
\value{
NULL.
}
\description{
This function will create an Examples folder in the current working
directory. The folder will contain all files and folder necessary to
conduct the tutorial. After the Examples folder is created, open the
examples.R script to begin the tutorial.
}
