% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_valid.R
\name{PM_valid}
\alias{PM_valid}
\title{Pmetrics validation object}
\description{
Contains results of internal validation by simulation to permit generation of
visual predictive checks (VPCs), prediction corrected visual predictive checks,
(pcVPCs), normalized prediction distribution errors (NPDE), and
numerical predictive checks.
}
\details{
This object is created by running the \code{make_valid} method in a
\link{PM_result} object. It contains all the information necessary
to internally validate the result by simulation methods.
}
\seealso{
\link{PM_result}, \link{makeValid}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simdata}}{Simulated data created in the validation process}

\item{\code{timeBinMedian}}{Median times for cluster bins}

\item{\code{tadBinMedian}}{Median times after previous doses for cluster bins}

\item{\code{opDF}}{Observed-predicted data frame}

\item{\code{npde}}{Data for Normalized Prediction Distribution Error}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PM_valid$new()}}
\item \href{#method-plot}{\code{PM_valid$plot()}}
\item \href{#method-clone}{\code{PM_valid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new PM_valid object from a PM_result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_valid$new(result, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{The PM_result object}

\item{\code{...}}{Additional arguments to ultimately pass to makeValid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method. Calls \link{plot.PMvalid}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_valid$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments to pass to [plot.PMvalid].}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_valid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
