% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMbuild}
\alias{PMbuild}
\title{Build Fortran}
\usage{
PMbuild(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Building is no longer required; models are compiled just-in-time at run time.
}
\author{
Michael Neely
}
\keyword{internal}
