% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmetrics-package.R
\name{Pmetrics-package}
\alias{Pmetrics-package}
\alias{Pmetrics}
\title{Parametric and non-parametric modeling and simulation of pharmacokinetic-pharmacodynamic systems.}
\description{
This package contains functions to run and analyze the output from all three components
of the Pmetrics software suite for population pharmacometric data analysis:
\enumerate{
\item IT2B (Iterative Two-Stage Bayesian) for parametric models; 2) NPAG (Non-parametric
Adaptive Grid) for non-parametric models; 3) Simulator for semi-parametric Monte-Carlo
simulations.
}
}
\author{
Michael Neely MD
\url{http://www.lapk.org}
}
\keyword{"_PACKAGE"}
\keyword{internal}
