% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{PMex3}
\alias{PMex3}
\title{Pmetrics data file with errors}
\format{badData is a PMmatrix object as read by \code{\link{PMreadMatrix}}}
\usage{
PMex3
}
\description{
Example dataset for an NPAG run, which has been corrupted with errors.
}
\details{
Errors include missing covariate on first line for subject 1, alphanumeric covariate for
gender, and trailing dose for subject 1.
}
\author{
Michael Neely
}
\keyword{datasets}
