% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMfixMatrix}
\alias{PMfixMatrix}
\title{Fix data matrix}
\usage{
PMfixMatrix(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use PM_data$new() and PM_model$new() instead.
}
\author{
Michael Neely
}
\keyword{internal}
