% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMaddCRCL.R
\name{PMgetCRCL}
\alias{PMgetCRCL}
\title{Add Jelliffe Creatinine Clearance}
\usage{
PMgetCRCL(
  mdata,
  idCol = "id",
  wtCol = "wt",
  maleCol = "male",
  ageCol = "age",
  scrCol = "scr",
  SI = F
)
}
\arguments{
\item{mdata}{A \link{PM_data} data object}

\item{idCol}{A character vector with the name
of the id column in \code{mdata}. The default is \dQuote{id}.}

\item{wtCol}{A character vector with the name of the weight
column in \code{data}.  The default is \dQuote{wt}.}

\item{maleCol}{A character vector with the name of the gender column in \code{mdata}.
Male should be 1 and female should be 0. The default is \dQuote{male}.}

\item{ageCol}{A character vector with the name of the age column in \code{mdata}.
The default is \dQuote{age}.}

\item{scrCol}{A character vector with the name of the serum creatinine column in \code{mdata}.
Default units are mg/dL, and the the default name is \dQuote{scr}.}

\item{SI}{Boolean value, if true, will expect serum creatinine to be in micromol/L.
Default is \code{FALSE}.}
}
\value{
A vector of length \code{nrow(mdata)} with Jelliffe CRCL values for every dose in \code{mdata}. Vector values
for observation events in \code{mdata} are NA.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Gets creatinine clearance as estimated by the Jelliffe equation.
}
\details{
The equation depends on age, sex, weight, and serum creatinine.
ESS = wt * (29.3 - (0.203 * age)) for males
ESS = wt * (25.1 - (0.175 * age)) for females
scrAve = (Scr1 + Scr2) / 2
ESS_cor = ESS * (1.035 - (0.0337 * scrAve))
E = ESS_cor - 4 * wt * (Scr2 - Scr1) / (time2 - time1)
CRCL = E / (14.4 * scrAve) in ml/min/1.73m^2
}
\author{
Michael Neely
}
