% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMremote.R
\name{PMlogout}
\alias{PMlogout}
\title{Login a user in the remote server}
\usage{
PMlogout(server_address)
}
\arguments{
\item{server_address}{remote server address. If omitted,  Pmetrics will use the the default one. See \code{\link{setPMoptions}}.}
}
\value{
A message that specifies if the user was successfuly logged out, or an explanation of the error.
}
\description{
Logout a logged in user onto the server.
}
\details{
If no server_address is provided Pmetrics will use the the default one. See \code{\link{setPMoptions}}.
}
\author{
Michael Neely
}
