% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMmanual.R
\name{PMmanual}
\alias{PMmanual}
\title{Open user and function manuals.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
PMmanual()
}
\description{
Opens the Pmetrics User reference online
}
\details{
Help for Pmetrics.
}
