% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMmb2csv.R
\name{PMmb2csv}
\alias{PMmb2csv}
\title{Convert Old .mb or USC*PACK Files to .csv Matrix File}
\usage{
PMmb2csv(oldFiles, newFile = "data")
}
\arguments{
\item{oldFiles}{A character vector of files in the current working directory to convert.
This could be easily obtained with \code{\link{list.files}}.}

\item{newFile}{A single character vector with the basename (without any file extension) of the new file to be created.}
}
\value{
A new file will be created with the name equal to \code{newFile} and
an extension of "csv".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Convert old style, single drug .mb or USC*PACK files into
a \link{PM_data} object and write a .csv file to the current working directory.

IDs will be suffixed with .1 to .9 for <10 subjects, .01 to .99 for <100 subjects and .001 to .999 for <1000 subjects,
as needed to ensure unique ID numbers.
}
\author{
Michael Neely
}
