% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMnews.R
\name{PMnews}
\alias{PMnews}
\title{Pmetrics changelog}
\usage{
PMnews(PMversion = packageVersion("Pmetrics"))
}
\arguments{
\item{PMversion}{Default is the current version, otherwise a character string with the starting version you wish to see up to
the current, e.g. \dQuote{0.21}.  Use \dQuote{all} for all versions.}
}
\value{
The changelog for the requested version.
}
\description{
See changelog for Pmetrics
}
\examples{
PMnews()
}
\author{
Michael Neely
}
