% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMreadDefaults}
\alias{PMreadDefaults}
\title{Read Defaults}
\usage{
PMreadDefaults(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use getPMoptions() instead.
}
\author{
Michael Neely
}
\keyword{internal}
