% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMremote.R
\name{PMregister}
\alias{PMregister}
\title{Register a new user in the remote server}
\usage{
PMregister(email, server_address)
}
\arguments{
\item{email}{User's email.}

\item{server_address}{remote server address. If omitted,  Pmetrics will use the the default one. See \code{\link{setPMoptions}}.}
}
\value{
A message that specifies if the user was registered or if some error happened.
}
\description{
Creates a new user in the remote server
}
\details{
This function receives an unused email and optionally the server adress.
}
\author{
Michael Neely
}
