% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMreport.R
\name{PMreport}
\alias{PMreport}
\title{Summarize NPAG or IT2B Run}
\usage{
PMreport(wd, rdata, icen = "median", type = "NPAG", parallel = F)
}
\arguments{
\item{wd}{The working directory containing the NP_RFxxxx.TXT or IT_RFxxxx.TXT file}

\item{rdata}{The processed output of an IT2B or NPAG run, depending on local or server runs.}

\item{icen}{Median (default), mean or mode of Bayesian posterior to be used to calculate predictions.}

\item{type}{\dQuote{NPAG} (default) or \dQuote{IT2B} report type}

\item{parallel}{Boolean parameter which indicates the type of run done.  Default is \code{FALSE} for serial.}
}
\value{
Several files are placed in the \code{wd}
\item{NPAGreport.html or IT2Breport.html }{An .html file containing a summary of all the results}
\item{poppoints.csv }{NPAG only: A .csv file containing the population support points and probabilities}
\item{poparam.csv }{A .csv file containing a summary of the population parameter values, including
mean, standard deviation, coefficient of variation, variance, and median}
\item{popcor.csv }{A .csv file containing the population parameter correlation matrix}
\item{popcov.csv }{A .csv file containing the population parameter covariance matrix}
\item{cycle.pdf }{A .pdf file containing the run cycle information (see \code{\link{plot.PMcycle}})}
\item{cycle.png }{A thumbnail of the run cycle information for the .html file}
\item{final.pdf }{A .pdf file containing the population final cycle information (see \code{\link{plot.PMfinal}})}
\item{final.png }{A thumbnail of the population final cycle information for the .html file}
\item{opx.pdf }{One or more .pdf files, where \emph{x} is the number of the output equation, each containing
two observed vs. predicted plots: population and individual Bayesian posterior predictions (see \code{\link{plot.PMop}})}
\item{opx.png }{One or more thumnails of the observed vs. predicted plots for the .html file}
\item{NPAGout.Rdata or IT2Bout.Rdata }{An R data file containing the output of \code{\link{NPparse}} or \code{\link{ITparse}}, \code{\link{makeFinal}},
\code{\link{makeCycle}}, \code{\link{makeOP}}, \code{\link{makeCov}}, \code{\link{makePop}}, \code{\link{makePost}}, and
the data file for the run read by \code{\link{PMreadMatrix}}.
This file can be loaded using \code{\link{PMload}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Generates a summary of a Pmetrics NPAG or IT2B run
}
\details{
Creates an HTML page and several files summarizing an NPAG or IT2B run.  This report is generated
automatically at the end of a successful run.
}
\author{
Michael Neely
}
