% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMsave.R
\name{PMsave}
\alias{PMsave}
\title{Save Pmetrics objects}
\usage{
PMsave(run, ..., quiet = F)
}
\arguments{
\item{run}{The numerical value of the run number of the objects to be saved.
This parameter must be specified, as it also determines where to save the revised output.}

\item{\dots}{Additional objects to be saved, which do not need to be suffixed with the run number,
e.g. var1, var2, var3.}

\item{quiet}{Suppress written report.  Default is \code{FALSE}.}
}
\description{
Saves Pmetrics objects
}
\details{
Any objects that are made during the course of analysis in R can be added to the saved data
that are automatically generated at the end of an NPAG or IT2B run and loaded with \code{\link{PMload}}.
Objects with the same run number will be saved as a group.  So if a user has made a new object called
lm.1 that contains regressions related to run 1, it will be saved with any other object
that also has .1 at the end.

Additionally, other objects can be saved via the \dots argument.  For exmaple PMsave(1,lm) will
save any object with .1 at the end, plus an object named "lm".  All objects will be suffixed
with the run number when loaded back with \code{\link{PMload}}.
}
\seealso{
\code{\link{PMload}}
}
\author{
Michael Neely
}
