% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMstep.R
\name{PMstep}
\alias{PMstep}
\title{Stepwise covariate-parameter regressions}
\usage{
PMstep(x, icen = "median", direction = "backward")
}
\arguments{
\item{x}{A PMcov object loaded by \code{\link{PMload}}
or made by \code{\link{makeCov}}.}

\item{icen}{A character vector to summarize covariate values.  Default is \dQuote{median}, but can also be
\dQuote{mean}.}

\item{direction}{The direction for covariate elmination can be \dQuote{backward}, \dQuote{forward}, or \dQuote{both}.
\emph{backward} is the default.}
}
\value{
A matrix with covariates in the rows and parameters in the columns.  Values for the matrix are the multi-variate P-values.
A value of \code{NA} indicates that the variable was not retained in the final model.
}
\description{
Perform a stepwise linear regression on all covariates and Bayesian posterior parameters
}
\details{
This function will perform stepwise linear regressions on a PMcov object loaded by \code{\link{PMload}},
or made by \code{\link{makeCov}}.  Every covariate in the model will be tested in a stepwise linear regression for their relationships
to each parameter in the model.  Bayesian posterior parameters and individual covariates are used.
}
\seealso{
\code{\link{step}}
}
\author{
Michael Neely
}
