% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMtest.R
\name{PMtest}
\alias{PMtest}
\title{Test Pmetrics}
\usage{
PMtest()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Check Pmetrics fortran installation by trying to compile sample files
}
\author{
Michael Neely
}
