% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMupdate.R
\name{PMupdate}
\alias{PMupdate}
\title{Download and install Pmetrics updates}
\usage{
PMupdate(force = F)
}
\arguments{
\item{force}{Boolean operator to force downloading and installing.  Default is false.}
}
\value{
The latest system-specific Pmetrics update will be downloaded to a temporary
folder and then installed.  You need to restart R (Rstudio) and then reload Pmetrics with
the \code{library(Pmetrics)} command to complete the installation.
}
\description{
Download and install Pmetrics updates from LAPK website
}
\author{
Michael Neely
}
