% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMwriteMatrix.R
\name{PMwriteMatrix}
\alias{PMwriteMatrix}
\title{Write a Pmetrics .csv Matrix File}
\usage{
PMwriteMatrix(data, filename, override = F, version = "DEC_11", header = T)
}
\arguments{
\item{data}{Must be a data.frame with appropriate structure (see \code{\link{PMcheck}}.}

\item{filename}{Name of file to create.}

\item{override}{Boolean operator to write even if errors are detected.  Default is \code{False}.}

\item{version}{Which matrix data format version to write.  Default is the current version.}

\item{header}{Is there a header row? Default is \code{TRUE}.}
}
\value{
Returns the error report (see \code{\link{PMcheck}} for details).
}
\description{
\code{PMwriteMatrix} is the companion function to \code{\link{PMreadMatrix}}.
It will write an appropriate R data object to a formatted .csv file.
This is now largely an internal function accessed with the \verb{$write()} method
for \link{PM_data} objects.
}
\details{
\code{PMwriteMatrix} will first run \code{\link{PMcheck}} to determine
if there are any errors in the structure of  \code{data}.  If the error check
fails, the file will not be written and a message will be printed on the console.
}
\examples{
\dontrun{
#write to the current directory
NPex$data$write("data.csv")
}
}
\seealso{
\code{\link{PMcheck}}, \code{\link{PMreadMatrix}}
}
\author{
Michael Neely
}
