% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMMopt.R
\name{print.MMopt}
\alias{print.MMopt}
\title{Print Pmetrics Multiple-Model Optimal Sampling Objects}
\usage{
\method{print}{MMopt}(x, ...)
}
\arguments{
\item{x}{The name of an \emph{MMopt} data object generated by \link{MM_opt}}

\item{...}{Not used}
}
\value{
Prints the optimal sampling times and Bayes Risk.
}
\description{
Print \emph{MMopt} objects
}
\details{
Simulated observations are plotted on the y-axis vs. time on the x.axis.
Optimal sampling times are indicated as vertical lines.
}
\seealso{
\link{MM_opt}
}
\author{
Michael Neely
}
