% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_compare.R
\name{print.PM_compare}
\alias{print.PM_compare}
\title{Print method for PM_compare objects}
\usage{
\method{print}{PM_compare}(x, ...)
}
\arguments{
\item{x}{A PM_compare object.}

\item{...}{Additional arguments (not used).}
}
\description{
\verb{ r lifecycle::badge("stable")}
Prints a summary of the PM_compare object.
}
