% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPMdopt.R
\name{print.PMdopt}
\alias{print.PMdopt}
\title{Print PMdopt}
\usage{
\method{print}{PMdopt}(x, ...)
}
\arguments{
\item{x}{A PMdopt object made by \code{\link{Dopt}}.}

\item{\dots}{Other parameters which are not necessary.}
}
\value{
A printed object.
}
\description{
Print a Pmetrics PMdopt Object
}
\details{
Print a Pmetrics D-optimal object, made by \code{Dopt}.
}
\seealso{
\code{\link{Dopt}}
}
\author{
Michael Neely
}
