% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPMerr.R
\name{print.PMerr}
\alias{print.PMerr}
\title{Print Data Errors}
\usage{
\method{print}{PMerr}(x, ...)
}
\arguments{
\item{x}{A PMerr object made by validation in \link{PM_data}.}

\item{...}{Other parameters which are not necessary.}
}
\value{
A printed object.
}
\description{
Print a Pmetrics Error Object
}
\details{
Print the errors in a Pmetrics data file or \link{PM_data} object.
}
\seealso{
\link{PM_data}
}
\author{
Michael Neely
}
