% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPMnpc.R
\name{print.PMnpc}
\alias{print.PMnpc}
\title{Print NPC}
\usage{
\method{print}{PMnpc}(x, ...)
}
\arguments{
\item{x}{A PMnpc object made by \code{\link{plot.PMsim}}.}

\item{\dots}{Other parameters which are not necessary.}
}
\value{
A printed object.
}
\description{
Print a Pmetrics NPC Object
}
\details{
Print a Pmetrics NPC (Numerical Predictive Check) object, made by \code{plot.PMsim}.
}
\seealso{
\code{\link{plot.PMsim}}
}
\author{
Michael Neely
}
