% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_cycle.R
\name{print.summary.PM_cycle}
\alias{print.summary.PM_cycle}
\title{Print Summary of Cycle Statistics}
\usage{
\method{print}{summary.PM_cycle}(x, ...)
}
\arguments{
\item{x}{An object made by \link{summary.PM_cycle}.}

\item{...}{Not used.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print a Pmetrics Cycle Summary Object
}
\details{
Print a summary a summary.PM_cycle object
made by \link{summary.PM_cycle}. Users do not normally need to call this
function directly, as it will be the default method to display the object.
}
\examples{
#'
\dontrun{
NPex$cycle$summary()
}
}
\seealso{
\link{summary.PM_cycle}
}
\author{
Michael Neely
}
