% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_data.R
\name{print.summary.PM_data}
\alias{print.summary.PM_data}
\title{Print Summary of Pmetrics Data}
\usage{
\method{print}{summary.PM_data}(x, ...)
}
\arguments{
\item{x}{An object made by \link{summary.PM_data}.}

\item{...}{Not used.}
}
\value{
A printed object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Print the summary of \link{PM_data} object.

Summarize the raw data used for a Pmetrics run.
}
\examples{
\dontrun{
dataEx$summary()
}
}
\seealso{
\link{summary.PM_data}
}
\author{
Michael Neely
}
