% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_final.R
\name{print.summary.PM_final}
\alias{print.summary.PM_final}
\title{Print Summary of Parameter Values and Credibility Intervals}
\usage{
\method{print}{summary.PM_final}(x, digits = 3, ...)
}
\arguments{
\item{x}{A summary.PM_final object made by \link{summary.PM_final}.}

\item{digits}{Integer, used for number of digits to print.}

\item{...}{Not used.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print a Pmetrics Final Summary Object
}
\details{
Print a summary of parameter medians and MAWD, with point estimates and credibilty intervals
from an object made by \link{summary.PM_final}. Users do not normally need to call this
function directly, as it will be the default method to display the object.
}
\examples{
\dontrun{
NPex$final$summary
}
}
\seealso{
\link{summary.PM_final}
}
\author{
Michael Neely
}
