% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_op.R
\name{print.summary.PM_op}
\alias{print.summary.PM_op}
\title{Print Summary of Observations and Predictions}
\usage{
\method{print}{summary.PM_op}(x, digits = getPMoptions("digits"), embed = FALSE, ...)
}
\arguments{
\item{x}{An object made by \link{summary.PM_op}.}

\item{digits}{Integer, used for number of digits to print.}

\item{embed}{If \code{TRUE}, will embed the summary in the R Markdown document. Default is \code{FALSE}.}

\item{...}{Not used.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print a Pmetrics Observed vs. Predicted Summary Object
}
\details{
Print a summary of observations, predictions and errors in a \link{summary.PM_op} object
made by \link{summary.PM_op}. Users do not normally need to call this
function directly, as it will be the default method to display the object.
}
\examples{
\dontrun{
NPex$op$summary()
}
}
\seealso{
\link{summary.PM_op}
}
\author{
Michael Neely
}
