% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_sim.R
\name{print.summary.PM_sim}
\alias{print.summary.PM_sim}
\title{Print Summary of Simulations}
\usage{
\method{print}{summary.PM_sim}(x, ...)
}
\arguments{
\item{x}{An object made by \link{summary.PM_sim}.}

\item{...}{Not used.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print a Pmetrics Observed vs. Predicted Summary Object
}
\details{
Print a summary of simulations
made by \link{summary.PM_sim}. Users do not normally need to call this print
function directly, as it will be the default method to display the object.
}
\examples{
\dontrun{
simEx$summary()
}
}
\seealso{
\link{summary.PM_sim}
}
\author{
Michael Neely
}
