% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSummaryPMfinal.R
\name{print.summary.PMfinal}
\alias{print.summary.PMfinal}
\title{Print Summary of Parameter Values and Credibility Intervals}
\usage{
\method{print}{summary.PMfinal}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A summary.PMfinal object made by \code{\link{summary.PMfinal}}.}

\item{digits}{Integer, used for number of digits to print.}

\item{\dots}{Other parameters which are not necessary.}
}
\value{
A printed object.
}
\description{
Print a Pmetrics Final Summary Object
}
\details{
Print a summary of parameter medians and MAWD, with point estimates and credibilty intervals
from a summary.PMfinal object made by \code{\link{summary.PMfinal}}.
}
\seealso{
\code{\link{summary.PMfial}}
}
\author{
Michael Neely
}
