% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSummaryPMmatrix.R
\name{print.summary.PMmatrix}
\alias{print.summary.PMmatrix}
\title{Summarize Covariates and Bayesian Posterior Parameter Values}
\usage{
\method{print}{summary.PMmatrix}(x, ...)
}
\arguments{
\item{x}{An object made by \link{summary.PMmatrix}, aliased by \link{summary.PM_data}.}

\item{...}{Not used.}
}
\value{
A formatted printing of a \link{summary.PMmatrix} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Print the summary of \link{PM_data} object.

Summarize the raw data used for a Pmetrics run.
}
\seealso{
\link{summary.PMmatrix}, \link{summary.PM_data}
}
\author{
Michael Neely
}
