% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData.R
\name{read.npdeData,NpdeData-method}
\alias{read.npdeData,NpdeData-method}
\title{Read observed data}
\usage{
\S4method{read.npdeData}{NpdeData}(object, header, sep, na.strings, detect, verbose)
}
\arguments{
\item{header}{boolean indicating whether the file has a header (mandatory if 
detect is TRUE)}

\item{sep}{field separator (for files on disk)}

\item{na.strings}{strings to be considered as indicating NA}

\item{detect}{a boolean; if TRUE, automatic recognition of names will be attempted to detect necessary items (longitudinal data structure, missing data and censoring information)}
}
\value{
an object of class \code{"\linkS4class{NpdeData}"}
}
\description{
Creates an NpdeData object containing the observed data, either from disk or from a dataframe
}
\keyword{internal}
