% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData.R
\name{read.npdeSimData,NpdeSimData-method}
\alias{read.npdeSimData,NpdeSimData-method}
\title{Read simulated data}
\usage{
\S4method{read.npdeSimData}{NpdeSimData}(object, header, sep, na.strings, verbose)
}
\arguments{
\item{header}{boolean indicating whether the file has a header (mandatory if 
detect is TRUE)}

\item{sep}{field separator (for files on disk)}

\item{na.strings}{strings to be considered as indicating NA}
}
\value{
an object of class NpdeSimData
}
\description{
Creates an NpdeSimData object containing the simulated data, either from disk or from a dataframe
}
\keyword{internal}
