% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_Utils.R
\name{rgba_to_rgb}
\alias{rgba_to_rgb}
\title{Convert RGBA to RGB}
\usage{
rgba_to_rgb(rgba_str, alpha = NULL)
}
\arguments{
\item{rgba_str}{A string in the format \code{rgba(r, g, b, a)} where r, g, and b are
integers between 0 and 255, and a is a float between 0 and 1.}

\item{alpha}{Optional numeric value to replace the alpha channel in the rgba string.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Converts a CSS \code{rgba()} string to an RGB color string.
}
