% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{save_flextable}
\alias{save_flextable}
\title{Save a flextable object to a file}
\usage{
save_flextable(x)
}
\arguments{
\item{x}{A \link[flextable:flextable]{flextable::flextable} object.}
}
\value{
A message indicating the file was saved.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Saves flextable objects to a file based on the \code{file} attribute
in the object, set when the flextable generator function is called.
Allowable file types are 'docx', 'pptx', 'html', 'png', and 'svg'.
}
\keyword{internal}
