% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeObject.R
\name{set.plotoptions}
\alias{set.plotoptions}
\alias{set.plotoptions.NpdeObject}
\alias{set.plotoptions,NpdeData-method}
\alias{set.plotoptions,NpdeObject-method}
\alias{set.plotoptions.NpdeData}
\title{Set graphical preferences}
\usage{
set.plotoptions(object, ...)
}
\arguments{
\item{object}{an object of class NpdeData or NpdeObject}

\item{\dots}{arguments to replace default arguments (currently ignored)}
}
\value{
a list of options for graphs
}
\description{
This function is used to set options for graphs
}
\details{
See documentation for a list of available options.
}
\seealso{
\code{\link{npde}}, \code{\link{autonpde}}
}
\author{
Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\keyword{plot}
