% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{setup_logs}
\alias{setup_logs}
\title{Initialize the tracing subscriber with the custom R formatter}
\usage{
setup_logs()
}
\description{
Initialize the tracing subscriber with the custom R formatter
}
