% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData.R
\name{showall}
\alias{showall}
\alias{showall.NpdeData}
\alias{showall,NpdeData-method}
\alias{showall.NpdeSimData}
\alias{showall,NpdeSimData-method}
\alias{showall.NpdeRes}
\alias{showall.NpdeObject}
\title{Brief summary of an object}
\usage{
\method{showall}{NpdeData}(object)
}
\arguments{
\item{object}{an object}
}
\description{
Prints a brief summary of an object
}
\keyword{print}
