% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{simEx}
\alias{simEx}
\title{Example simulator output}
\format{
R6 \link{PM_sim}
}
\usage{
simEx
}
\description{
Example simulator output
}
\details{
This is an R6 Pmetrics \link{PM_sim} object created by
created by running \verb{$sim()} on a \link{PM_result} object, e.g.
\verb{NPex$sim(include = 1:4, limits = NA, nsim = 100).}
}
\author{
Michael Neely
}
\keyword{datasets}
