% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simulate_all}
\alias{simulate_all}
\title{Simulates all subjects in the data set using the given model.}
\usage{
simulate_all(data_path, model_source, theta, solver = NULL)
}
\arguments{
\item{data_path}{Path to the data file.}

\item{model_source}{Model definition written in the pharmsol DSL.}

\item{theta}{Data frame of support points.}

\item{solver}{Optional ODE solver name.}
}
\value{
Simulation results.
}
\description{
Simulates all subjects in the data set using the given model.
}
