% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{start_live_session}
\alias{start_live_session}
\title{Start a local live reporting session.}
\usage{
start_live_session()
}
\value{
Live session connection info.
}
\description{
Start a local live reporting session.
}
\keyword{internal}
