% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData.R
\name{[,NpdeData,ANY,ANY,ANY-method}
\alias{[,NpdeData,ANY,ANY,ANY-method}
\title{Get/set methods for NpdeData object}
\usage{
\S4method{[}{NpdeData,ANY,ANY,ANY}(x, i, j, drop)
}
\description{
Access slots of a NpdeData using the object["slot"] format
}
\keyword{methods}
