% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeObject.R
\name{[,NpdeObject,ANY,ANY,ANY-method}
\alias{[,NpdeObject,ANY,ANY,ANY-method}
\title{Get/set methods for NpdeData object}
\usage{
\S4method{[}{NpdeObject,ANY,ANY,ANY}(x, i, j, drop)
}
\description{
Access slots of a NpdeData using the object["slot"] format
}
\keyword{methods}
