% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{summary.PM_cov}
\alias{summary.PM_cov}
\title{Wrapper function for summmary.PMcov}
\usage{
\method{summary}{PM_cov}(object, ...)
}
\arguments{
\item{object}{The \emph{PM_cov} object to summarize}

\item{...}{Arguments passed to \link{summary.PMcov}}
}
\value{
A \link{summary.PMcov} object
}
\description{
This redirects to summary.PMcov for PM_final R6 objects
}
\details{
See \link{summary.PMcov}. Alternative way to summarize is
\code{PM_result$cov$summary()}.
}
\author{
Michael Neely, Julian Otalvaro
}
