% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_cycle.R
\name{summary.PM_cycle}
\alias{summary.PM_cycle}
\title{Summarize Final Cycle Statistics}
\usage{
\method{summary}{PM_cycle}(object, cycle = NULL, digits = 3, ...)
}
\arguments{
\item{object}{A \link{PM_cycle} object}

\item{cycle}{Cycle number to summarize. Default is last cycle.}

\item{digits}{Number of digits to round to. Default is 3.}

\item{...}{Not used.}
}
\value{
A list of class \emph{summary.PM_cycle} whose elements are the last cycle values for the following fields
in a \link{PM_cycle} object.
\itemize{
\item \strong{cycle} Maximum cycle number
\item \strong{ll} Log likelihood
\item \strong{aic} Akaike Information Criterion
\item \strong{bic} Bayesian Information Criterion
\item \strong{gamlam} Value of gamma or lambda for each output equation
\item \strong{mean} Normalized mean parameter values compared to initial value
\item \strong{sd} Normalized standard deviation of parameter values compared to initial value
\item \strong{median} Normalized median parameter values compared to initial value
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize a Pmetrics Cycle object
}
\details{
This is a function usually called by the \verb{$summary()} method for \link{PM_cycle} objects
within a \link{PM_result} to summarize final cycle statistics. The function can
be called directly on a \link{PM_cycle} object. See examples.
}
\examples{
#'
\dontrun{
NPex$cycle$summary() # preferred
summary(NPex$cycle) # alternative
}
}
\seealso{
\link{PM_cycle}
}
\author{
Michael Neely
}
