% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_cycle.R
\name{summary.PM_cycle_data}
\alias{summary.PM_cycle_data}
\title{Summarize PM_cycle_data objects}
\usage{
\method{summary}{PM_cycle_data}(object, ...)
}
\arguments{
\item{object}{A \code{PM_cycle_data} object}

\item{...}{Additional arguments passed to \link{summary.PM_cycle}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Summarizes the raw data (\code{class: PM_cycle_data}) from a \link{PM_cycle} object in the same way as summarizing a \link{PM_cycle} object.
Both use \link{summary.PM_cycle}.
}
\examples{
# There is no example we can think of to filter or otherwise process a PM_cycle object,
# but we provide this function for completeness.
NPex$cycle$data \%>\% summary()
# all the below are the same
# summary(NPex$cycle$data) 
# summary(NPex$cycle)
# NPex$cycle$summary()
}
