% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_data.R
\name{summary.PM_data}
\alias{summary.PM_data}
\title{Summarize a Pmetrics PM_data object}
\usage{
\method{summary}{PM_data}(object, ...)
}
\arguments{
\item{object}{Data to be summarized}

\item{...}{Arguments to pass to \link{summary.PMmatrix}}
}
\description{
Summarize a PM_data object using S3 method.
Calls \code{\link{summary.PMmatrix}}
}
