% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{summary.PM_final}
\alias{summary.PM_final}
\title{Wrapper function for summmary.PMfinal}
\usage{
\method{summary}{PM_final}(object, ...)
}
\arguments{
\item{object}{The \emph{PM_final} object to summarize}

\item{...}{Arguments passed to \link{summary.PMfinal}}
}
\value{
A \link{summary.PMfinal} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This redirects to summary.PMfinal for PM_final R6 objects
}
\details{
See \link{summary.PMfinal}. Alternative way to summarize is
\code{PM_result$final$summary()}.
}
