% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_final.R
\name{summary.PM_final_data}
\alias{summary.PM_final_data}
\title{Summarize PM_final_data objects}
\usage{
\method{summary}{PM_final_data}(object, ...)
}
\arguments{
\item{object}{A \code{PM_final_data} object}

\item{...}{Additional arguments passed to \link{summary.PM_final}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Summarizes the raw data (\code{class: PM_final_data}) from a \link{PM_final} object in the same way as summarizing a \link{PM_final} object.
Both use \link{summary.PM_final}.
}
\examples{
NPex$final$data \%>\% summary()
}
