% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{summary.PM_op}
\alias{summary.PM_op}
\title{Wrapper function for summmary.PMop}
\usage{
\method{summary}{PM_op}(obj, ...)
}
\arguments{
\item{obj}{The \emph{PM_op} object to summarize}

\item{...}{Arguments passed to \link{summary.PMop}}
}
\value{
A \link{summary.PMop} object
}
\description{
This redirects to summary.PMop for PM_op R6 objects
}
\details{
See \link{summary.PMop}. Alternative way to summarize is
\code{PM_result$op$summary()}.
}
