% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_op.R
\name{summary.PM_op_data}
\alias{summary.PM_op_data}
\title{Summarize PM_op_data objects}
\usage{
\method{summary}{PM_op_data}(object, ...)
}
\arguments{
\item{object}{A \code{PM_op_data} object}

\item{...}{Additional arguments passed to \link{summary.PM_op}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Summarizes the raw data (\code{class: PM_op_data}) from a \link{PM_op} object in the same way as summarizing a \link{PM_op} object.
Both use \link{summary.PM_op}.
}
\examples{
\dontrun{
NPex$op$data \%>\%
dplyr::filter(pred > 5) \%>\%
dplyr::filter(pred < 10) \%>\%
summary()
}
}
