% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_post.R
\name{summary.PM_post}
\alias{summary.PM_post}
\title{Summarize Observations and Predictions}
\usage{
\method{summary}{PM_post}(
  object,
  digits = max(3, getOption("digits") - 3),
  icen = "median",
  outeq = 1,
  ...
)
}
\arguments{
\item{object}{A \link{PM_post} object}

\item{digits}{Integer, used for number of digits to print.}

\item{icen}{Can be either "median" for the predictions based on medians of the posterior parameter value
distributions, or "mean".  Default is "median".}

\item{outeq}{Output equation number.  Default is 1.}

\item{...}{Not used.}
}
\value{
A data frame with the minimum, first quartile, median, third quartile, maximum,
mean and standard deviation for times and predictions in \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize a Pmetrics Observed vs. Predicted object
}
\details{
This is a function usually called by the \verb{$summary()} method for \link{PM_op} objects
within a \link{PM_result} to summarize observations, predictions and errors. The function can
be called directly on a \link{PM_op} object. See examples.
}
\examples{
\dontrun{
NPex$post$summary() # preferred
summary(NPex$post) # alternative
}
}
\seealso{
\link{PM_post}
}
\author{
Michael Neely
}
