% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_sim.R
\name{summary.PM_sim}
\alias{summary.PM_sim}
\title{Plot Pmetrics Simulation Objects}
\usage{
\method{summary}{PM_sim}(
  object,
  at = 1,
  field = "obs",
  by,
  individual = FALSE,
  statistics = c("mean", "sd", "median", "min", "max")
)
}
\arguments{
\item{at}{Index of the PM_sim object to be summarized. Default is 1.}

\item{field}{Quoted character value, one of
\itemize{
\item \strong{obs} for simulated observations (the default)
\item \strong{amt} for simulated amounts in each compartment
\item \strong{parValues} for simulated parameter values
}}

\item{by}{Optional quoted column name(s) to group by, e.g. \code{by = "id"} or
\code{by = c("id", "outeq")}.}

\item{individual}{If "id" is included within \code{by}, the summary statistics
will be calculated for each individual, i.e. each individual's mean, sd, median,
min and max observations. If \code{individual} is \code{FALSE} (the default), then those
summaries will again be summarized to present, for example, the mean of
all subjects' mean observations, or the mean of their maximum observations.
This argument will be ignored if "id" is not in \code{by}.}

\item{statistics}{Default is  \code{c("mean", "sd", "median", "min", "max")}, but
can include specific quantiles, e.g. \code{c(25, "median", 75)}.}
}
\value{
If \code{by} is ommitted, a data frame with rows for each data element except ID,
and columns labeled as mean, sd, median, min and max. If \code{by} is specified,
return will be a list with named elements mean, sd, median, min and max, each containing
the corresponding value for each group in \code{by}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize simulation
}
\details{
Summarizes simulated observations, compartment amounts or parameter values.
Can be ungrouped (i.e. the entire simulation), average values grouped by simulated id,
outeq, or both, as well as summary statistics for each individual.
Default statistics reported
are mean, sd, median, min, max, but could include individual quantiles. See
\code{statistics} below.
}
