% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPMfinal.R
\name{summary.PMfinal}
\alias{summary.PMfinal}
\title{Summary Statistics for PMfinal Objects}
\usage{
\method{summary}{PMfinal}(object, ..., lower = 0.025, upper = 0.975)
}
\arguments{
\item{object}{The PMfinal object made after an NPAG or IT2B, e.g. final.1 after run 1.}

\item{\dots}{Other parameters which are not necessary.}

\item{lower}{Desired lower confidence interval boundary.  Default is 0.025. Ignored for IT2B objects.}

\item{upper}{Desired upper confidence interval boundary.  Default is 0.975. Ignored for IT2B objects.}
}
\value{
The output is a data frame.
For NPAG this has 4 columns:
\item{value }{The value of the summary statistic}
\item{par }{The name of the parameter}
\item{type }{Either \emph{WtMed} for weighted median, or \emph{MAWD} for MAWD (see details)}
\item{quantile }{Requested \code{lower}, 0.5 (median), and \code{upper} quantiles}
For IT2B this has 5 columns:
\item{mean }{Parameter mean value}
\item{se.mean }{Standard error of the mean}
\item{cv.mean }{Error of the mean divided by mean}
\item{var }{Variance of the parameter values}
\item{se.var }{Standard error of the variance}
}
\description{
Generates summary statistics of final population model parameters.
}
\details{
For NPAG runs, this function will generate weighted medians as central tendencies of the
population points with a 95\% confidence interval (95\% CI) around the median, 
and the median absolute weighted deviation (MAWD) from the median as a measure 
of the variance, with its 95\% CI.  These estimates correspond to weighted mean, 
95\% CI of the mean, variance, and 95\% CI of the variance, respectively, for a 
sample from a normal distribution.  To estimate these non-parametric summaries, 
the function uses a Monte Carlo simulation approach, creating  1000 x npoint samples 
with replacement from the weighted marginal distribution of each parameter, 
where npoint is the number of support points in the model.  As an example, 
if there are 100 support points, npoint = 100, and for Ka, there will be 
1000 sets of 100 samples drawn from the weighted marginal distribution of the 
values for Ka.  For each of the 1,000 sets of npoint values, the median and MAWD are 
calculated, with MAWD equal to the median absolute difference between each point 
and the median of that set.  The output is npoint estimates of the weighted median 
and npoint estimates of the MAWD for each parameter, from which the median, 2.5th, 
and 97.5th percentiles can be found as point estimates and 95\% confidence 
interval limits, respectively, of both the weighted median and MAWD.

For IT2B runs, the function will return the mean and variance of each parameter,
and the standard errors of these terms, using SE (mean) = SD/sqrt(nsub) and 
SE (var) = var * sqrt(2/(nsub-1)).
}
\examples{
data(NPdata.1)
final <- makeFinal(NPdata.1)
summary(final)
data(ITdata.2)
final2 <- makeFinal(ITdata.2)
summary(final2)
}
\seealso{
\code{\link{makeFinal}}, \code{\link{ITparse}},  \code{\link{plot.PMfinal}}
}
\author{
Michael Neely
}
