% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{template}
\alias{template}
\title{Import recycled text into documentation}
\usage{
template(name)
}
\arguments{
\item{name}{The name of the text file}
}
\description{
Import recycled text into documentation
}
\keyword{internal}
