% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{temporary_path}
\alias{temporary_path}
\title{Retrieves the temporary path used for building models.}
\usage{
temporary_path()
}
\value{
Temporary build path.
}
\description{
Retrieves the temporary path used for building models.
}
