% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{three_comp_bolus_cl}
\alias{three_comp_bolus_cl}
\title{Three-compartment bolus clearance model template}
\usage{
three_comp_bolus_cl()
}
\value{
A \code{PM_model} object.
}
\description{
Create a \code{PM_model} object for the \code{three_comp_bolus_cl} model template.
}
