% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{three_comp_iv}
\alias{three_comp_iv}
\title{Three-compartment IV model template}
\usage{
three_comp_iv()
}
\value{
A \code{PM_lib} object.
}
\description{
Create a \code{PM_lib} object for the \code{three_comp_iv} model template.
}
