% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{two_comp_bolus}
\alias{two_comp_bolus}
\title{Two-compartment bolus model template}
\usage{
two_comp_bolus()
}
\value{
A \code{PM_lib} object.
}
\description{
Create a \code{PM_lib} object for the \code{two_comp_bolus} model template.
}
