% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{two_comp_iv}
\alias{two_comp_iv}
\title{Two-compartment IV model template}
\usage{
two_comp_iv()
}
\value{
A \code{PM_model} object.
}
\description{
Create a \code{PM_model} object for the \code{two_comp_iv} model template.
}
