% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{two_comp_iv_cl}
\alias{two_comp_iv_cl}
\title{Two-compartment IV clearance model template}
\usage{
two_comp_iv_cl()
}
\value{
A \code{PM_model} object.
}
\description{
Create a \code{PM_model} object for the \code{two_comp_iv_cl} model template.
}
