% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGfortran.R
\name{update_gfortran}
\alias{update_gfortran}
\title{Download and install Gfortran updates}
\usage{
update_gfortran()
}
\value{
A boolean that represents if the latest version of Gfortran for your OS is installed.
}
\description{
Download and install Gfortran updates
IMPORTANT: Only use this function if you are using either brew on Mac or chocolatey on windows to install gfortran.
}
\author{
Michael Neely
}
