% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{validate_model_source}
\alias{validate_model_source}
\title{Validates DSL model source against the runtime compiler.}
\usage{
validate_model_source(model_source, kind, solver = NULL)
}
\arguments{
\item{model_source}{DSL source code for the model.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}

\item{solver}{Optional ODE solver name.}
}
\value{
Result of the validation process.
}
\description{
Validates DSL model source against the runtime compiler.
}
\keyword{internal}
