% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgrowth.R
\name{zBMI}
\alias{zBMI}
\title{Extract CDC pediatric BMI z-scores}
\usage{
zBMI(agemos, sex, bmi)
}
\arguments{
\item{agemos}{The age in months.  Should be between 24 and 240.5.}

\item{sex}{A single quoted character: \dQuote{M} for males, \dQuote{F} for females.  Default is \dQuote{M}.}

\item{bmi}{The individual's BMI.}
}
\value{
A list with objects calculated for \code{agemos} and \code{sex}.
\item{z }{Z-score}
\item{mod_z }{Modified Z-score for extreme BMI}
\item{per }{BMI percentile}
\item{mod_per }{Modifed BMI percentile}
}
\description{
Will extract BMI z-scores based on age in months and sex. Overweight is
a z-score of >1.04, and obese is a z-score > 1.64. Calculations are based on
CDC data/formulae as found here:
\link{https://www.cdc.gov/nccdphp/dnpa/growthcharts/resources/biv-cutoffs.pdf}
}
\author{
Michael Neely
}
