% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_Utils.R
\name{add_shapes}
\alias{add_shapes}
\title{Add shapes to plotly plot}
\usage{
add_shapes(p = plotly::last_plot(), shape)
}
\arguments{
\item{p}{The plot to which the shape should be added. Default is the
\code{last_plot()}.}

\item{shapes}{A list of attributes that specify a shape.}
}
\description{
Modifies the layout object of an existing plot to include a new shape.
}
\details{
This function adds a shape to the layout element of a plotly plot.
Type \code{schema()} in the console and expand the list for
layout > layoutAttributes > shapes for details on how
to specify a shape. A convenient Pmetrics helper function to add line shapes
is \link{ab_line}. Other shapes such as circles, rectangles and paths can be added,
but must be done manually as outlined in the plotly schema documentation.
An example of a circle shape is below in examples.
}
\examples{
NPex$op$plot()
add_shapes(shapes = ab_line(v = 12))

NPex$data$plot()
add_shapes(shapes = list(type = "circle", x0 = 125, y0 = 10, x1 = 135, y1 = 15))
}
\seealso{
\link{ab_line}
}
