% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{additive}
\alias{additive}
\title{Additive error model}
\usage{
additive(add, constant = FALSE)
}
\arguments{
\item{add}{Initial value for lambda}

\item{constant}{Estimate if \code{FALSE} (default).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Create an additive (lambda) error model
}
