% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{additive}
\alias{additive}
\title{Additive error model}
\usage{
additive(initial, coeff, fixed = FALSE)
}
\arguments{
\item{initial}{Initial value for lambda}

\item{coeff}{Vector of coefficients defining assay error polynomial}

\item{fixed}{Estimate if \code{FALSE} (default).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Create an additive (lambda) error model
}
