% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Launch Model Builder app}
\usage{
build_model(data, ...)
}
\arguments{
\item{data}{Optional \link{PM_data} object which will be used to read
covariates.}

\item{...}{Currently unused}
}
\value{
Launches the shiny app.
}
\description{
Opens the shiny model builder app.
}
\examples{
\dontshow{if () (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
${1:interactive()}
${1:build_model(NPex$data)} 
\dontshow{\}) # examplesIf}
}
\author{
Michael Neely
}
