% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Launch Model Builder app}
\usage{
build_model(..., update = FALSE)
}
\arguments{
\item{...}{Optional \link{PM_data} and/or \link{PM_model} object(s). \emph{PM_data} objects
supply covariates. \emph{PM_model} objects supply any other defined model element,
and covariates only if there is no \emph{PM_data} object or it has no covariates.
If the \emph{PM_model} object contains covariates, they will be superseded by those in
the \emph{PM_data} object, if supplied.}

\item{update}{Logical. If \code{TRUE}, the \code{Save} button in the app will update the model,
rather than write the model to a file.}
}
\value{
Launches the shiny app.
}
\description{
Open the shiny model builder app.
}
\details{
The app will open in a separate window.
}
\author{
Michael Neely
}
