% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{clear_build}
\alias{clear_build}
\title{Clear build files}
\usage{
clear_build()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Deletes rust template files from the Pmetrics template directory.
}
\keyword{internal}
