% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compile_model}
\alias{compile_model}
\title{Compiles the text representation of a model into a binary file.}
\usage{
compile_model(model_path, output_path, params)
}
\description{
Compiles the text representation of a model into a binary file.
}
