% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compile_model}
\alias{compile_model}
\title{Compiles the text representation of a model into a binary file.}
\usage{
compile_model(model_path, output_path, params, template_path, kind)
}
\arguments{
\item{model_path}{Path to the model file.}

\item{output_path}{Path to save the compiled model.}

\item{params}{List of model parameters.}

\item{template_path}{Path to the template directory.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}
}
\value{
Result of the compilation process.
}
\description{
Compiles the text representation of a model into a binary file.
}
