% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMoptions.R
\name{editPMoptions}
\alias{editPMoptions}
\title{Edit Pmetrics User Options}
\usage{
editPMoptions()
}
\value{
A list with the current options.
}
\description{
Edit user options for Pmetrics
}
\details{
This function will open the user options file for Pmetrics in the system
default editor for JSON. Use \strong{caution} directly editing the file to avoid errors.
This is the easiest way to remove specific Pmetrics function argument overrides.
See further details in the \code{func_defaults} argument to \link{setPMoptions}.
}
\author{
Michael Neely
}
