% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fit}
\alias{fit}
\title{Fits the given model to the data using the provided settings.}
\usage{
fit(model_source, data, params, output_path, solver = NULL)
}
\arguments{
\item{model_source}{Model definition written in the pharmsol DSL.}

\item{data}{Path to the data file.}

\item{params}{List of fitting parameters.}

\item{output_path}{Path to save the fitting results.}

\item{solver}{Optional ODE solver name.}
}
\value{
Result of the fitting process.
}
\description{
Fits the given model to the data using the provided settings.
}
\keyword{internal}
