% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fit}
\alias{fit}
\title{Fits the model DSL source to the data using the provided parameters.}
\usage{
fit(model_source, data, params, output_path, kind, solver = NULL)
}
\arguments{
\item{model_source}{DSL source code for the model.}

\item{data}{Path to the data file.}

\item{params}{List of fitting parameters.}

\item{output_path}{Path to save the fitting results.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}

\item{solver}{Optional ODE solver name.}
}
\value{
Result of the fitting process.
}
\description{
Fits the model DSL source to the data using the provided parameters.
}
\keyword{internal}
