% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fit}
\alias{fit}
\title{Fits the model at the given path to the data at the given path using the provided parameters.}
\usage{
fit(model_path, data, params, output_path, kind)
}
\arguments{
\item{model_path}{Path to the compiled model file.}

\item{data}{Path to the data file.}

\item{params}{List of fitting parameters.}

\item{output_path}{Path to save the fitting results.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}
}
\value{
Result of the fitting process.
}
\description{
Fits the model at the given path to the data at the given path using the provided parameters.
}
