% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMutilities.R
\name{getCov}
\alias{getCov}
\title{Extract covariate information}
\usage{
getCov(mdata)
}
\arguments{
\item{mdata}{A \link{PM_data} object. It can be other data frames but the results
will likely not be meaningful.}
}
\value{
A list with named items: \emph{ncov, covnames, covstart, covend}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts covariate information from a Pmetrics data object.
}
\details{
The function subtracts the number of fixed columns in a standard data format,
currently 14, from the total number of columns in \code{mdata} and queries
the remaining columns.  When given a \link{PM_data} object, will return a list
with the number of covariates, their names, and the starting and
ending column numbers
}
\examples{
\dontrun{
getCov(dataEx)
}
}
\author{
Michael Neely
}
