% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_Utils.R
\name{getDefaultColors}
\alias{getDefaultColors}
\title{Get a list of default colors}
\usage{
getDefaultColors(n)
}
\arguments{
\item{n}{The number of colors to return from the list.}
}
\value{
A character vector of color names, which is recycled as needed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Generate list of default color names.
}
\details{
Used for Pmetrics plots. The following list is recycled as necessary to generate the
requested number of colors:
\code{c(red(), green(), blue(), brown(), black(), purple(), pink(), gold(), orange(), gray())}.
Each value is a function that returns a color name.
}
\examples{
\dontrun{
getDefaultColors(6)
}
}
\author{
Michael Neely
}
