% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMoptions.R
\name{getPMoptions}
\alias{getPMoptions}
\title{Get Pmetrics User Options}
\usage{
getPMoptions(opt, warn = T)
}
\arguments{
\item{opt}{The option to retrieve.  If omitted, all option values will be returned.}

\item{warn}{Warn if options file doesn't exist. Default \code{TRUE}.}
}
\value{
The user options file will be updated.  This will persist from session to session.
}
\description{
Get user options for Pmetrics
}
\details{
This function will get user options for Pmetrics.  Current user options are
\itemize{
\item sep Field separator in data files
\item dec Decimal separator in numbers
}
}
\author{
Michael Neely
}
