% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMoptions.R
\name{getPMoptions}
\alias{getPMoptions}
\title{Get Pmetrics User Options}
\usage{
getPMoptions(opt, warn = T, quiet = F)
}
\arguments{
\item{opt}{The option to retrieve.  If omitted, all option values will be returned.}

\item{warn}{Warn if options file doesn't exist. Default \code{TRUE}.}

\item{quiet}{Suppress warning messages. Default \code{FALSE}.}
}
\value{
A list with the current options.
}
\description{
Get user options for Pmetrics
}
\details{
This function will get user options for Pmetrics.
}
\author{
Michael Neely
}
