% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMoptions.R
\name{getPMoptions}
\alias{getPMoptions}
\title{Get Pmetrics User Options}
\usage{
getPMoptions(opt, warn = TRUE, quiet = FALSE)
}
\arguments{
\item{opt}{The option to retrieve.  If omitted, all option values will be returned.}

\item{warn}{Warn if options file doesn't exist. Default \code{TRUE}.}

\item{quiet}{Suppress warning messages. Default \code{FALSE}.}
}
\value{
A list with the current options.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get user options for Pmetrics
}
\details{
This function will get user options for Pmetrics. It will look for a \emph{PMoptions.json} file
in a hidden folder outside of the Pmetrics package. If that does not exist,
it will look for a default options file in the package options folder. See \link{setPMoptions} for
details on where the options file is stored and how to set options.
}
\author{
Michael Neely
}
