% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAUC.R
\name{makeAUC}
\alias{makeAUC}
\title{Calculation of AUCs}
\usage{
makeAUC(
  data,
  formula,
  include,
  exclude,
  start = 0,
  end = Inf,
  icen = "median",
  outeq = 1,
  block = 1,
  method = "linear",
  addZero = F
)
}
\arguments{
\item{data}{A suitable data object, i.e. \link{PM_pop}, \link{PM_post}, \link{PM_op},
\link{PM_sim}, or some other suitable dataframe
with at least time/observation columns referred to by \code{formula}, with an
"id" column (so named) if necessary.}

\item{formula}{A formula of the form \code{obs ~ time}.
This is only required with data that is not of class PM_pop, PM_post, PM_op or PM_sim.}

\item{include}{A vector of subject IDs to include in the plot, e.g. c(1:3,5,15)}

\item{exclude}{A vector of subject IDs to exclude in the plot, e.g. c(4,6:14,16:20)}

\item{start}{Specify the time to begin AUC calculations. Default is 0.}

\item{end}{Specify the time to end AUC calculations so that AUC is calculated
from \code{start} to \code{end}.  Default for end is the maximum observation
time for each subject.  Subjects with insufficient data for a specified interval will have
AUC calculated for the available data, not to exceed the specified interval.}

\item{icen}{Can be either "median" for the predictions based on medians of \code{pred.type} parameter value distributions, or "mean".  Default is "median". Only relevant for PMpost or PMpop objects.}

\item{outeq}{Which output equation to plot. Default is 1.}

\item{block}{Which block to plot, where blocks are defined by dose reset events (EVID = 4) in the data. Default is 1.}

\item{method}{Default is "linear" for AUC trapezoidal calculation.  Any other value will result in
linear up, log down.}

\item{addZero}{Boolean to add a zero concentration at time 0. Default is \code{FALSE}.}
}
\value{
A dataframe of class \emph{PMauc}, which has 2 columns:
\itemize{
\item id - Subject identification
\item tau - AUC from \code{start} to \code{end}
}
}
\description{
Calculates AUC from a variety of inputs
}
\details{
Calculates the area under the time concentration curve using the
trapezoidal approximation from a variety of inputs.
If a PM_pop, PM_post, PM_op, or PMsim object is specified,
\code{formula} is not required.
}
\examples{
NPex$op$auc()
Indometh \%>\% mutate(id = Subject) \%>\% makeAUC(conc ~ time)
}
\seealso{
\link{PM_result}, \link{PM_sim}, \link{PM_op}
}
\author{
Michael Neely
}
