% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAUC.R
\name{makeAUC}
\alias{makeAUC}
\title{Calculation of AUCs}
\usage{
makeAUC(
  data,
  formula,
  include,
  exclude,
  start = 0,
  end = Inf,
  icen = "median",
  outeq = 1,
  block = 1,
  method = "linear"
)
}
\arguments{
\item{data}{A suitable data object of the \emph{PMpop} class (see \code{\link{makePop}}),
\emph{PMpost} class (see \code{\link{makePost}}),
\emph{PMop} class (see \code{\link{makeOP}}),
the \emph{PMsim} class (see \code{\link{SIMparse}}), or some other suitable dataframe
with at least time/observation columns referred to by \code{formula}, with an \dQuote{id} column (so named) if necessary.}

\item{formula}{A formula of the form \code{obs ~ time}.  This is only required with data that is not of class PMpop, PMpost, PMop or PMsim.}

\item{include}{A vector of subject IDs to include in the AUC calculations, e.g. c(1:3,5,15)}

\item{exclude}{A vector of subject IDs to exclude in the AUC calculations, e.g. c(4,6:14,16:20)}

\item{start}{Specify the time to begin AUC calculations. Default is 0.}

\item{end}{Specify the time to end AUC calculations so that AUC is calculated
from \code{start} to \code{end}.  Default for end is the maximum observation
time for each subject.  Subjects with insufficient data for a specified interval will have
AUC calculated for the available data, not to exceed the specified interval.}

\item{icen}{Only relevant for PMpost or PMpop objects which have predictions based on median or mean of each
subject's Bayesian posterior parameter distribution.  Default is "median", but could be "mean".}

\item{outeq}{Specify which output equation is to be used.  Default is 1.}

\item{block}{Specify which observation block (separated by EVID=4) is to be used for each subject.  Default is 1.}

\item{method}{Default is "linear" for AUC trapezoidal calculation.  Any other value will result in
linear up, log down.}
}
\value{
The output of \code{makeAUC} is a dataframe of class \emph{PMauc},
which has 2 columns:
\item{id }{subject identification}
\item{tau }{AUC from \code{start} to \code{end}}
}
\description{
Calculates AUC from a variety of inputs
}
\details{
\code{makeAUC} will calculate the area under the time concentration curve using the
trapezoidal approximation from a variety of inputs.  If a PMpost, PMop, or PMsim object is specified, 
\code{formula} is not required.  AUCs from PMop objects are based on observations.
For AUCs based on predictions, use a PMpost object.
}
\examples{
data(NPdata.1)
op <- makeOP(NPdata.1)
makeAUC(op)
}
\seealso{
\code{\link{makeOP}}, \code{\link{SIMparse}}
}
\author{
Michael Neely
}
