% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeCov}
\alias{makeCov}
\title{Extract covariate data}
\usage{
makeCov(data)
}
\arguments{
\item{data}{A suitable data object of the \emph{NPAG} or \emph{IT2B} class (see \code{\link{NPparse}} or \code{\link{ITparse}}).}
}
\value{
The output of \code{makeCov} is a dataframe of class \emph{PMcov},
which has \emph{nsub} rows and the following columns:
\item{id }{Subject identification}
\item{time }{Times of covariate observations}
\item{covnames... }{Columns with each covariate observations in the dataset for each subject and \code{time} }
\item{parnames... }{Columns with each parameter in the model and the \code{icen} summary
for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian
parameter means and medians }
A plot method exists in \code{\link{plot}} for \emph{makeCov} objects.
}
\description{
Generates an data.frame with subject-specific covariate data from an \emph{NPAG} or \emph{IT2B} object
}
\details{
For each subject, \code{makeCov} extracts covariate information and Bayesian posterior parameter estimates.
This output of this function is suitable for exploration of covariate-parameter, covariate-time, or parameter-time relationships.
}
\examples{
data(PMex1)
cov <- makeCov(PMex1)
cov
names(cov)
}
\author{
Michael Neely
}
\seealso{
\code{\link{NPparse}}, \code{\link{ITparse}}, \code{\link{plot.PMcov}}
}

