% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFinal.R
\name{makeFinal}
\alias{makeFinal}
\title{Summarize NPAG or IT2B Final Cycle Population Values}
\usage{
makeFinal(data)
}
\arguments{
\item{data}{A suitable data object of the \emph{NPAG} or \emph{IT2B} class (see \code{\link{NPparse}} or \code{\link{ITparse}}).}
}
\value{
The output of \code{makeFinal} is a list of class \emph{PMfinal}, which contains the following:
\item{popPoints }{(NPAG only) Dataframe of the final cycle joint population density of grid points
 with column names equal to the name of each random parameter plus \emph{prob} for the
 associated probability of that point}
\item{popMean }{The final cycle mean for each random parameter distribution}
\item{popSD }{The final cycle standard deviation for each random parameter distribution}
\item{popCV }{The final cycle coefficient of variation (SD/Mean) for each random parameter distribution}
\item{popVar }{The final cycle variance for each random parameter distribution}
\item{popCov }{The final cycle random parameter covariance matrix}
\item{popCor }{The final cycle random parameter correlation matrix}
\item{popMedian }{The final cycle median values for each random parameter}
#' \item{postPoints}{(NPAG only) Dataframe of posterior population points for each of the first 100 subject,
with columns id, point, parameters and probability.  The first column is the subject, the second column has the population
point number, followed by the values for the parameters in that point and the probability.}
\item{postMean }{A \emph{nsub} x \emph{npar} data frame containing 
the means of the posterior distributions for each parameter.}
\item{postSD }{A \emph{nsub} x \emph{npar} data frame containing 
the SDs of the posterior distributions for each parameter.}
\item{postVar }{A \emph{nsub} x \emph{npar} data frame containing 
the variances of the posterior distributions for each parameter.}
\item{postCov }{NPAG only: An array of dimensions \emph{npar} x \emph{npar} x \emph{nsub} that
contains the covariances of the posterior distributions for each parameter and subject.}
\item{postCor }{NPAG only: An array of dimensions \emph{npar} x \emph{npar} x \emph{nsub} that
contains the correlations of the posterior distributions for each parameter and subject.}
\item{postMed }{A \emph{nsub} x \emph{npar} data frame containing 
the medians of the posterior distributions for each parameter.}
\item{shrinkage }{A data frame with the shrinkage for each parameter.  \code{popVar}
is comprised of variance(EBE) + variance(EBD), where EBE is the Emprical Bayes Estimate or mean of the posterior
distribution for the parameter. EBD is the Empirical Bayes Distribution, or
the full Bayesian posterior distribution. In other words, if Bayesian posterior distributions are wide
for a given parameter due to sparse or uninformative sampling, then most of the population variance is due
to this variance and shrinkage of the EBE variance is high because individual posterior estimates
shrink towards the population mean.}
\item{gridpts }{(NPAG only) Initial number of support points}
\item{nsub }{Number of subjects}
\item{ab }{Matrix of boundaries for random parameter values}
A plot method exists in \code{\link{plot}} for \emph{PMfinal} objects.
}
\description{
Extracts final cycle information from NPAG or IT2B run.
}
\details{
This function will parse the output of \code{\link{NPparse}} or \code{\link{ITparse}} to generate a
list suitable for analysis and plotting of NPAG  or IT2B final cycle population values.
}
\examples{
data(PMex1)
final <- makeFinal(NPdata.1)
final
names(final)
}
\seealso{
\code{\link{NPparse}}, \code{\link{ITparse}},  \code{\link{plot.PMfinal}}
}
\author{
Michael Neely
}
