% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeFinal}
\alias{makeFinal}
\title{Summarize NPAG or IT2B Final Cycle Population Values}
\usage{
makeFinal(data)
}
\arguments{
\item{data}{A suitable data object of the \emph{NPAG} or \emph{IT2B} class (see \code{\link{NPparse}} or \code{\link{ITparse}}).}
}
\value{
The output of \code{makeFinal} is a list of class \emph{PMfinal}, which has 11 objects from NPAG,
or 9 objects from IT2B:
\item{popPoints }{(NPAG only) Dataframe of the final cycle joint population density of grid points
 with column names equal to the name of each random parameter plus \emph{prob} for the
 associated probability of that point}
\item{popMean }{The final cycle mean for each random parameter distribution}
\item{popSD }{The final cycle standard deviation for each random parameter distribution}
\item{popCV }{The final cycle coefficient of variation (SD/Mean) for each random parameter distribution}
\item{popVar }{The final cycle variance for each random parameter distribution}
\item{popCov }{The final cycle random parameter covariance matrix}
\item{popCor }{The final cycle random parameter correlation matrix}
\item{popMedian }{The final cycle median values for each random parameter}
\item{gridpts }{(NPAG only) Initial number of support points}
\item{nsub }{Number of subjects}
\item{ab }{Matrix of boundaries for random parameter values}
A plot method exists in \code{\link{plot}} for \emph{PMfinal} objects.
}
\description{
Extracts final cycle information from NPAG or IT2B run.
}
\details{
This function will parse the output of \code{\link{NPparse}} or \code{\link{ITparse}} to generate a
list suitable for analysis and plotting of NPAG  or IT2B final cycle population values.
}
\examples{
data(PMex1)
final <- makeFinal(NPdata.1)
final
names(final)
}
\author{
Michael Neely
}
\seealso{
\code{\link{NPparse}}, \code{\link{ITparse}},  \code{\link{plot.PMfinal}}
}

