% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeNCA}
\alias{makeNCA}
\title{Non-compartmental analysis}
\usage{
makeNCA(post, data, input = 1, icen = "median", outeq = 1, block = 1,
  start = 0, end = Inf)
}
\arguments{
\item{post}{An \emph{PMpost} object created by \code{\link{makePost}} or loaded with \code{\link{NPload}}}

\item{data}{An \emph{NPAG} object created by \code{\link{NPparse}} or loaded with \code{\link{NPload}}}

\item{input}{The number of the input (e.g. drug) to analyze; default 1.}

\item{icen}{Only relevant for PMpost or PMpop objects which have predictions based on median or mean of each
subject's Bayesian posterior parameter distribution.  Default is "median", but could be "mean".}

\item{outeq}{The number of the output equation to analyze; default 1}

\item{block}{The number of the observation block within subjects, with each block delimited by EVID=4 in the data file; default 1}

\item{start}{If the \code{start} time is not 0 (default), then it is assumed that steady state (multiple dose) conditions apply.}

\item{end}{Set this equal to the dosing interval for steady state (multiple dose) analysis.}
}
\value{
A dataframe of class \emph{PMnca} with columns
 \item{id }{Subject identification}
 \item{auc }{Area under the time-observation curve, using the trapezoidal approximation, from time 0 until the second dose,
or if only one dose, until the last observation}
 \item{aumc }{Area under the first moment curve}
 \item{k }{Slope by least-squares linear regression of the final 6 log-transformed observations vs. time}
 \item{auclast }{Area under the curve from the time of the last observation to infinity, calculated as [Final obs]/k}
 \item{aumclast }{Area under the first moment curve from the time of the last observation to infinity}
 \item{aucinf }{Area under the curve from time 0 to infinity, caluculated as auc + auclast}
 \item{aumcinf }{Area under the first moment curve from time 0 to infinity}
 \item{mrt }{Mean residence time, calculated as 1/k}
 \item{cmax }{Maximum predicted concentration after the first dose}
 \item{tmax }{Time to cmax}
 \item{cl }{Clearance, calculated as dose/aucinf}
 \item{vdss }{Volume of distribution at steady state, calculated as cl*mrt}
 \item{thalf }{Half life of elimination, calculated as ln(2)/k}
 \item{dose }{First dose amount for each subject}
}
\description{
Performs a non-compartmental analysis on individual Bayesian posterior predicted
time-observation profiles generated after an NPAG run by the \code{\link{makePost}}
command.
}
\author{
Michael Neely
}

