% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{makeNPDE}
\alias{makeNPDE}
\title{makeNPDE}
\usage{
makeNPDE(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use $validate() on a PM_result object or make_valid() instead.
}
\author{
Michael Neely
}
\keyword{internal}
