% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makePTA}
\alias{makePTA}
\title{Calculation of PTAs}
\usage{
makePTA(simdata, targets, target.type, success, outeq = 1,
  free.fraction = 1, start, end)
}
\arguments{
\item{simdata}{A vector of simulator output filenames, e.g. c(\dQuote{simout1.txt},\dQuote{simout2.txt}),
with wildcard support, e.g. \dQuote{simout*} or \dQuote{simout?}, or
a list of PMsim objects made by \code{\link{SIMparse}} with suitable simulated doses and observations.  The number and times of simulated
observations does not have to be the same in all objects.}

\item{targets}{A vector of pharmacodynamic targets, such as Minimum Inhibitory Concentrations (MICs), e.g. c(0.25, 0.5,1,2,4,8,16,32)}

\item{target.type}{A numeric or character vector, length 1.  If numeric, must correspond to an observation time common to all PMsim objects in
\code{simdata}, rounded to the nearest hour.  In this case, the target statistic will be the ratio of observation at time \code{target.type} to target.  This enables
testing of a specific timed concentration (e.g. one hour after a dose or C1) which may be called a peak, but is not actually the maximum drug
concentration.  Be sure that the time in the simulated data is used, e.g. 122 after a dose given at 120.  Character values may be one of
\dQuote{time}, \dQuote{auc}, \dQuote{peak}, or \dQuote{min}, for, respectively, percent time above target within the time range
specified by \code{start} and \code{end}, ratio of area under the curve within the time range to target, ratio of peak concentration within the time range
to target, or ratio of minimum concentration within the time range to target.}

\item{success}{A single value specifying the success statistic, e.g. 0.4 for proportion time (end-start) above target, or 100 for peak:target.}

\item{outeq}{An integer specifying the number of the simulated output equation to use. Default is 1.}

\item{free.fraction}{Proportion of free, active drug.  Default is 1, i.e. 100\% free drug or 0\% protein binding.}

\item{start}{Specify the time to begin PTA calculations. Default is a vector with the first observation time for subjects
in each element of \code{simdata}, e.g. dose regimen. If specified as a vector, values will be recycled as necessary.}

\item{end}{Specify the time to end PTA calculations so that PTA is calculated
from \code{start} to \code{end}.  Default for end is the maximum observation
time for subjects in each element of \code{simdata}, e.g. dose regimen.  If specified as a vector, values will be recycled
as necessary. Subjects with insufficient data (fewer than 5 simulated observations) for a specified interval will trigger a warning.
Ideally then, the simulated datset should contain sufficient observations within the interval specified by \code{start} and \code{end}.}
}
\value{
The output of \code{makePTA} is a list of class \emph{PMpta},
which has 2 objects:
\item{results }{A data frame with the following columns: simnum,id,target,ratio.
\emph{simnum} is the number of the simulation; \emph{id} is the simulated profile number
within each simulation; \emph{target} is the specified target; and \emph{ratio} is
the target ratio, e.g. time > target, auc:target, etc.}
\item{outcome }{A data frame summarizing the results with the following columns: simnum, target, success, meanratio, and sdratio.
\emph{simnum} and \emph{target} are as for \code{outcome}.  The \emph{prop.success} column has the proportion with a ratio > \code{success},
as specified in the function call.  The \emph{mean.stat} and \emph{sd.stat} columns have the
mean and standard deviation of the target statistic (e.g. proportion end-start above target, ratio of Cmax to target) for each simulation and target.}
}
\description{
Calculates the Percent Target Attainment (PTA)
}
\details{
\code{makePTA} will calculate the PTA for any number of simulations, targets and definitions of success.
Simulations typically differ by dose, but may differ by other features such as children vs. adults.
}
\author{
Michael Neely
}
\seealso{
\code{\link{plot.PMpta}}, \code{\link{SIMparse}}
}

