% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_pta.R
\name{makePTAtarget}
\alias{makePTAtarget}
\title{Make a Percent Target Attainment (PTA) Target}
\usage{
makePTAtarget(x)
}
\arguments{
\item{x}{A data.frame or name of .csv file in working directory whose first two
columns are targets and the number of samples for each target.  An example can be
seen for Staphylococcus aureus susceptibility to vancomycin at
\href{http://mic.eucast.org/Eucast2/regShow.jsp?Id=1214}{EUCAST}.}
}
\value{
A data frame with two columns named targets and n, of class \emph{PMpta.targ}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Generates an object of class \emph{PMpta.targ} which can
be used in the \verb{$new()} method for \link{PM_pta} or \verb{$pta()} method for \link{PM_sim} for targets sampled from a distribution.
}
\examples{
\dontrun{
makePTAtarget(mic1)
}
}
\seealso{
\link{PM_pta}
}
