% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValid.R
\name{makeValid}
\alias{makeValid}
\title{Simulation-based model diagnostics}
\usage{
makeValid(run, outeq = 1, input = 1, icen = "median", binCov, clusters,
  lower = 0.025, upper = 0.975, ...)
}
\arguments{
\item{run}{When the current working directory is the Runs folder, the folder name of a previous run that you wish to use for the npde,
which will typically be a number, e.g. 1.}

\item{outeq}{The number of the output equation to simulate/test.  Default is missing, which will test all output equations.}

\item{icen}{Central tendency of parameter value distributions to be used for predictions.  Default is median.}

\item{binCov}{Character vector of covariates in model equations which should be binned.}

\item{clusters}{Numeric vector of length 2 with number of clusters to use for covariates in \code{binCov} and for time. If
not specified, Pmetrics will guide user through process to choose the number of clusters.}

\item{lower}{Lower percentile for pcVPC}

\item{upper}{Upper percentile for pcVPC}

\item{\dots}{Other parameters to be passed to \code{\link{SIMrun}}, for example \code{limits}.}
}
\value{
The output of \code{makeValid} is a list of class \code{PMvalid} with the following objects
\item{upper}{Data Frame with plot locations for upper confidence region}
\item{med}{Data Frame with plot locations for median confidence region}
\item{lower}{Data Frame with plot locations for lower confidence region}
\item{quant_Obs}{Data Frame with quantiles for each observation}
\item{quant_pcObs}{Data Frame with quantiles for each prediction-corrected observation}
\item{med2}{Data Frame with central time for each time bin}
\item{tempDF}{Data Frame with subject ID, times, observations, predictions, and bin-corrected predictions}
\item{npde}{A list of class \code{PMnpde}, itself a list of of \code{NpdeObject} class.  Additionally, two objects with run numbers appended will be saved to the output directory of the run for subsequent
loading with \code{\link{PMload}}: npde and sim.  \emph{npde} is the PMnpde object, and \emph{sim} is a 
PMsim object of all simulations combined which can be used for visual predictive checks
(see \code{\link{plot.PMsim}}).}
}
\description{
Use simulations to create an object which can be used to internally validate a model.
}
\details{
This function generates the data necessary to plot a Visual Predictive Check (VPC), a prediction-corrected VPC (pcVPC) using methods by
Bergstrand et al, and/or a Normalized Prediction Distribution Error (NPDE) using the \code{npde}
package of Comets et al.
Output from a loaded NPAG or IT2B run will be used as the population model supplied to the
simulator. The function will iterate through the .csv file, using each subject as a template to simulate \code{nsim} new individuals. 
The \code{icen} population values will be used for each parameter and the covariance matrix.  
Errors may arise if extreme or negative concentrations are simulated
from excessively large covariance matrices, so limits on the simulations may be necessary.
}
\references{
Wang DD, Zhang S. (2012) Standardized visual predictive check versus visual predictive check for model evaluation. \emph{J Clin Pharmacol}, 52(1):39–54. 

Comets E, Brendel K, Mentre F. (2011) Why Should Prediction Discrepancies Be Renamed Standardized Visual Predictive Check? \emph{J Clin Pharmacol}, epub Sep 10.

Bergstrand M, Hooker AC, Wallin JE, Karlsson MO. (2011) Prediction-corrected visual predictive checks for diagnosing nonlinear mixed-effects models. \emph{The AAPS Journal}, 13(2):143–51.

Brendel K, Comets E, Laffont CM, Laveille C, Mentre F (2006) Metrics for external model evaluation with an application to the population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research}, 23:2036-49

Mentre F, Escolano S (2006) Prediction discrepancies for the evaluation of nonlinear mixed-effects models. \emph{J Pharmacokinet Pharmacodyn}, 33:345-67
}
\seealso{
\code{\link{SIMrun}}, \code{\link{autonpde}}, \code{\link{plot.PMvalid}}
}
\author{
Michael Neely
}
