% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{model_lib}
\alias{model_lib}
\title{Pharmacokinetic model library}
\usage{
model_lib(name = NULL, show = TRUE)
}
\arguments{
\item{name}{The name of the model to display. If \code{NULL}, the entire list is displayed.}

\item{show}{If \code{TRUE}, the model is displayed in the console. If \code{FALSE}, the model is only returned as a tibble.}
}
\value{
A tibble with the model equations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function provides a list of available pharmacokinetic models.
}
\examples{
\dontrun{
model_lib()
model_lib("one_comp_iv")
}
}
\seealso{
\link{PM_model}
}
\author{
Michael Neely
}
