% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{model_lib}
\alias{model_lib}
\title{Model Library}
\usage{
model_lib(show = TRUE)
}
\arguments{
\item{show}{Logical indicating if the table should be printed. Default is TRUE.}
}
\value{
Invisibly, a tibble containing the model templates or a specific model template.
}
\description{
Returns a table of all available model templates in the Pmetrics library.
}
