% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{model_parameters}
\alias{model_parameters}
\title{Retrieves the model parameters from compiled DSL metadata.}
\usage{
model_parameters(model_source, kind)
}
\arguments{
\item{model_source}{DSL source code for the model.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}
}
\value{
List of model parameters.
}
\description{
Retrieves the model parameters from compiled DSL metadata.
}
\keyword{internal}
