% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{model_parameters}
\alias{model_parameters}
\title{Retrieves the model parameters from the compiled model at the given path.}
\usage{
model_parameters(model_path, kind)
}
\arguments{
\item{model_path}{Path to the compiled model file.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}
}
\value{
List of model parameters.
}
\description{
Retrieves the model parameters from the compiled model at the given path.
}
