% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMMopt.R
\name{plot.MMopt}
\alias{plot.MMopt}
\title{Plot Pmetrics Multiple-Model Optimal Sampling Objects}
\usage{
\method{plot}{MMopt}(x, probs = NA, line = T, ...)
}
\arguments{
\item{x}{The name of an \emph{MMopt} data object generated by \code{\link{MMopt}}}

\item{probs}{Default is NA. See \link{plot.PMsim} for details.}

\item{line}{Controls characteristics of lines.
This argument maps to plotly trace line attributes.
\code{TRUE} will plot default lines. \code{FALSE} will suppress lines.
If a list, can control many line characteristics, including overriding defaults.
Use the plotly \code{plotly::schema()} command in the console and navigate
to traces > scatter > attributes > line to see all the ways the line
can be formatted. Most common will be:
\itemize{
\item color Line color.
\item dash Plotting character. See \code{plotly::schema()}, traces > scatter > attributes > line > dash > values.
\item width Thickness in points.
Example: \code{line = list(color = "red", dash = "longdash", width = 2)}
}}

\item{\dots}{Other parameters to pass to \code{\link{plot.PMsim}}.}
}
\value{
Plots the simulation profiles with MMoptimal times indicated as vertical lines.
}
\description{
Plots \emph{MMopt} objects
}
\details{
Simulated observations are plotted on the y-axis vs. time on the x.axis.
Optimal sampling times are indicated as vertical lines. Defaults for optimal
sample times are red, dash, width 2. Defaults for the \code{line} format are as
for \link{plot.PMsim}.
}
\seealso{
\code{\link{plot.PMsim}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{axis}}

Other PMplots: 
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PMcov}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMdopt}()},
\code{\link{plot.PMpta}()},
\code{\link{plot.PMsim}()},
\code{\link{plot.PMvalid}()}
}
\author{
Michael Neely
}
\concept{PMplots}
