% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMMopt.R
\name{plot.MMopt}
\alias{plot.MMopt}
\title{Plot Pmetrics Multiple-Model Optimal Sampling Objects}
\usage{
\method{plot}{MMopt}(x, line = list(probs = NA), times = T, ...)
}
\arguments{
\item{x}{The name of an \emph{MMopt} data object generated by \code{\link{MMopt}}}

\item{line}{Passed to \link{plot.PM_sim} with default as \code{list(probs = NA)}.}

\item{times}{Format the vertical lines for optimal times. Default is
dashed red line. r template("line")`}

\item{\dots}{Other parameters to pass to \link{plot.PM_sim}.}

\item{probs}{Default is NA. See \link{plot.PMsim} for details.}
}
\value{
Plots the simulation profiles with MMoptimal times indicated as vertical lines.
}
\description{
Plots \emph{MMopt} objects
}
\details{
Simulated observations are plotted on the y-axis vs. time on the x.axis.
Optimal sampling times are indicated as vertical lines. Defaults for optimal
sample times are red, dash, width 2. Defaults for the \code{line} format are as
for \link{plot.PMsim}.
}
\seealso{
\link{plot.PM_sim}

Other PMplots: 
\code{\link{plot.PM_cycle}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_pta}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()},
\code{\link{plot.PMcov}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMdopt}()}
}
\author{
Michael Neely
}
\concept{PMplots}
