% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMcycle.R
\name{plot.PM_cycle}
\alias{plot.PM_cycle}
\title{Plot Cycle Information}
\usage{
\method{plot}{PM_cycle}(
  x,
  line = T,
  marker = T,
  colors,
  linetypes,
  omit,
  grid = T,
  xlab,
  ylab,
  ...
)
}
\arguments{
\item{x}{The name of a [PM_cycle]  object generated by [makeCycle].}

\item{line}{`r template("line")` Default = `list(color = "dodgerblue", width = 2, dash = "solid")`.
**Note** The width will apply to all plots, but `color` and `dash` will only apply
to the first three plots (log-likelihood, AIC, gamma/lambda). Use `colors` and `linetypes`
below to control the appearance
of the line traces for the normalized plots, because each of those traces is mapped to 
a parameter.}

\item{marker}{`r template("marker")` Here, the observation controlled is the value of a given
trace at a specific cycle number. Default = `list(symbol = "circle", color = "dodgerblue", size = 4)`. 
**Note** the marker color for the normalized parameter value plots will be controlled by the `colors`
parameter below, but size and symbol will apply to all plots.}

\item{colors}{to use for normalized parameter value line traces. This can be a palette or a vector of colors.
For accepted palette names see `RColorBrewer::brewer.pal.info`. Examples include
"BrBG", or "Set2". An example vector could be `c("red", "green", "blue")`. It is not
necessary to specify the same number of colors parameters to be plotted, as colors
will be interpolated to generate the correct number. The default when `color`
is not specified is the "Spectral" palette.
#' @param linetypes to use for normalized parameter value line traces. 
See `plotly::schema()`, traces > scatter > 
attributes > line > dash > values.
An example vector could be `c("solid", "dash", "longdash")`. It is not
necessary to specify the same number of linetype parameters to be plotted, as they
will be recycled to generate the correct number. The default when `linetypes`
is not specified is "solid".}

\item{omit}{Decimal between 0 and 1 specifying the proportion of "burn-in" cycles to omit from the plots.  If missing,
the first 20\% will be omitted.}

\item{grid}{`r template("grid")`}

\item{xlab}{Controls the formatting of the x-axis label. The text is fixed by the function and cannot be altered.
Use the plotly `plotly::schema()` command in the console and navigate
to layout > layoutAttributes > xaxis > title to see the ways to customize
this axis label.
In addition to the plotly attributes, a custom Pmetrics attribute `bold` 
may be included as a list element, either on its own or within the font
list. The default for `bold` is `TRUE`.<br>
<br>
Examples:
\itemize{
\item{`xlab = list(bold = F, font = list(color = "red", family = "Arial", size = 10))`}
\item{`xlab = list(font = list(bold = T))`}
}}

\item{ylab}{Format for y-axis label.
This argument maps to the the yaxis title element of the layout object in plotly. 
See `xlab` for details. If `xlab` is specified as a list with formatting,
then the formatting for the 
`xlab` will be applied to the `ylab`. To format `ylab` independently,
specify a formatting list as for `xlab`.<br>}

\item{\dots}{Additional R plotting parameters.}
}
\value{
Plots a panel with the following windows: -2 times the log-likelihood at each cycle, gamma/lambda at
each cycle; Akaike Information Criterion at each cyle and Bayesian (Schwartz) Information Criterion
at each cycle, the mean parameter values at each cycle (normalized to starting values); the normalized
standard deviation of the population distribution for each parameter at each cycle; and
the normalized median parameter values at each cycle.
}
\description{
Plot *PM_cycle* objects
}
\examples{
NPex$cycle$plot()
NPex$cycle$plot(omit = 0, marker = list(symbol = "cross"), line = list(width = 1))
NPex$cycle$plot(linetypes = "dash", colors = "Blues", marker = list(size = 1), line = list(width = 3))#' NPex$cycle$plot(grid = F, xlab = list(bold = F, font = list(color = "red", family = "Arial", size = 10)))
}
\seealso{
\code{\link{makeCycle}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{axis}}

Other PMplots: 
\code{\link{plot.MMopt}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_pta}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()},
\code{\link{plot.PMcov}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMdopt}()}
}
\author{
Michael Neely
}
\concept{PMplots}
