% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMmatrix.R
\name{plot.PM_data}
\alias{plot.PM_data}
\title{Plot PM_data Time-Output Data}
\usage{
\method{plot}{PM_data}(
  x,
  include,
  exclude,
  pred = NULL,
  mult = 1,
  outeq = 1,
  group,
  block = 1,
  log = F,
  join = T,
  marker = T,
  grid = F,
  legend = F,
  tad = F,
  overlay = T,
  xlim,
  ylim,
  xlab = "Time",
  ylab = "Output",
  ...
)
}
\arguments{
\item{x}{The name of an \code{PM_data} data object created by \code{\link[=PM_data$new]{PM_data$new()}} or loaded as a field
in a \link{PM_result} object}

\item{include}{A vector of subject IDs to include in the plot, e.g. c(1:3,5,15)}

\item{exclude}{A vector of subject IDs to exclude in the plot, e.g. c(4,6:14,16:20)}

\item{pred}{The name of a population or posterior prediction object in a \link{PM_result} object, eg.
\code{run1$post} or \code{run1$pop}. Can be a list, with the prediction object first, followed by named options to control the
prediction plot
\itemize{
\item icen Chooses the median or mean of each
subject's Bayesian posterior parameter distribution.  Default is "median", but could be "mean".
\item Other parameters to pass to plotly to control line characteristics, including \code{color}, \code{width}, and \code{dash}.
For example: \code{pred = list(run1$post, color = "blue", width = 2)}. Defaults are "dodgerblue", 1, and 1 (solid line).
Type \code{schema()} into the console (with the plotly package loaded), select traces > scatter > attributes > line
for more details on the available options for line formatting.
}}

\item{mult}{Multiplication factor for y axis, e.g. to convert mg/L to ng/mL}

\item{outeq}{A vector of output equation(s) to plot; if missing, plot all.  E.g. outeq=1, outeq=2, outeq=c(1,3).}

\item{group}{\strong{Not currently implemented.} Quoted name of a covariate in \code{data} by which
to distinguish groups with color in the plot. Note that if covariates do not have values on observation
rows, those observations will be unable to be grouped.  Grouping is only applicable if \code{outeq} is
specified; otherwise there would be a confusing mix of colors for groups and output equations.}

\item{block}{Which block to plot, where a new block is defined by dose resets (evid=4); default is 1.}

\item{log}{Boolean operator to plot in log-log space; the default is \code{FALSE}.}

\item{join}{Either \code{TRUE} (the default), \code{FALSE} to suppress, or a list
controlling the joining line segments between observations. Type \code{schema()}
into the console (with the plotly package loaded), select traces > scatter > attributes > line
for more details on the available options for marker formatting. The most relevant are \code{color}, \code{width},
and \code{dash}. For example: \code{join = list(color = "green", width = 2)}
Defaults are the same as for \code{pred}: "dodgerblue", 1, and 1 (solid).}

\item{marker}{Either \code{TRUE} (the default), \code{FALSE} to suppress, or a list
controlling the plotting symbol for observations. Type \code{schema()}
into the console (with the plotly package loaded), select traces > scatter > attributes > marker
for more details on the available options for marker formatting. The most relevant are \code{color}, \code{size},
\code{symbol}, and \code{opacity}. For example: \code{marker = list(color = "green", size = 15)}
Defaults are red, 10, circle, and 0.5.}

\item{grid}{Either \code{TRUE} or \code{FALSE} (the default) to plot or suppress the grid.  It may also be a list with either
or both of \code{gridcolor} or \code{gridwidth}. Type \code{schema()}
into the console (with the plotly package loaded), select layout > layoutAttributes > xaxis > gridcolor or gridwidth
for more details on the available options for formatting. Defaults if plotted are grey and 1.}

\item{legend}{Either a boolean operator (most common) or a list of parameters to be supplied to plotly.
If \code{FALSE} (the default), a legend will not be plotted. Type \code{schema()}
into the console (with the plotly package loaded), select layout > layoutAttributes > legend
for more details on the available options for formatting.}

\item{tad}{Boolean operator to use time after dose rather than time after start.  Default is \code{FALSE}.}

\item{overlay}{Boolean operator to overlay all time concentration profiles in a single plot.
The default is \code{TRUE}.}

\item{xlim}{Optional to specify the limits for the x axis. This is a shortcut for the plotly method
of specifying xaxis = list(range = c(min, max)).}

\item{ylim}{Optional to specify the limits for the y axis. It's a shortcut similar to \code{xlim}.}

\item{xlab}{Label for the x axis.  Default is "Time". This is a shortcut for the plotly method
of specifying xaxis = list(title = xlab).}

\item{ylab}{Label for the y axis.  Default is "Observation". It's a shortcut similar to \code{xlab}.}

\item{...}{Not currently implemented.}
}
\value{
Plots the object.
}
\description{
\code{plot.PM_data} plots \emph{PM_data} objects
}
\details{
This function will plot raw and fitted time and concentration data with a variety of options.
}
\seealso{
\link{PM_data}, \link{PM_result}
}
\author{
Michael Neely
}
