% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMfinal.R
\name{plot.PM_final}
\alias{plot.PM_final}
\title{Plot Pmetrics Final Cycle Parameter Value Distributions}
\usage{
\method{plot}{PM_final}(
  x,
  formula,
  marker = T,
  line = T,
  density = F,
  grid = T,
  xlab,
  ylab,
  standardize
)
}
\arguments{
\item{x}{The name of an \link{PM_final} data object as a field in a \link{PM_result} R6
object, e.g \code{PM_result$final}.}

\item{formula}{An optional formula of the form \code{y ~ x}, where \code{y} and \code{x}
are two model parameters to plot in a 3-dimensional bivariate plot.  See details.}

\item{marker}{See details for which objects the \code{marker} argument controls.
This argument maps to the plotly marker object.
It can be boolean or a list.
\code{TRUE} will plot the marker with default characteristics.
\code{FALSE} will suppress marker plotting.
If a list, can control many marker characteristics, including overriding defaults.
Use the plotly \code{plotly::schema()} command in the console and navigate
to traces > scatter > attributes > marker to see all the ways the marker
can be formatted. Most common will be:
\itemize{
\item \code{color} Fill color for NPAG bars, marker color for bivariate NPAG plots.
Ignored for IT2B plots.
\item \code{symbol} Plotting character. See \code{plotly::schema()}, traces > scatter >
attributes > marker > symbol > values. Only relevant for bivariate NPAG plots.
\item \code{size} Character size in points. Only relevant for bivariate NPAG plots.
\item \code{opacity} Bar fill color for univariate NPAG plots or marker opacity for
bivariate NPAG plots. Ignored for IT2B plots.
Ranges between 0 (fully transparent) to 1 (fully opaque).
\item \code{line} A list of  additional attributes governing the outline for bars in
univariate NPAG plots or markers in bivariate NPAG plots. Ignored for IT2B plots.
\itemize{
\item \code{color} Outline color. Default is "black".
\item \code{width} Outline width. Default is 1.
Example: \code{marker = list(color = "red", symbol = "triangle", opacity = 0.8, line = list(color = "black", width = 2))}
}
}}

\item{line}{Controls characteristics of lines.
This argument maps to plotly trace line attributes.
\code{TRUE} will plot default lines. \code{FALSE} will suppress lines.
If a list, can control many line characteristics, including overriding defaults.
Use the plotly \code{plotly::schema()} command in the console and navigate
to traces > scatter > attributes > line to see all the ways the line
can be formatted. Most common will be:
\itemize{
\item color Line color.
\item dash Plotting character. See \code{plotly::schema()}, traces > scatter > attributes > line > dash > values.
\item width Thickness in points.
Example: \code{line = list(color = "red", dash = "longdash", width = 2)}
}}

\item{density}{Boolean operator to plot a kernel density function overlying the histogram
of a univariate marginal parameter distribution from NPAG; the default is \code{FALSE}.
See \link{density}.  Ignored for IT2B output.}

\item{grid}{Controls grid display.
This argument maps to the xaxis and yaxis layout objects in plotly.
Use the plotly \code{plotly::schema()} command in the console and navigate
to layout > layoutAttributes > xaxis or yaxis > gridcolor or gridwidth.
It is a Boolean operator. If \code{FALSE}, no grid is plotted.
If \code{TRUE}, the default color \emph{grey50} and width 1 will be plotted at
major tick marks. If a list,
color and width can be customized.
Examples:
\itemize{
\item \code{grid = F}
\item \code{grid = list(gridcolor = "black", gridwidth = 2)}
}}

\item{xlab}{Define x-axis label for bivariate NPAG or IT2B plot.  Default is the name of the plotted x-variable.}

\item{ylab}{Define y-axis label for bivariate NPAG or IT2B plot.  Default is the name of the plotted y-variable.}

\item{standardize}{Standardize the normal parameter distribution plots from IT2B to the same
scale x-axis.  Ignored for NPAG output.}
}
\value{
Plots the object.
}
\description{
Plot R6 \link{PM_final} objects made by \link{makeFinal} and loaded as a field in the
\link{PM_result} object, e.g. \code{PM_result$final}.
}
\details{
If \code{formula} is omitted, this will generate a marginal plot for each parameter.
For NPAG data, this will be a histogram of marginal values for each parameter and the associated probability
of that value.  For IT2B, this will be a series of normal distributions with mean and standard deviation
equal to the mean and standard deviation of each parameter marginal distribution.  IF \code{formula} IS specified,
this will generate one of two plots.

On the other hand, if \code{formula} is two parameters, e.g. CL~V, this will generate a bivariate plot.
For NPAG data, it will be support point with size proportional to the probability
of each point.  For IT2B, it will be an elliptical distribution of a bivariate normal distribution centered at the mean
of each plotted variable and surrounding quantiles of the bivariate distribution plotted in decreasing shades of grey.

The \code{line} argument is used to format:
\itemize{
\item the density line drawn from an NPAG \link{PM_final object} if \code{density = T}
\item the drop lines from an NPAG \link{PM_final object} when a \code{formula} is specified
to generate a bivariate plot
\item the lines drawing the normal distribution
of parameter values from an IT2B \link{PM_Final} object.
The \code{marker} argument is used to format:
\item the bar fill color and opacity, and the outline color and width
from an NPAG \link{PM_final object}
\item the markers used from an NPAG \link{PM_final object} when a \code{formula} is specified
to generate a bivariate plot
}
}
\examples{
#NPAG
NPex$final$plot()
NPex$final$plot(density = T)
NPex$final$plot(Ke ~ V)
#IT2B
ITex$final$plot()
ITex$final$plot(Ke ~ V)
}
\seealso{
\link{PM_final}, \link{schema}

Other PMplots: 
\code{\link{plot.MMopt}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_pta}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()},
\code{\link{plot.PMcov}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMdopt}()}
}
\author{
Michael Neely
}
\concept{PMplots}
