% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_op.R
\name{plot.PM_op}
\alias{plot.PM_op}
\title{Plot Pmetrics Observed vs. Predicted Objects}
\usage{
\method{plot}{PM_op}(
  x,
  line = list(lm = NULL, loess = NULL, ref = NULL),
  marker = TRUE,
  resid = FALSE,
  icen = "median",
  pred.type = "post",
  outeq = 1,
  block,
  include,
  exclude,
  mult = 1,
  legend,
  log = FALSE,
  grid = TRUE,
  xlab,
  ylab,
  title,
  stats = TRUE,
  print = TRUE,
  xlim,
  ylim,
  ...
)
}
\arguments{
\item{x}{The name of a \link{PM_op} data object
and loaded with \link{PM_load} as a field in a \link{PM_result}, e.g. \code{PM_result$op}.}

\item{line}{Controls characteristics of lines. Unlike
some other Pmetrics plots, for plot.PM_op, \code{line} is a list of
three elements:
\itemize{
\item \code{lm}  If set to \code{TRUE} (default) or a list of plotly line attributes,
will generate a linear regression of the form obs ~ pred.
Line attributes will control the appearance of the regression
line and the confidence interval around the line. If set to
\code{FALSE}, no linear regression will be generated. The default
values for the elements of the \code{lm} list, all of which can be
overriden are:
\itemize{
\item \code{ci} Confidence interval around the regression, default 0.95.
\item \code{color} Color of the regression line and the confidence area around
the line, but at opacity = 0.2. Default is "dodgerblue".
\item \code{width }Width of the regression line, default 1.
\item \code{dash} See \code{plotly::schema()}, traces > scatter > attributes >
line > dash > values. Default is "solid".
Example: \code{line = list(lm = list(color = "red", dash = "longdash", width = 2))}
}
\item \code{loess} If set to \code{TRUE} (default is \code{FALSE}) or a list of plotly line attributes,
will generate a loess regression of the form obs ~ pred.
The list elements and default values in the \code{loess} list are the
same as for \code{lm} except the default style is "dash".
Example: \code{line = list(lm = FALSE, loess = TRUE)}
\item \code{ref} If set to \code{TRUE} (default) or a list of plotly line attributes,
will generate a reference line with slope = 1 and intercept = 0.
The default values for the elements of the \code{ref} list are:
\itemize{
\item \code{color} "grey".
\item \code{width} 1.
\item \code{dash} "dot".
Note that there is no \emph{ci} argument for the \emph{ref} list.
Example: \code{line = list(lm = FALSE, loess = TRUE, ref = list(color = "lightgrey"))}
If the \code{line} argument is missing, it will be set to
\code{line = list(lm = TRUE, loess = FALSE, ref = TRUE)}, i.e. there will be a linear
regression with reference line, but no loess regression. However, if \code{resid = T},
the default will become \code{line = list(lm = FALSE, loess = TRUE, ref = TRUE)}, i.e.,
loess regression with reference line, but no linear regression.
}
}}

\item{marker}{\verb{r template("marker")} Default is
\code{marker = list(color = orange, shape = "circle", size = 10, opacity = 0.5, line = list(color = black, width = 1))}.
The color of any BLQ points is set to a color 90 degrees different on the color wheel from the
color of the other points using \link{opposite_color} to ensure good contrast. The symbol for BLQ points is fixed to "triangle-down", and
the opacity is fixed to 1. Size is the same as for other points.}

\item{resid}{Boolean operator to generate a plot of weighted prediction error vs. time,
a plot of weighted prediction error vs. prediction. Prediction error is
pred - obs. By default a loess regression will indicate deviation from
zero prediction error.}

\item{icen}{\verb{r template("icen")}}

\item{pred.type}{Either 'post' for a posterior object or 'pop' for a population object.  Default is 'post'.}

\item{outeq}{\verb{r template("outeq")}}

\item{block}{\verb{r template("block")} Default is missing, which results in all blocks included.}

\item{include}{\verb{r template("include")}}

\item{exclude}{\verb{r template("exclude")}}

\item{mult}{\verb{r template("mult")}}

\item{legend}{Ignored for this plot.}

\item{log}{\verb{r template("log")}}

\item{grid}{\verb{r template("grid")}}

\item{xlab}{\verb{r template("xlab")}   If missing, will default to "Predicted" for
plots when \code{resid = F} and either "Time" or "Predicted" for residual plots.}

\item{ylab}{\verb{r template("ylab")}   If missing, will default to "Observed" for
plots when \code{resid = F} and either "Individual weighted residuals" or
"Population weighted residuals" for residual plots, depending on the value of
\code{pred.type}.}

\item{title}{\verb{r template("title")} Default is to have no title.}

\item{stats}{Add the statistics from linear regression to the plot. If
\code{FALSE}, will be suppressed. Default is \code{TRUE} which results in default format of
\code{list(x= 0.8, y = 0.1, font = list(color = "black", family = "Arial", size = 14, bold = FALSE))}.
The coordinates are relative to the plot with lower left = (0,0), upper right = (1,1). This
argument maps to \code{plotly::add_text()}.}

\item{print}{If \code{TRUE}, will print the plotly object and return it. If \code{FALSE}, will only return the plotly object.}

\item{xlim}{\verb{r template("xlim")}}

\item{ylim}{\verb{r template("ylim")}}

\item{...}{\verb{r template("dotsPlotly")}}
}
\value{
Plots the object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Plot PM_op objects
}
\details{
This is a function usually called by the \verb{$plot()} method for \link{PM_op} objects
within a \link{PM_result} to generate a plot of Observed vs. Predicted observations.
The function can be called directly on a \link{PM_op} object. The default
is to generate an observed vs. predicted plot of predictions based on the
median of the Bayesian posterior distributions for each subject.
Missing observations are excluded. Observations reported as BLQ (below the limit of quantification)
are indicated as downward triangles, and colored differently from other observations. They are plotted
at the reported LOQ on the observed axis and the predicted value on the predicted axis. They are not
included in any regression lines or statistics.

Clicking on a point in the plot will highlight all points from that subject. The color of the highlight
is 180 degrees different on the color wheel from the color of the other points (\link{opposite_color}), ensuring good contrast.
Clicking on the plot background will remove the highlighting.
}
\examples{
\dontrun{
NPex$op$plot()
NPex$op$plot(pred.type = "pop")
NPex$op$plot(line = list(lm = TRUE, ref = TRUE, loess = FALSE))
NPex$op$plot(line = list(loess = list(ci = 0.9, color = "green")))
NPex$op$plot(marker = list(color = "blue"))
NPex$op$plot(resid = TRUE)
NPex$op$plot(stats = list(x = 0.5, y = 0.2, font = list(size = 7, color = "blue")))
}
}
\seealso{
\link{PM_result}, \link{PM_op}, \link[plotly:schema]{schema}

Other PMplots:
\code{\link[=plot.PM_cov]{plot.PM_cov()}},
\code{\link[=plot.PM_cycle]{plot.PM_cycle()}},
\code{\link[=plot.PM_data]{plot.PM_data()}},
\code{\link[=plot.PM_final]{plot.PM_final()}},
\code{\link[=plot.PM_model]{plot.PM_model()}},
\code{\link[=plot.PM_opt]{plot.PM_opt()}},
\code{\link[=plot.PM_pop]{plot.PM_pop()}},
\code{\link[=plot.PM_post]{plot.PM_post()}},
\code{\link[=plot.PM_pta]{plot.PM_pta()}},
\code{\link[=plot.PM_sim]{plot.PM_sim()}},
\code{\link[=plot.PM_valid]{plot.PM_valid()}}
}
\author{
Michael Neely
}
\concept{PMplots}
