% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_pop.R
\name{plot.PM_pop}
\alias{plot.PM_pop}
\title{Plot PM_pop Prediction Data}
\usage{
\method{plot}{PM_pop}(
  x,
  include = NULL,
  exclude = NULL,
  line = TRUE,
  marker = TRUE,
  color = NULL,
  colors = "Set1",
  names = NULL,
  mult = 1,
  icen = "median",
  outeq = 1,
  block = 1,
  overlay = TRUE,
  legend = FALSE,
  log = FALSE,
  grid = FALSE,
  xlab = "Time",
  ylab = "Output",
  title = "",
  xlim,
  ylim,
  ...
)
}
\arguments{
\item{x}{The name of a \link{PM_pop}  object, e.g. \code{NPex$pop}.
in a \link{PM_result} object}

\item{include}{\verb{r template("include")}}

\item{exclude}{\verb{r template("exclude")}}

\item{line}{Controls characteristics of lines as for all plotly plots.
It can either be a boolean or a list. If set to \code{TRUE} or
a list of plotly line attributes, it
will generate line segments joining observations. If set to
\code{FALSE}, no segments will be generated. The default
values for the elements of formatting list, all of which can be
overriden are:
- \code{color} Color of the segments. Default is "dodgerblue".
- \code{width }Width of the segments, default 1.
- \code{dash} See \code{plotly::schema()}, traces > scatter > attributes >
line > dash > values. Default is "solid".
Example: \code{line = list(color = "red", dash = "longdash", width = 2)}}

\item{marker}{Formats the symbols plotting observations. \verb{r template("marker")}}

\item{color}{Character vector naming a column to \strong{group} by, e.g. "outeq" or "icen".}

\item{colors}{to use for \strong{groups} when there are multiple \code{outeq}, \code{block}, or \code{icen}.
This can be a palette or a vector of colors.
For accepted palette names see \code{RColorBrewer::brewer.pal.info}. Examples include
"BrBG", or "Set2". An example vector could be \code{c("red", "green", "blue")}. It is not
necessary to specify the same number of colors as groups within \code{color}, as colors
will be interpolated to generate the correct number. The default when \code{color}
is specified is the "Set1" palette.}

\item{names}{A character vector of names to label the \strong{groups} if \code{legend = TRUE}.
This vector does need to be the same length as the number of groups.
Example: \code{c("Mean", "Median")} if \code{icen = c("mean", "median")}.}

\item{mult}{\verb{r template("mult")}}

\item{icen}{Can be "median" for the predictions based on median of
parameter value distributions, "mean", or both, e.g. \code{c("mean", "median")}.  Default is "median".}

\item{outeq}{\verb{r template("outeq")} Default is 1, but can be multiple if present in the data, e.g. \code{1:2} or \code{c(1, 3)}.}

\item{block}{\verb{r template("block")} Default is 1, but can be multiple if present in the data, as for \code{outeq}.}

\item{overlay}{Operator to overlay all time prediction profiles in a single plot.
The default is \code{TRUE}. If \code{FALSE}, will trellisplot subjects one at a time. Can also be
specified as a vector with number of rows and columns, e.g. \code{c(3, 2)} for 3 rows and
2 columns of subject splots to include in each trellis.}

\item{legend}{Not used for this plot.}

\item{log}{\verb{r template("log")}}

\item{grid}{\verb{r template("grid")}}

\item{xlab}{\verb{r template("xlab")} Default is "Time".}

\item{ylab}{\verb{r template("ylab")} Default is "Output".}

\item{title}{\verb{r template("title")} Default is to have no title.}

\item{xlim}{\verb{r template("xlim")}}

\item{ylim}{\verb{r template("ylim")}}

\item{...}{\verb{r template("dotsPlotly")}}
}
\value{
Plots the object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Plots \emph{PM_pop} objects
}
\details{
This is a function usually called by the \verb{$plot()} method for \link{PM_pop} objects
within a \link{PM_result} to generate the plot.
However, the function can be called directly on a \link{PM_pop} object.
This function will plot time and population predictions with a variety of options.
By default markers are included and  have the following plotly properties:
\code{list(symbol = "circle", color = "red", size = 10, opacity = 0.5, line = list(color = "black", width = 1))}.
Markers can be joined by lines, default is \code{TRUE}. If \code{TRUE},
the joining lines will have the following properties:
\verb{list(color = "dodgerblue", width = 1, dash = "solid"}.
The grid and legend are omitted by default.
}
\examples{
\dontrun{
# basic spaghetti plot
NPex$pop$plot()
# format line and marker
NPex$pop$plot(
  marker = list(color = "blue", symbol = "square", size = 12, opacity = 0.4),
  line = list(color = "orange")
)
}
}
\seealso{
\link{PM_pop}, \link{PM_result}

Other PMplots: 
\code{\link{plot.PM_cov}()},
\code{\link{plot.PM_cycle}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_model}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_opt}()},
\code{\link{plot.PM_post}()},
\code{\link{plot.PM_pta}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()}
}
\author{
Michael Neely
}
\concept{PMplots}
