% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMpta.R
\name{plot.PM_pta}
\alias{plot.PM_pta}
\title{Plot PM_pta Percent Target Attainment objects}
\usage{
\method{plot}{PM_pta}(
  x,
  include,
  exclude,
  type = "pta",
  mult = 1,
  outeq = 1,
  line = T,
  marker = T,
  ci = 0.9,
  legend = T,
  log = F,
  grid = T,
  xlab,
  ylab,
  xlim,
  ylim,
  ...
)
}
\arguments{
\item{x}{The name of an *PM_pta" data object read by \link{makePTA}}

\item{include}{A vector of subject IDs to include in the plot, e.g. c(1:3,5,15)}

\item{exclude}{A vector of subject IDs to exclude in the plot, e.g. c(4,6:14,16:20)}

\item{type}{Character vector controlling type of plot.
Default is "pta", which plots proportion with success on the y-axis and target on the x-axis.
The other choice is "pdi", which plots the median pdi (pharmacodynamic index), e.g. AUC/MIC, on the
y-axis, and target on the x-axis.}

\item{mult}{Multiplication factor for y axis, e.g. to convert mg/L to ng/mL}

\item{outeq}{Which output equation to plot. Default is 1.}

\item{line}{Controls characteristics of lines.
This argument maps to the plotly line object.
It can be boolean or a list.
\code{TRUE} will plot the line with default characteristics for each simulated regimen.
\code{FALSE} will suppress line plotting.
If a list, it functions a little differently than other Pmetrics plotly functions.
Rather than controlling individual line characteristics, for this plot,
the \code{line} argument should be a list of the options for group based plotting,
where each group corresponds to a simulated regimen. The possible elements of the
\code{line} list should be exactly named:
\itemize{
\item color Maps to the \link{plot_ly} \code{colors} argument to override default colors
applied to the lines for each regimen. This can be a named palette, which
can be obtained with \code{RColorBrewer::display.brewer.all()} or a vector of hexadecimal
color names. One way to ensure reliable color palettes is to use the
\href{https://colorbrewer2.org/#type=qualitative&scheme=Accent&n=6}{ColorBrewer} site.
Choosing the number of data classes to correspond to regimens, and qualitative data
results in a distinct palette. Easiest importing into R is to copy/paste the Export
of JavaScript on the ColorBrewer website. The default is "Spectral". Palettes
with fewer colors than regimens will be recycled. A color can also be a character
vector of color names, recycled as needed. For example, a print-friendly choice
is \code{line = list(color = "black")}.
\item width Maps to the \link{plot_ly} \code{width} argument to override default widths
applied to the lines for each regimen. All lines will have the same width.
The default value is 2.
\item dash Maps to the \link{plot_ly} \code{linetypes} argument to override default styles
applied to the lines for each regimen. If numeric, will map to \code{lty} \link{par} values.
It can also be a character vector of dash names as listed in \link{plot_ly}.
Example: \code{line = list(color = "Blues", width = 1, dash = 2)}, whicb will result
in dotted lines (dash  = 2) all with width 1 but in different shades of blue.
}}

\item{marker}{Controls the plotting symbol.
This argument maps to the plotly marker object.
It can be boolean or a list.
\code{TRUE} will plot the profiles with default characteristics for each simulated regimen.
\code{FALSE} will suppress line plotting.
If a list, it functions a little differently than other Pmetrics plotly functions.
Rather than controlling individual marker characteristics, for this plot,
the \code{marker} argument should be a list of the options for group based plotting,
where each group corresponds to a simulated regimen. The possible elements of the
\code{marker} list should be exactly named:
\itemize{
\item color Default marker color is the same as the line color. If line color is specified,
marker color does not need to also be specified. Even if line plotting is suppressed
with \code{line = F}, the default color value of "Spectral" will be applied to markers,
unless specified, e.g. \code{marker = list(color = "Blues")}.
\item symbol Maps to the \link{plot_ly} \code{symbols} argument to override default symbols
applied to the markers for each regimen. If only one value is supplied for this,
it will be recycled for each regimen, i.e. all will have the same symbol.
See \code{plotly::schema()}, traces > scatter > attributes > marker > symbol > values
for options.
\item size Maps to the \link{plot_ly} \code{size} argument to override default size
applied to the markers for each regimen. All markers will have the same size.
The default value is 12.
}}

\item{ci}{Confidence interval around curves on \code{type = "pdi"} plot, on scale of 0 to 1. Default is 0.9.}

\item{legend}{Controls display of legend. This argument maps to the plotly layout showlegend and legend arguments. It is either a boolean operator (most common) or a list of parameters to be supplied to plotly. See \code{plotly::schema()} > layout > layoutAttributes > legend and showlegend for more details on the available options for formatting.  If legend is supplied as a list,  the plotly layout > layoutAttributes > showlegend value will be set to \code{TRUE} automatically. Examples: * \code{legend = T} * \code{legend = list(orientation = "h", font = list(color = "blue"))} Default will be the labeled regimen names supplied during \link{makePTA},
or if missing, "Regimen 1, Regimen 2,...Regimen n", where \emph{n} is the number of
regimens in the PM_pta object.}

\item{log}{Boolean operator to plot the y axis in log base 10. This argument maps to the   the yaxis type value in the layout object in plotly.  Use the plotly \code{plotly::schema()} command in the console and navigate to layout > layoutAttributes > yaxis > type. Example: \code{log = T}}

\item{grid}{Controls grid display. This argument maps to the xaxis and yaxis layout objects in plotly.  Use the plotly \code{plotly::schema()} command in the console and navigate to layout > layoutAttributes > xaxis or yaxis > gridcolor or gridwidth. It is a Boolean operator. If \code{FALSE}, no grid is plotted.  If \code{TRUE}, the default color \emph{grey50} and width 1 will be plotted at  major tick marks. If a list, color and width can be customized. Examples:  * \code{grid = F} * \code{grid = list(gridcolor = "black", gridwidth = 2)}}

\item{xlab}{Value for x axis label. This argument maps to the the xaxis layout object in plotly.  Use the plotly \code{plotly::schema()} command in the console and navigate to layout > layoutAttributes > xaxis to see other ways to customize this axis using lists as additional arguments to the plot function. Example: \code{xaxis = list(title = list(text = "Time", font = list(color = "red", family = "Arial", size = 10)))}  Default is "Target" when targets are discrete,
and "Regimen" when targets are sampled.}

\item{ylab}{Value for y axis label. This argument maps to the yaxis layout object in plotly.  Use the plotly \code{plotly::schema()} command in the console and navigate to layout > layoutAttributes > yaxis to see other ways to customize this axis using lists as additional arguments to the plot function. Example: \code{yaxis = list(title = list(text = "Concentration", font = list(color = "blue", family = "Times New Roman", size = 12)))}  Default is "Proportion with success" for
plot \code{type = "pta"} and "Pharmacodynamic Index" for plot \code{type = "pdi"}.}

\item{xlim}{Limits of the x axis as a vector.   This argument maps to the   the xaxis range in the layout object in plotly.  Use the plotly \code{plotly::schema()} command in the console and navigate to layout > layoutAttributes > xaxis > range. Example: \code{xlim = c(0,1)}}

\item{ylim}{Limits of the y axis as a vector.   This argument maps to the   the yaxis range in the layout object in plotly.  Use the plotly \code{plotly::schema()} command in the console and navigate to layout > layoutAttributes > yaxis > range. Example: \code{ylim = c(0,100)}}

\item{`r}{template("dotsPlotly")`}
}
\value{
Plots the object.
}
\description{
Plots PM_pta objects
}
\details{
This function will plot the percent target attainment for objects made with the \link{makePTA} function.
}
\seealso{
\link{makePTA}

Other PMplots: 
\code{\link{plot.MMopt}()},
\code{\link{plot.PM_data}()},
\code{\link{plot.PM_final}()},
\code{\link{plot.PM_op}()},
\code{\link{plot.PM_sim}()},
\code{\link{plot.PM_valid}()},
\code{\link{plot.PMcov}()},
\code{\link{plot.PMcycle}()},
\code{\link{plot.PMdopt}()}
}
\author{
Michael Neely
}
\concept{PMplots}
