% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMcycle.R
\name{plot.PMcycle}
\alias{plot.PMcycle}
\title{Plot NPAG Cycle Information}
\usage{
\method{plot}{PMcycle}(x, x.leg = 0, y.leg = 1, cex.leg = 1.2, omit, col, out = NA, ...)
}
\arguments{
\item{x}{The name of an \emph{PMcycle} data object generated by \code{\link{makeCycle}}}

\item{x.leg}{Porportionate location along the X-axis to place legend; 0 (default) is at left, 1 at right.}

\item{y.leg}{Porportionate location along the X-axis to place legend;  0 is at bottom, 1 (default) at top.}

\item{cex.leg}{Porportionate size of legend text.}

\item{omit}{Deceimal between 0 and 1 specifying the proportion of \dQuote{burn-in} cycles to omit from the plots.  If missing,
the first 20\% will be omitted.}

\item{col}{A vector of colors for the curves, which will be recycled if too short.  Not mandatory.}

\item{out}{Direct output to a PDF, EPS or image file.  Format is a named list whose first argument, 
\code{type} is one of the following character vectors: \dQuote{pdf}, \dQuote{eps} (maps to \code{postscript}),
\dQuote{\code{png}}, \dQuote{\code{tiff}}, \dQuote{\code{jpeg}}, or \dQuote{\code{bmp}}.  Other named items in the list
are the arguments to each graphic device. PDF and EPS are vector images acceptable to most journals
in a very small file size, with scalable (i.e. infinite) resolution.  The others are raster images which may be very
large files at publication quality dots per inch (DPI), e.g. 800 or 1200. Default value is \code{NA} which means the 
output will go to the current graphic device (usually the monitor). For example, to output an eps file,
out=list(\dQuote{eps}) will generate a 7x7 inch (default) graphic.}

\item{\dots}{Additional R plotting parameters.}
}
\value{
Plots a panel with the following windows: -2 times the log-likelihood at each cycle, gamma/lambda at
each cycle; Akaike Information Criterion at each cyle and Bayesian (Schwartz) Information Criterion
at each cycle, the mean parameter values at each cycle (normalized to starting values); the normalized
standard deviation of the population distribution for each parameter at each cycle; and
the normalized median parameter values at each cycle.
}
\description{
\code{plot.PMcycle} plots \emph{PMcycle} objects
}
\examples{
data(PMex1)
plot(cycle.1)
plot(cycle.1,omit=0)
}
\seealso{
\code{\link{makeCycle}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{axis}}
}
\author{
Michael Neely
}
