% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMvalid.R
\name{plot.PMvalid}
\alias{plot.PMvalid}
\title{Plot Pmetrics Validation Objects}
\usage{
\method{plot}{PMvalid}(
  x,
  type = "vpc",
  tad = F,
  icen = "median",
  lower = 0.025,
  upper = 0.975,
  log = F,
  pch.obs = 1,
  col.obs = "black",
  cex.obs = 1,
  theme = "color",
  col.obs.ci = "blue",
  col.obs.med = "red",
  col.sim.ci = "dodgerblue",
  col.sim.med = "lightpink",
  xlab = "Time",
  ylab = "Observation",
  xlim,
  ylim,
  ...
)
}
\arguments{
\item{x}{The name of an \emph{PMvalid} data object generated by \code{\link{makeValid}}.}

\item{type}{Default is \dQuote{vpc} for a visual prective check, but could be \dQuote{pcvpc} for a 
prediction-corrected visual predictive check.}

\item{tad}{Plot using time after dose if \code{TRUE}.  Default is \code{FALSE} which plots using standard
relative time.  This will be the only option if \code{tad} was not set to \code{TRUE} when making the 
PMvalid object.}

\item{icen}{Can be either \dQuote{median} for the predictions based on medians of the population parameter value
distributions, or \dQuote{mean}.  Default is \dQuote{median}.}

\item{lower}{The lower quantile displayed for the observed and simulated profiles. Default is 0.025.}

\item{upper}{The upper quantile displayed for the observed and simulated profiles. Default is 0.975.}

\item{log}{Boolean operator to plot in semilog space.  The default is \code{FALSE}.}

\item{pch.obs}{Control the plotting character used for observations.  Default is 1, i.e. an open circle.
See \code{\link{points}} for other values of \code{pch}.}

\item{col.obs}{Color for observations.  Default is black.}

\item{cex.obs}{Size for observatins.  Default is 1.}

\item{theme}{Default is \dQuote{color}, but could be \dQuote{grey} or \dQuote{gray}.}

\item{col.obs.ci}{Color of the observation confidence interval (set by \code{lower} and \code{upper}).
Default is blue.}

\item{col.obs.med}{Color of the observation median.
Default is red.}

\item{col.sim.ci}{Color of the simulation confidence interval (set by \code{lower} and \code{upper}).
Default is dodgerblue.}

\item{col.sim.med}{Color of the simulation median.
Default is lightpink.}

\item{xlab}{Label for x axis.  Default is \dQuote{Time}.}

\item{ylab}{Label for y axis.  Default is \dQuote{Observation}.}

\item{xlim}{Limits for x axis}

\item{ylim}{Limits for y axis}

\item{\dots}{Not currently used}
}
\value{
Plots the object using ggplot2.
}
\description{
Plot PMvalid objects
}
\seealso{
\code{\link{makeValid}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{points}}
}
\author{
Michael Neely
}
