% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_valid.R
\name{plot.PMvalid}
\alias{plot.PMvalid}
\title{Plot Pmetrics Validation Objects}
\usage{
\method{plot}{PMvalid}(
  x,
  type = "vpc",
  tad = FALSE,
  icen = "median",
  outeq = 1,
  lower = 0.025,
  upper = 0.975,
  log = FALSE,
  pch.obs = 1,
  col.obs = "black",
  cex.obs = 1,
  data_theme = "color",
  plot_theme = theme_grey(),
  col.obs.ci = "blue",
  col.obs.med = "red",
  col.sim.ci = "dodgerblue",
  col.sim.med = "lightpink",
  axis.x = NULL,
  axis.y = NULL,
  ...
)
}
\arguments{
\item{x}{The name of an \emph{PMvalid} data object generated by \code{\link{make_valid}}.}

\item{type}{Default is \dQuote{vpc} for a visual prective check, but could be \dQuote{pcvpc} for a
prediction-corrected visual predictive check.}

\item{tad}{Plot using time after dose if \code{TRUE}.  Default is \code{FALSE} which plots using standard
relative time.  This will be the only option if \code{tad} was not set to \code{TRUE} when making the
PMvalid object.}

\item{icen}{Can be either \dQuote{median} for the predictions based on medians of the population parameter value
distributions, or \dQuote{mean}.  Default is \dQuote{median}.}

\item{outeq}{The number of the output equation to simulate/test.  Default is 1.}

\item{lower}{The lower quantile displayed for the observed and simulated profiles. Default is 0.025.}

\item{upper}{The upper quantile displayed for the observed and simulated profiles. Default is 0.975.}

\item{log}{Boolean operator to plot in semilog space.  The default is \code{FALSE}.}

\item{pch.obs}{Control the plotting character used for observations.  Default is 1, i.e. an open circle.
See \code{\link{points}} for other values of \code{pch}.}

\item{col.obs}{Color for observations.  Default is black.}

\item{cex.obs}{Size for observatins.  Default is 1.}

\item{data_theme}{Default is \dQuote{color}, but could be \dQuote{grey} or \dQuote{gray}.}

\item{plot_theme}{Default is \code{theme_grey()} but could be any complete ggplot2 theme, e.g.
\code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.}

\item{col.obs.ci}{Color of the observation confidence interval (set by \code{lower} and \code{upper}).
Default is blue.}

\item{col.obs.med}{Color of the observation median.
Default is red.}

\item{col.sim.ci}{Color of the simulation confidence interval (set by \code{lower} and \code{upper}).
Default is dodgerblue.}

\item{col.sim.med}{Color of the simulation median.
Default is lightpink.}

\item{axis.x}{List of \verb{$name} and \verb{$limits}. Default name is \dQuote{Time}.}

\item{axis.y}{List of \verb{$name} and \verb{$limits}. Default name is \dQuote{Observation}.}

\item{...}{Not currently used}
}
\value{
Plots the object using ggplot2.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

This is largely now a legacy plotting function,
with a variety of options. It has been superseded by \link{plot.PM_valid}.
}
\seealso{
\code{\link{make_valid}}, \code{\link{plot}}, \code{\link{par}}, \code{\link{points}}
}
\author{
Michael Neely
}
