% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMcompare.R
\name{PM_compare}
\alias{PM_compare}
\title{Compare NPAG or IT2B runs}
\usage{
PM_compare(x, y, ..., icen = "median", outeq = 1, plot = F)
}
\arguments{
\item{x}{The first \link{PM_result} object you wish to compare. Unlike the legacy
\link{PMcompare} this function only uses objects already loaded with \link{PM_load}.
This will serve as the reference output for P-value testing (see details).}

\item{y}{The second \link{PM_result} object to compare.}

\item{...}{Additional \link{PM_result} objects to compare.  See details.
Also, parameters to be passed to \link{plot.PM_op}
if \code{plot} is true as well as to \link{mtsknn.eq}.  Order does not matter.}

\item{icen}{Can be either "median" for the predictions based on medians of
\code{pred.type} parameter value distributions, or "mean".  Default is "median".}

\item{outeq}{Number of the output equation to compare; default is 1}

\item{plot}{Boolean operator selecting whether to generate observed vs.
predicted plots for each data object as in \link{plot.PM_op}.}
}
\value{
A data frame with the following objects for each model to analyze:
\item{run }{The run number of the data}
\item{type }{NPAG or IT2B data}
\item{nsub }{Number of subjects in the model}
\item{nvar }{Number of random parameters in the model}
\item{par }{Names of random parameters}
\item{cycles }{Number of cycles run}
\item{converge }{Boolean value if convergence occurred.}
\item{ll }{Final cycle -2*Log-likelihood }
\item{aic }{Final cycle Akaike Information Criterion}
\item{bic }{Final cycle Bayesian (Schwartz) Information Criterion }
\item{popBias }{Bias, or mean weighted prediction error of predictions based on population parameters minus observations}
\item{popImp }{Imprecision, or bias-adjusted mean weighted squared error of predictions based on population parameters minus observations }
\item{popPerRMSE}{Percent root mean squared error of predictions based on population parameters minus observations}
\item{postBias }{Bias, or mean weighted prediction error of predictions - observations  based on posterior parameters}
\item{postImp }{Imprecision, or bias-adjusted mean weighted squared error of predictions - observations based on posterior parameters}
\item{postPerRMSE}{Percent root mean squared error of predictions based on posterior parameters minus observations}
\item{pval }{P-value for each model compared to the first. See details.}
}
\description{
\verb{ r lifecycle::badge("stable")}

Compare parameters, convergence, -2*log likelihood, AIC and  bias
and imprecision of population and posterior predictions.
}
\details{
Objects can be specified separated by commas, e.g. \code{PM_compare(run1, run2, run3)}
followed by any arguments you wish to \link{plot.PMop}, \link{mtsknn.eq}.
P-values are based on comparison using the nearest neighbors
approach if all models are non-parametrics.  Models may only be compared on
parameters that are included in the first model.  The P-value is the
comparison between each model and the first model in
the list.  Missing P-values are when a model has no parameter names
in common with the first model, and for the first model compared to itself,
or when models from IT2B runs are included.  Significant P-values indicate
that the null hypothesis should be rejected, i.e. the joint distributions
between the two compared models for that parameter are significantly different.
}
\seealso{
\link{PM_load}, \link{plot.PM_op}, \link{mtsknn.eq}
}
\author{
Michael Neely
}
