% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_cov.R
\name{PM_cov}
\alias{PM_cov}
\title{Contains covariate data}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Contains a data frame with subject-specific covariate data output.
}
\details{
The \link{PM_cov} object is both a data field within a \link{PM_result}, and itself an R6 object
comprising data fields and associated methods suitable for analysis and
plotting relationships between covariates and posterior parameters,
covariates over time, or parameter values over time.

Because \link{PM_cov} objects are automatically added to the \link{PM_result} at the end of a
successful run, it is generally not necessary for users to generate \link{PM_cov} objects
themselves.

The  results are contained in the \verb{$data} field,
and it is this field which is passed to the \verb{$plot} and \verb{$summary} methods.
You can use this \verb{$data} field for custom manipulations, although usually this
is better done on the \code{summary} object, e.g.,
\code{run1$cov$summary() \%>\% select(africa, gender) \%>\% table()}.
If you are unfamiliar with the \verb{\%>\%} pipe function, please type \code{help("\%>\%", "magrittr")}
into the R console and look online for instructions/tutorials in tidyverse, a
powerful approach to data manipulation upon which Pmetrics is built.

This output of this function is suitable for exploration of covariate-
parameter, covariate-time, or parameter-time relationships.
}
\seealso{
\link{plot.PM_cov}, \link{PM_step}
}
\author{
Michael Neely, Julian Otalvaro
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data frame with the following columns
\itemize{
\item id Subject identification
\item time Times of covariate observations
\item covnames... Columns with each covariate observations in the dataset for each subject and \code{time}
\item parnames... Columns with each parameter in the model and the \code{icen} summary
for each subject, replicated as necessary for covariate observation times and duplicated for Bayesian
parameter means and medians
\item icen The type of summarized Bayesian posterior individual parameter values: mean or median
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_cov-new}{\code{PM_cov$new()}}
\item \href{#method-PM_cov-step}{\code{PM_cov$step()}}
\item \href{#method-PM_cov-summary}{\code{PM_cov$summary()}}
\item \href{#method-PM_cov-plot}{\code{PM_cov$plot()}}
\item \href{#method-PM_cov-print}{\code{PM_cov$print()}}
\item \href{#method-PM_cov-clone}{\code{PM_cov$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cov-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cov-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with covariate-parameter information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$new(PMdata = NULL, path = ".", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PMdata}}{include Saved, parsed output of prior run, used when source files are not available.
.}

\item{\code{path}}{include Path to the folder containing the raw results of the run. Default is the current working directory.
.}

\item{\code{...}}{Not currently used.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creation of new \code{PM_cov} object is automatic and not generally necessary
for the user to do.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cov-step"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cov-step}{}}}
\subsection{Method \code{step()}}{
Stepwise linear regression of covariates and Bayesian posterior
parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$step(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{PM_step}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{PM_step}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cov-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cov-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PM_cov}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{summary.PM_cov}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cov-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cov-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_cov}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{plot.PM_cov}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cov-print"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cov-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{print}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Print method for \link{PM_cov}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cov-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cov-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cov$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
