% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_cycle}
\alias{PM_cycle}
\title{Pmetrics Run Cycle Information}
\description{
Pmetrics Run Cycle Information

Pmetrics Run Cycle Information
}
\details{
Contains the cycle information after a run.

This contains the output of \link{makeCycle} after a run, which
generates information suitable for analysis and plotting of cycle information.
Each field corresponds to a column in the complete data frame.

To manipulate the entire data frame,
use the \code{data} field, e.g. \code{final <- run1$cycle$data \%>\% slice_tail(n=1)}. If
you are unfamiliar with the \verb{\%>\%} pipe function, please type \code{help("\%>\%", "magrittr")}
into the R console and look online for instructions/tutorials in tidyverse, a
powerful approach to data manipulation upon which Pmetrics is built.

Creation of new \code{PM_cycle} object is automatic and not generally necessary
for the user to do.

See \link{plot.PM_cycle}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data frame with the following columns
\itemize{
\item names Vector of names of the random parameters
\item cynum Vector cycle numbers, which may start at numbers greater
than 1 if a non-uniform prior was specified for the run (NPAG only)
\item ll Matrix of cycle number and -2*Log-likelihood at each cycle
\item gamlam A matrix of cycle number and gamma or lambda at each cycle
\item mean A matrix of cycle number and the mean of each random parameter
at each cycle, normalized to initial mean
\item sd A matrix of cycle number and the standard deviation of each random parameter
at each cycle,  normalized to initial standard deviation
\item median A matrix of cycle number and the median of each random
parameter at each cycle,  normalized to initial median
\item aic A matrix of cycle number and Akaike Information Criterion at each cycle
\item bic A matrix of cycle number and Bayesian (Schwartz) Information Criterion at each cycle
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_cycle-new}{\code{PM_cycle$new()}}
\item \href{#method-PM_cycle-plot}{\code{PM_cycle$plot()}}
\item \href{#method-PM_cycle-clone}{\code{PM_cycle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cycle-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cycle-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with  cycle information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cycle$new(cycle)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cycle}}{The parsed output from \link{makeCycle}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cycle-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cycle-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cycle$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_cycle}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cycle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cycle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cycle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
