% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_cycle}
\alias{PM_cycle}
\title{Pmetrics Run Cycle Information}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Contains the cycle information after a run.
}
\details{
#' @title Wrapper function for summmary.PMfinal
#'
#' @description
#' \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
#'
#' This redirects to summary.PMfinal for PM_final R6 objects
#'
#' @details
#' See \link{summary.PMfinal}. Alternative way to summarize is
#' \code{PM_result$final$summary()}.
#'
#' @param object The \emph{PM_final} object to summarize
#' @param ... Arguments passed to \link{summary.PMfinal}
#' @return A \link{summary.PMfinal} object
#' @export
summary.PM_final <- function(object, ...) {
object$summary(...)
}

This contains the output of \link{makeCycle} after a run, which
generates information suitable for analysis and plotting of cycle information.
Each field corresponds to a column in the complete data frame.

To manipulate the entire data frame,
use the \code{data} field, e.g. \code{final <- run1$cycle$data \%>\% slice_tail(n=1)}. If
you are unfamiliar with the \verb{\%>\%} pipe function, please type \code{help("\%>\%", "magrittr")}
into the R console and look online for instructions/tutorials in tidyverse, a
powerful approach to data manipulation upon which Pmetrics is built.

Creation of new \code{PM_cycle} object is automatic and not generally necessary
for the user to do.

See \link{plot.PM_cycle}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{names}}{Vector of names of the random parameters}

\item{\code{cycnum}}{Vector cycle numbers, which may start at numbers greater
than 1 if a non-uniform prior was specified for the run (NPAG only)}

\item{\code{ll}}{Vector of -2*Log-likelihood at each cycle}

\item{\code{gamlam}}{A tibble of cycle number and gamma or lambda at each cycle for each output equation}

\item{\code{mean}}{A tibble of cycle number and the mean of each random parameter
at each cycle, normalized to initial mean}

\item{\code{median}}{A tibble of cycle number and the median of each random
parameter at each cycle,  normalized to initial median}

\item{\code{sd}}{A tibble of cycle number and the standard deviation of each random parameter
at each cycle,  normalized to initial standard deviation}

\item{\code{aic}}{A vector of Akaike Information Criterion at each cycle}

\item{\code{bic}}{A vector of Bayesian (Schwartz) Information Criterion at each cycle}

\item{\code{data}}{A data frame combining all the above fields as its columns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_cycle-new}{\code{PM_cycle$new()}}
\item \href{#method-PM_cycle-plot}{\code{PM_cycle$plot()}}
\item \href{#method-PM_cycle-clone}{\code{PM_cycle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cycle-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cycle-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with  cycle information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cycle$new(cycle)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cycle}}{The parsed output from \link{makeCycle}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cycle-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cycle-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cycle$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_cycle}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_cycle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_cycle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_cycle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
