% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_data.R
\name{PM_data}
\alias{PM_data}
\title{Defines the PM_data object}
\description{
PM_data R6 objects containing raw, standardized and valid data, and methods
to process the data
}
\details{
PM_data objects are passed to \link{PM_fit} objects to initiate a
population analysis. The object is created by reading a delimited file in
the current working directory. The data will be transformed into the standard
format which is the same for all engines, with a report of any assumptions
that were necessary to standardize the data. \link{PMcheck} is called
on the standard data to evaluate for errors. There are a number of methods
defined for a PM_data object, including to write the standard data back
to a file for future use, to summarize and to plot the object, and to
conduct a non-compartmental analysis on the raw data using
\link{makeNCA}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Data frame containing the data to be modeled}

\item{\code{standard_data}}{Data frame containing standardized version of the data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PM_data$new()}}
\item \href{#method-write}{\code{PM_data$write()}}
\item \href{#method-nca}{\code{PM_data$nca()}}
\item \href{#method-plot}{\code{PM_data$plot()}}
\item \href{#method-print}{\code{PM_data$print()}}
\item \href{#method-summary}{\code{PM_data$summary()}}
\item \href{#method-clone}{\code{PM_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$new(data, dt = c("Ymd", "HS"), quiet = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A quoted name of a file with full path if not
in the working directory, or an unquoted name of a data frame
in the current R environment.}

\item{\code{dt}}{Date time format as a character vector whose
first element is date format and second is time. Default is
\code{c("Ymd", "HS")}. Use the following abbreviations:
\itemize{
\item Y = 4 digit year
\item y = 2 digit year
\item m = decimal month (1, 2, ..., 12)
\item d = decimal day (1, 2, ..., 31)
\item H = hours (0-23)
\item M = minutes (0-59)
Other formats are possible. See \link{strptime} for these,
omitting the \% before each abbreviation.
}}

\item{\code{quiet}}{Quietly validate. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creation of a new \code{PM_data} objects from a file or
a data frame. Data will be standardized and checked
automatically to a fully specified, valid data object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write data to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$write(file_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{A quoted name of the file to create
with full path if not
in the working directory.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Writes a delimited file (e.g. comma-separated)
from the \code{standard_data} field
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nca"></a>}}
\if{latex}{\out{\hypertarget{method-nca}{}}}
\subsection{Method \code{nca()}}{
Perform non-compartmental analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$nca(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeNCA}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{makeNCA}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_data}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{plot.PMmatrix}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$print(standard = F, viewer = T, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{standard}}{Display the standardized data if \code{TRUE}.
Default is \code{FALSE}.}

\item{\code{viewer}}{Display the Viewer if \code{TRUE}.
Default is \code{TRUE}.}

\item{\code{...}}{Other arguments to \link{print.data.frame}. Only
passed if \code{viewer = FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Displays the PM_data object in a variety of ways.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMmatrix}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{summary.PMmatrix}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
