% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_data.R
\name{PM_data}
\alias{PM_data}
\title{Defines the PM_data object}
\description{
PM_data R6 objects containing raw, standardized and valid data, and methods
to process the data
}
\details{
\emph{PM_data} objects are passed to \link{PM_fit} objects to initiate a
population analysis. The object is created by reading a delimited file in
the current working directory. The data will be transformed into the standard
format which is the same for all engines, with a report of any assumptions
that were necessary to standardize the data. \link{PMcheck} is called
on the standard data to evaluate for errors. If dates and times are converted
to relative decimal times in the standard data, automatic detection of the correct
format will be attempted using \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}}. In the case of failure
due to an unusual format, use the 'dt' argument to specify the correct format in your
data. In the case of successful automatic detection, the format used will be
included in the standardization report generated upon creation of a new \emph{PM_data}
object. Check carefully to make sure the correct format was chosen. Note that if
your clock times did not include seconds, they were appended as ":00" to the end
of each time and will appear that way in the copy of the original data.

There are a number of methods
defined for a PM_data object, including to write the standard data back
to a file for future use, to summarize and to plot the object, and to
conduct a non-compartmental analysis on the raw data using
\link{makeNCA}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Data frame containing the data to be modeled}

\item{\code{standard_data}}{Data frame containing standardized version of the data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_data-new}{\code{PM_data$new()}}
\item \href{#method-PM_data-write}{\code{PM_data$write()}}
\item \href{#method-PM_data-auc}{\code{PM_data$auc()}}
\item \href{#method-PM_data-nca}{\code{PM_data$nca()}}
\item \href{#method-PM_data-plot}{\code{PM_data$plot()}}
\item \href{#method-PM_data-print}{\code{PM_data$print()}}
\item \href{#method-PM_data-summary}{\code{PM_data$summary()}}
\item \href{#method-PM_data-clone}{\code{PM_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-new}{}}}
\subsection{Method \code{new()}}{
Create new data object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$new(data, dt = NULL, quiet = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A quoted name of a file with full path if not
in the working directory, or an unquoted name of a data frame
in the current R environment.}

\item{\code{dt}}{Pmetrics will try a variety of date/time formats. If all 16 of
them fail, use this parameter to specify the correct format as a
character vector whose
first element is date format and second is time. Use the following abbreviations:
\itemize{
\item Y = 4 digit year
\item y = 2 digit year
\item m = decimal month (1, 2, ..., 12)
\item d = decimal day (1, 2, ..., 31)
\item H = hours (0-23)
\item M = minutes (0-59)
Example: \code{format = c("myd", "mh")}. Not one of the tried combinations!
Always check to make sure that dates/times were parsed correctly and the
relative times in the \code{PM_data$standard_data} field look correct.
Other date/time formats are possible. See \code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}} for these.
}}

\item{\code{quiet}}{Quietly validate. Default is \code{FALSE}.}

\item{\code{...}}{Arguments to be passed further}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creation of a new \code{PM_data} objects from a file or
a data frame. Data will be standardized and checked
automatically to a fully specified, valid data object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-write"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-write}{}}}
\subsection{Method \code{write()}}{
Write data to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$write(file_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{A quoted name of the file to create
with full path if not
in the working directory.}

\item{\code{...}}{ARguments passed to PMwriteMatrix}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Writes a delimited file (e.g. comma-separated)
from the \code{standard_data} field
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-auc"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-auc}{}}}
\subsection{Method \code{auc()}}{
Calculate AUC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$auc(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeAUC}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{makeAUC}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-nca"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-nca}{}}}
\subsection{Method \code{nca()}}{
Perform non-compartmental analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$nca(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeNCA}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{makeNCA}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_data}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{plot.PMmatrix}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-print"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-print}{}}}
\subsection{Method \code{print()}}{
Print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$print(standard = F, viewer = T, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{standard}}{Display the standardized data if \code{TRUE}.
Default is \code{FALSE}.}

\item{\code{viewer}}{Display the Viewer if \code{TRUE}.
Default is \code{TRUE}.}

\item{\code{...}}{Other arguments to \link{print.data.frame}. Only
passed if \code{viewer = FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Displays the PM_data object in a variety of ways.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMmatrix}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{summary.PMmatrix}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
