% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMload.R
\name{PM_load}
\alias{PM_load}
\title{Load Pmetrics NPAG or IT2B output}
\usage{
PM_load(run = 1, remote = F, server_address)
}
\arguments{
\item{run}{The numerical value of the folder number containing the run results.
This parameter is \code{1} by default.}

\item{remote}{Default is \code{FALSE}.  Set to \code{TRUE} if loading results of an NPAG run on remote server.
See \code{\link{NPrun}}. Currently remote runs are not configured for IT2B or the Simulator.}

\item{server_address}{If missing, will use the default server address returned by getPMoptions().
Pmetrics will prompt the user to set this address the first time the \code{remote} argument is set to \code{TRUE}
in \code{\link{NPrun}}.}
}
\value{
An R6 \code{\link{PM_result}}.
}
\description{
Loads all the data from an \emph{NPAG} or \emph{IT2B} run.
}
\seealso{
\code{\link{PMreport}}, \code{\link{NPparse}}, \code{\link{ITparse}},
\code{\link{makeFinal}}, \code{\link{makeCycle}}, \code{\link{makeOP}}, \code{\link{makeCov}},
\code{\link{makePop}}, \code{\link{makePost}}
}
\author{
Michael Neely and Julian Otalvaro
}
