% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{PM_model}
\alias{PM_model}
\title{Defines the PM_model class}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

PM_model objects contain the variables, covariates, equations and error models
necessary to run a population analysis.
}
\details{
PM_model objects are one of two fundamental objects in Pmetrics, along with
\code{\link[=PM_data]{PM_data()}} objects. Defining a PM_model allows for fitting it to the data
via the \verb{$fit()} method to conduct a
population analysis, i.e. estimating the probability distribution of model equation
paramter values in the population. The PM_model object is created using the
\code{\link[=build_model]{build_model()}} function, by defining a list of lists
directly in R, or by reading a model text file. See the vignette on models
for details.
}
\author{
Michael Neely
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_model-new}{\code{PM_model$new()}}
\item \href{#method-PM_model-write_model_file}{\code{PM_model$write_model_file()}}
\item \href{#method-PM_model-from_file}{\code{PM_model$from_file()}}
\item \href{#method-PM_model-print}{\code{PM_model$print()}}
\item \href{#method-PM_model-plot}{\code{PM_model$plot()}}
\item \href{#method-PM_model-get_primary}{\code{PM_model$get_primary()}}
\item \href{#method-PM_model-write_rust}{\code{PM_model$write_rust()}}
\item \href{#method-PM_model-clone}{\code{PM_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$new(
  pri,
  cov = NULL,
  eqn = NULL,
  sec = NULL,
  tem = NULL,
  lag = NULL,
  f = NULL,
  ini = NULL,
  out = NULL,
  error = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-write_model_file"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-write_model_file}{}}}
\subsection{Method \code{write_model_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$write_model_file(file_path = "model.rs")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-from_file"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-from_file}{}}}
\subsection{Method \code{from_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$from_file(file_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-print"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$plot(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-get_primary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-get_primary}{}}}
\subsection{Method \code{get_primary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$get_primary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-write_rust"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-write_rust}{}}}
\subsection{Method \code{write_rust()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$write_rust()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
