% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_op}
\alias{PM_op}
\title{Observed vs. predicted data}
\description{
Observed vs. predicted data

Observed vs. predicted data
}
\details{
Contains observed vs. predicted data after a run

Contains the results of \link{makeOP}, which is a
data frame suitable for analysis and plotting of observed vs. population or
or individual predicted outputs. To provide a more traditional experience in R,
the data frame is separated by columns into fields, e.g. \code{id} or \code{time}. This
allows you to access them in an S3 way, e.g. \code{run1$op$time} if \code{run1} is a
\code{PM_result} object.

However, if you wish to manipulate the entire data frame,
use the \code{data} field, e.g. \code{trough <- run1$op$data \%>\% filter(time == 24)}. If
you are unfamiliar with the \verb{\%>\%} pipe function, please type \code{help("\%>\%", "magrittr")}
into the R console and look online for instructions/tutorials in tidyverse, a
powerful approach to data manipulation upon which Pmetrics is built.

Creation of new \code{PM_op} object is automatic and not generally necessary
for the user to do.

See \link{plot.PM_op}.

See \link{summary.PMop}.

See \link{makeAUC}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{subject identification}

\item{\code{time}}{observation time in relative units, usually hours}

\item{\code{obs}}{observation}

\item{\code{pred}}{prediction}

\item{\code{pred.type}}{Population predictions based on Bayesian prior parameter value distribution,
or individual predictions based on Bayesian posterior parameter value distributions}

\item{\code{icen}}{Predictions based on mean or median of Bayesian \code{pred.type}parameter values}

\item{\code{outeq}}{output equation number}

\item{\code{block}}{dosing block number for each subject, as defined by dose resets (evid=4).}

\item{\code{obsSD}}{standard deviation of the observation based on the assay error polynomial}

\item{\code{d}}{prediction error, \code{pred} - \code{obs}}

\item{\code{ds}}{squared prediction error}

\item{\code{wd}}{weighted prediction error, which is the prediction error divided by the \code{obsSD}}

\item{\code{wds}}{weighted squared prediction error}

\item{\code{data}}{A data frame combining all the above fields as its columns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_op-new}{\code{PM_op$new()}}
\item \href{#method-PM_op-plot}{\code{PM_op$plot()}}
\item \href{#method-PM_op-summary}{\code{PM_op$summary()}}
\item \href{#method-PM_op-auc}{\code{PM_op$auc()}}
\item \href{#method-PM_op-tidy}{\code{PM_op$tidy()}}
\item \href{#method-PM_op-clone}{\code{PM_op$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_op-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_op-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with observed vs. predicted data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_op$new(op)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{op}}{The parsed output from \link{makeOP}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_op-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_op-plot}{}}}
\subsection{Method \code{plot()}}{
Plot method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_op$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_op}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_op-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_op-summary}{}}}
\subsection{Method \code{summary()}}{
Summary method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_op$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMop}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_op-auc"></a>}}
\if{latex}{\out{\hypertarget{method-PM_op-auc}{}}}
\subsection{Method \code{auc()}}{
Calculate AUC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_op$auc(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeAUC}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_op-tidy"></a>}}
\if{latex}{\out{\hypertarget{method-PM_op-tidy}{}}}
\subsection{Method \code{tidy()}}{
Returns a structure compatible with tidy methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_op$tidy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_op-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_op-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_op$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
