% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_parse.R
\name{PM_parse}
\alias{PM_parse}
\title{Parse Pmetrics output}
\usage{
PM_parse(wd = getwd(), write = TRUE)
}
\arguments{
\item{wd}{The directory containing the output from the Rust-implementation of NPAG}

\item{write}{A logical value indicating if the results should be returned (\code{FALSE}, default) or written to disk (\code{TRUE})}
}
\value{
The output of \code{PM_parse} is a list containing the following elements
\item{op }{Written to the standard of PM_op}
\item{pop }{Written to the standard of PM_pop}
\item{post }{Written to the standard of PM_post}
\item{cycles }{Written to the standard of PM_cycle}
\item{final }{Written to the standard of PM_final}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
A flexible parser for Pmetrics output
}
\details{
Currently written for the Rust implementation of NPAG
}
\seealso{
\code{\link{NPparse}}
}
