% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_pop}
\alias{PM_pop}
\title{Population predictions at short intervals}
\description{
Contains the population predictions at short intervals
specified as an argument to \link{PM_fit$run}. Default is every 12 minutes.
}
\details{
Contains the results of \link{makePop}, which is a
data frame with population predicted outputs for all subjects.
To provide a more traditional experience in R,
the data frame is separated by columns into fields, e.g. \code{id} or \code{time}. This
allows you to access them in an S3 way, e.g. \code{run1$pop$time} if \code{run1} is a
\code{PM_result} object.

However, if you wish to manipulate the entire data frame,
use the \code{data} field, e.g. \code{trough <- run1$pop$data \%>\% filter(time == 24)}. If
you are unfamiliar with the \verb{\%>\%} pipe function, please type \code{help("\%>\%", "magrittr")}
into the R console and look online for instructions/tutorials in tidyverse, a
powerful approach to data manipulation upon which Pmetrics is built.
}
\author{
Michael Neely, Julian Otalvaro
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Subject id}

\item{\code{time}}{Time of predictions in decimal hours}

\item{\code{icen}}{Prediction based on mean or median of Bayesian posterior parameter distribution}

\item{\code{outeq}}{Output equation number}

\item{\code{pred}}{Predicted output for each outeq}

\item{\code{block}}{Observation blocks within subjects as defined by \emph{EVID=4} dosing events}

\item{\code{data}}{A data frame combining all the above fields as its columns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PM_pop$new()}}
\item \href{#method-auc}{\code{PM_pop$auc()}}
\item \href{#method-clone}{\code{PM_pop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with population predicted data at
regular, frequent intervals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pop$new(pop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{The parsed output from \link{makePop}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creation of new \code{PM_pop} object is automatic and not generally necessary
for the user to do.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auc"></a>}}
\if{latex}{\out{\hypertarget{method-auc}{}}}
\subsection{Method \code{auc()}}{
Calculate AUC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pop$auc(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeAUC}}

\item{\code{data}}{The object to use for AUC calculation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
See \link{makeAUC}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
