% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_pta.R
\name{PM_pta}
\alias{PM_pta}
\title{Define object for PTA}
\arguments{
\item{file_name}{The name of the .rds file to load.}
}
\value{
A \link{PM_pta} object
}
\description{
If the \verb{$save} method has previously been invoked on a \link{PM_pta}
object, this function will load those results.
}
\details{
This object class contains results of simulations and a probability of
target attainment analysis.

Creating this object first requires a simulation of appropriate regimens
from a model using the \verb{$sim} method for a \link{PM_result} object. That simulation
result is passed to the \link{makePTA} function when a new \code{PM_pta} is created.

The saved object is an .rds file. When loaded, it should be assigned to an R
object, e.g. \code{pta1 <- PM_pta$load("filename")}.
}
\author{
Julian Otalvaro and Michael Neely
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{Contains the raw results. See \link{makePTA}.}

\item{\code{outcome}}{Contains summary of raw results. See \link{makePTA}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_pta-new}{\code{PM_pta$new()}}
\item \href{#method-PM_pta-save}{\code{PM_pta$save()}}
\item \href{#method-PM_pta-summary}{\code{PM_pta$summary()}}
\item \href{#method-PM_pta-plot}{\code{PM_pta$plot()}}
\item \href{#method-PM_pta-load}{\code{PM_pta$load()}}
\item \href{#method-PM_pta-clone}{\code{PM_pta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PM_pta} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$new(simdata, targets, target.type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simdata}}{Output of \verb{$sim} method for \link{PM_result} object}

\item{\code{targets}}{See \link{makePTA}.}

\item{\code{target.type}}{See \link{makePTA}.}

\item{\code{...}}{Other arguments passed to \link{makePTA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-save"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_pta object into a .rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$save(file_name = "PMpta.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be created, the default is PMpta.rds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize the \code{PM_pta} object. See \link{summary.PMpta}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMpta}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the \code{PM_pta} object. See \link{plot.PM_pta}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_pta}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-load"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-load}{}}}
\subsection{Method \code{load()}}{
Returns a PM_pta object based on the information found in a specified rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$load(file_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be read, the default is PMpta.rds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
