% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_pta.R
\name{PM_pta}
\alias{PM_pta}
\title{Define object for PTA}
\description{
Define object for PTA

Define object for PTA
}
\details{
This object class contains results of simulations and a probability of
target attainment analysis.

Creating this object first requires a simulation of appropriate regimens
from a model using the \verb{$sim} method for a \link{PM_result} object. That simulation
result is passed to the \link{makePTA} function when a new \code{PM_pta} is created.
}
\author{
Julian Otalvaro and Michael Neely
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{Contains the raw results. See \link{makePTA}.}

\item{\code{outcome}}{Contains summary of raw results. See \link{makePTA}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PM_pta$new()}}
\item \href{#method-save}{\code{PM_pta$save()}}
\item \href{#method-summary}{\code{PM_pta$summary()}}
\item \href{#method-plot}{\code{PM_pta$plot()}}
\item \href{#method-clone}{\code{PM_pta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PM_pta} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$new(simdata, targets, target.type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simdata}}{Output of \verb{$sim} method for \link{PM_result} object}

\item{\code{targets}}{See \link{makePTA}.}

\item{\code{target.type}}{See \link{makePTA}.}

\item{\code{...}}{Other arguments passed to \link{makePTA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_pta object into a .rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$save(file_name = "PMpta.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be created, the default is PMpta.rds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize the \code{PM_pta} object. See \link{summary.PMpta}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMpta}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the \code{PM_pta} object. See \link{plot.PMpta}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PMpta}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
