% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_pta.R
\name{PM_pta}
\alias{PM_pta}
\title{Create Probability of Target Attainment (PTA) object}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This object class contains results of simulations and a probability of
target attainment analysis.
}
\details{
#' There are two ways of creating a \emph{PM_pta} object.
\itemize{
\item \strong{PM_sim$pta()} This way uses the simulation method directly from
a \link{PM_sim} object.
\item \strong{PM_pta$new()} This way takes an external \link{PM_sim} result as an argument
and creates the PTA. It is described here.
}

Both methods require the prior creation of a simulation of
appropriate regimens. They use \link{makePTA} to create the \emph{PM_pta} object.
}
\author{
Julian Otalvaro and Michael Neely
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{Contains the raw results. See \link{makePTA}.}

\item{\code{outcome}}{Contains summary of raw results. See \link{makePTA}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_pta-new}{\code{PM_pta$new()}}
\item \href{#method-PM_pta-save}{\code{PM_pta$save()}}
\item \href{#method-PM_pta-summary}{\code{PM_pta$summary()}}
\item \href{#method-PM_pta-plot}{\code{PM_pta$plot()}}
\item \href{#method-PM_pta-load}{\code{PM_pta$load()}}
\item \href{#method-PM_pta-clone}{\code{PM_pta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PM_pta} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$new(simdata, targets, target.type, success, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simdata}}{One of two possibilities:
\itemize{
\item A \link{PM_sim} object, which typically will contain the results of
multiple simulations OR
\item The quoted filename of a previously saved .rds file
(saved with the \verb{$save} method below) in the current
working directory (or the path plus filename) with results of a prior
PTA analysis.
}}

\item{\code{targets}}{See \link{makePTA}.}

\item{\code{target.type}}{See \link{makePTA}.}

\item{\code{success}}{See \link{makePTA}}

\item{\code{...}}{Other arguments passed to \link{makePTA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-save"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_pta object into a .rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$save(file_name = "PMpta.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be created, the default is PMpta.rds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize the \code{PM_pta} object. See \link{summary.PMpta}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{summary.PMpta}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the \code{PM_pta} object. See \link{plot.PM_pta}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PM_pta}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-load"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-load}{}}}
\subsection{Method \code{load()}}{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Deprecated method to load a prior PTA Replaced by \code{PM_pta$new()} to be
consistent with R6.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$load(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_pta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_pta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_pta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
