% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_report.R
\name{PM_report}
\alias{PM_report}
\title{Generate a report}
\usage{
PM_report(x, template, path, show = TRUE, quiet = TRUE)
}
\arguments{
\item{x}{A \link{PM_result} object obtained from \link{PM_load}.}

\item{template}{Deprecated for app mode. Used only for HTML fallback report
generation when app launch fails.}

\item{path}{Deprecated for app mode. Used only for HTML fallback report
generation when app launch fails.}

\item{show}{Controls whether the Shiny app should be opened automatically, defaults to \code{TRUE}.}

\item{quiet}{Retained for compatibility.}
}
\value{
Launches the Pmetrics reporting app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Launches the Pmetrics reporting app for a completed run
}
\seealso{
\link{PM_load}
}
\author{
Markus Hovd, Julian Otalvaro, and Michael Neely
}
