% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_report.R
\name{PM_report}
\alias{PM_report}
\title{Generate a report}
\usage{
PM_report(x, template, out_path, show = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A \link{PM_result} object obtained from \link{PM_load}.}

\item{template}{If missing, the default Pmetrics report template as specified in \link{getPMoptions}
is used. It can be changed with \link{setPMoptions}. Otherwise, the value for \code{template}
can be "plotly", "ggplot", or "none".}

\item{out_path}{The path for the generated report, defaults to a temporary file
in the current working directory.}

\item{show}{Controls if the report should be automatically opened on generation, defaults to \code{TRUE}}

\item{quiet}{If \code{TRUE}, suppresses the message about report generation, defaults to \code{FALSE}.}
}
\value{
Generates an HTML-report in the current working directory.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Generates a report from a specified Rmd template
}
\seealso{
\link{PM_load}
}
\author{
Markus Hovd, Julian Otalvaro, and Michael Neely
}
