% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_report.R
\name{PM_report}
\alias{PM_report}
\title{Generate a report}
\usage{
PM_report(PM_result, template, outfile, show = TRUE)
}
\arguments{
\item{PM_result}{A \code{PM_result} object obtained from \code{PM_load(x)}, where \code{x} is the run number.}

\item{template}{The path to a template. If not specified, uses the default Pmetrics report template.}

\item{outfile}{The location of the generated report, defaults to a temporary file.}

\item{show}{Controls if the report should be automatically opened on generation, defaults to \code{TRUE}
Pmetrics will prompt the user to set this address the first time the \code{remote} argument is set to \code{TRUE}
in \code{\link{NPrun}}.}
}
\value{
Generates an HTML-report in the current working directory.
}
\description{
Generates a report from a specified Rmd template
}
\seealso{
\code{\link{PM_load}}
}
\author{
Markus Hovd and Julian Otalvaro
}
