% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{PM_result}
\alias{PM_result}
\title{Results of a Pmetrics run}
\description{
Results of a Pmetrics run

Results of a Pmetrics run
}
\details{
This object contains all of the results after a Pmetrics runs. It is created
by using the \link{PM_load} function.

After a run completes, results are stored on your hard drive. They are loaded
back into R with \link{PM_load} to create the \link{PM_result} object, which contains both
the results and functions to analyze or plot the result.

Creation of new \code{PM_result} objects is via \link{PM_load}.

See \link{makeNCA}.

\code{PM_result} objects contain a \code{save} method which invokes \link{saveRDS} to write
the object to the hard drive as an .rds file. This is the corresponding load
function.
}
\examples{

## ------------------------------------------------
## Method `PM_result$load`
## ------------------------------------------------

\dontrun{newRes <- PM_result$load("PMresult.rds")}
}
\author{
Michael Neely, Julian Otalvaro
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{npdata}}{List with all output from NPAG, made by \link{NPparse}}

\item{\code{pop}}{A \link{PM_pop} object}

\item{\code{post}}{A \link{PM_post} object}

\item{\code{final}}{A \link{PM_final} object}

\item{\code{cycle}}{A \link{PM_cycle} object}

\item{\code{op}}{A \link{PM_op} object}

\item{\code{cov}}{A \link{PM_cov} object}

\item{\code{data}}{\link{PM_data} object representing the original .csv data file used in the run}

\item{\code{model}}{text string representing the original model file used in the run}

\item{\code{errfile}}{Name of error file if it exists}

\item{\code{success}}{Boolean if successful run}

\item{\code{valid}}{If the \verb{$validate} method has been executed after a run, this object will be added to
the \code{PM_result} object.  It contains the information required to plot visual predictive checks and normalized prediction
error discrepancies via the npde code developed by Comets et al. Use the
\verb{$save} method on the augmented \code{PM_result} object to save it with the
new validation results.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PM_result$new()}}
\item \href{#method-plot}{\code{PM_result$plot()}}
\item \href{#method-summary}{\code{PM_result$summary()}}
\item \href{#method-auc}{\code{PM_result$auc()}}
\item \href{#method-nca}{\code{PM_result$nca()}}
\item \href{#method-sim}{\code{PM_result$sim()}}
\item \href{#method-save}{\code{PM_result$save()}}
\item \href{#method-validate}{\code{PM_result$validate()}}
\item \href{#method-step}{\code{PM_result$step()}}
\item \href{#method-MM_opt}{\code{PM_result$MM_opt()}}
\item \href{#method-load}{\code{PM_result$load()}}
\item \href{#method-clone}{\code{PM_result$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new object populated with data from previous run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$new(out, quiet = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out}}{The parsed output from \link{PM_load}.}

\item{\code{quiet}}{Quietly validate. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot generic function based on type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$plot(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of plot based on class of object}

\item{\code{...}}{Plot-specific arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summary generic function based on type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$summary(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of summary based on class of object}

\item{\code{...}}{Summary-specific arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auc"></a>}}
\if{latex}{\out{\hypertarget{method-auc}{}}}
\subsection{Method \code{auc()}}{
AUC generic function based on type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$auc(type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of AUC based on class of object}

\item{\code{...}}{Summary-specific arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nca"></a>}}
\if{latex}{\out{\hypertarget{method-nca}{}}}
\subsection{Method \code{nca()}}{
Perform non-compartmental analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$nca(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeNCA}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sim"></a>}}
\if{latex}{\out{\hypertarget{method-sim}{}}}
\subsection{Method \code{sim()}}{
Simulates using the self$final object.
For parameter information refer to \link{SIMrun}. It will return a \code{PM_sim} object
by running \link{SIMparse} at the end of the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$sim(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to \link{SIMrun}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_result object into a .rds file. This is useful if you
have run the \verb{$make_valid} method on a \code{PM_result} object, which returns
an internal simulation based validation as a new \code{valid} field. To save this
validation, use the \verb{$save} method. You could also save the result if you wish
to share it with someone else.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$save(file_name = "PMresult.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be created, the default is PMresult.rds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the result by internal simulation methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$validate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{make_valid}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step"></a>}}
\if{latex}{\out{\hypertarget{method-step}{}}}
\subsection{Method \code{step()}}{
Conduct stepwise linear regression of Bayesian posterior parameter values
and covariates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$step(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{PMstep}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MM_opt"></a>}}
\if{latex}{\out{\hypertarget{method-MM_opt}{}}}
\subsection{Method \code{MM_opt()}}{
Calculate optimal sample times from result and template data file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$MM_opt(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{MM_opt}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
This function loads an rds file created using the \verb{$save} method on a
\code{PM_result} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$load(file_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be read, the default is "PMresult.rds".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{PM_result} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{newRes <- PM_result$load("PMresult.rds")}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_result$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
