% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_sim.R
\name{PM_sim}
\alias{PM_sim}
\title{Object to contain results of simulation}
\usage{
PM_sim$run
}
\arguments{
\item{file_name}{The name of the .rds file to load.}

\item{poppar}{A population parameter result, which is a PM_final object. This
can be found in \code{PM_result$final$data}.}

\item{...}{Additional parameters to be passed to \link{SIMrun} and optionally,
"combine = T" as an argument will be passed to \link{SIMparse}.}
}
\value{
A \link{PM_sim} object

A \code{PM_sim} object created by calling \link{SIMparse} at the completion of the
simulation.
}
\description{
If the \verb{$save} method has previously been invoked on a \link{PM_sim}
object, this function will load those results.

Provides an alternative method to call the simulator directly from output
of a model fitting run.
}
\details{
This object is created after a successful run of the simulator.

There are two methods of creating a PM_sim object.
\itemize{
\item \strong{PM_result$sim}
\item \strong{PM_sim$run}
}

These both call \link{SIMrun} to execute the simulation and \link{SIMparse} to process
the results and return the PM_sim objects. See help on both of these functions
for further details.

See \link{makeAUC}

Choose whether to sim

The saved object is an .rds file. When loaded, it should be assigned to an R
object, e.g. \code{sim1 <- PM_sim$load("filename")}.

Calling this function is equivalent to \code{PM_result$sim()}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obs}}{Observations}

\item{\code{amt}}{Amounts}

\item{\code{parValues}}{Retained simulated parameter values after discarding
any due to truncation limits}

\item{\code{totalSets}}{Number of all simulated parameter values}

\item{\code{totalMeans}}{Mean of all simulated parameter values}

\item{\code{totalCov}}{Covariance of all simulated parameter values}

\item{\code{data}}{A matrix that contains all the above elements as columns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_sim-new}{\code{PM_sim$new()}}
\item \href{#method-PM_sim-save}{\code{PM_sim$save()}}
\item \href{#method-PM_sim-plot}{\code{PM_sim$plot()}}
\item \href{#method-PM_sim-pta}{\code{PM_sim$pta()}}
\item \href{#method-PM_sim-auc}{\code{PM_sim$auc()}}
\item \href{#method-PM_sim-summary}{\code{PM_sim$summary()}}
\item \href{#method-PM_sim-clone}{\code{PM_sim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_sim-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_sim-new}{}}}
\subsection{Method \code{new()}}{
Create new simulation objects with results of \verb{$sim}
method for \link{PM_result}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_sim$new(list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{List of output passed by \verb{$sim}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_sim-save"></a>}}
\if{latex}{\out{\hypertarget{method-PM_sim-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_sim object into a .rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_sim$save(file_name = "PMsim.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be created, the default is PMsim.rds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_sim-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_sim-plot}{}}}
\subsection{Method \code{plot()}}{
Plot \code{PM_sim} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_sim$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{plot.PMsim}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_sim-pta"></a>}}
\if{latex}{\out{\hypertarget{method-PM_sim-pta}{}}}
\subsection{Method \code{pta()}}{
Estimates the Probability of Target Attaintment (PTA), based on the results
of the current Simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_sim$pta(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters, refer to \link{makePTA}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_sim-auc"></a>}}
\if{latex}{\out{\hypertarget{method-PM_sim-auc}{}}}
\subsection{Method \code{auc()}}{
Calculate AUC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_sim$auc(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \link{makeAUC}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_sim-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_sim-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_sim$summary(field, by, individual = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Quoted character value, one of
\itemize{
\item obs for simulated observations
\item amt for simulated amounts in each compartment
\item parValues for simulated parameter values
}}

\item{\code{by}}{Optional quoted column name(s) to group by, e.g. \code{by = "id"} or
\code{by = c("id", "outeq")}.}

\item{\code{individual}}{If "id" is included within \code{by}, the summary statistics
will be calculated for each individual, i.e. each individual's mean, sd, median,
min and max observations. If \code{individual} is \code{FALSE} (the default), then those
summaries will again be summarized to present, for example, the mean of
all subjects' mean observations, or the mean of their maximum observations.
This argument will be ignored if "id" is not in \code{by}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{by} is ommitted, a data frame with rows for each data element except ID,
and columns labeled as mean, sd, median, min and max. If \code{by} is specified,
return will be a list with named elements mean, sd, median, min and max, each containing
the corresponding value for each group in \code{by}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_sim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_sim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_sim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
