% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_sim.R
\name{PM_simlist}
\alias{PM_simlist}
\title{Object to contain list of results of simulation}
\description{
Object to contain list of results of simulation

Object to contain list of results of simulation
}
\details{
This object is created after a successful run of the simulator when
there are multiple subjects in the data template and \code{combine = F} is
used as an argument to \code{PM_sim$run}, which is the default.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{List of all the individual \link{PM_sim} objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PM_simlist-new}{\code{PM_simlist$new()}}
\item \href{#method-PM_simlist-plot}{\code{PM_simlist$plot()}}
\item \href{#method-PM_simlist-pta}{\code{PM_simlist$pta()}}
\item \href{#method-PM_simlist-auc}{\code{PM_simlist$auc()}}
\item \href{#method-PM_simlist-save}{\code{PM_simlist$save()}}
\item \href{#method-PM_simlist-summary}{\code{PM_simlist$summary()}}
\item \href{#method-PM_simlist-clone}{\code{PM_simlist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_simlist-new"></a>}}
\if{latex}{\out{\hypertarget{method-PM_simlist-new}{}}}
\subsection{Method \code{new()}}{
Create new PM_simlist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_simlist$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The list of PM_sim objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_simlist-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PM_simlist-plot}{}}}
\subsection{Method \code{plot()}}{
Plot \code{PM_sim} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_simlist$plot(at = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{Index of the PM_sim object to be plotted.}

\item{\code{...}}{Arguments passed to \link{plot.PM_sim}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_simlist-pta"></a>}}
\if{latex}{\out{\hypertarget{method-PM_simlist-pta}{}}}
\subsection{Method \code{pta()}}{
Estimates the Probability of Target Attaintment (PTA), based on the results
of the specified Simulation.d.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_simlist$pta(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters, refer to \link{makePTA}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_simlist-auc"></a>}}
\if{latex}{\out{\hypertarget{method-PM_simlist-auc}{}}}
\subsection{Method \code{auc()}}{
Calculates the AUC of the specified simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_simlist$auc(at = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{Index of the PM_sim object to be plotted.}

\item{\code{...}}{Arguments passed to \link{makeAUC}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_simlist-save"></a>}}
\if{latex}{\out{\hypertarget{method-PM_simlist-save}{}}}
\subsection{Method \code{save()}}{
Save the current PM_sim object into a .rds file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_simlist$save(file_name = "PMsim.rds")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{Name of the file to be created, the default is PMsim.rds}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_simlist-summary"></a>}}
\if{latex}{\out{\hypertarget{method-PM_simlist-summary}{}}}
\subsection{Method \code{summary()}}{
Summarizes  the specified simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_simlist$summary(at = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{Index of the PM_sim object to be summarized}

\item{\code{...}}{Arguments passed to the \verb{$summary} method for \link{PM_sim} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PM_simlist-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PM_simlist-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PM_simlist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
