% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_tutorial.R
\name{PM_tutorial}
\alias{PM_tutorial}
\title{Pmetrics tutorial}
\usage{
PM_tutorial(wd = getwd(), auto_confirm = FALSE)
}
\arguments{
\item{wd}{The working directory in which to create the Examples folder.}

\item{auto_confirm}{If true, the execution is not stopped to ask the
user for confirmation
Default is current working directory.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Introductory tutorial to Pmetrics
}
\details{
This function will create an Examples folder in the current working
directory. The folder will contain all files and folder necessary to
conduct the tutorial. After the Examples folder is created, open the
\emph{examples.R} script to begin the tutorial.
}
\keyword{internal}
