% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_tutorial.R
\name{PM_tutorial}
\alias{PM_tutorial}
\title{Pmetrics tutorial}
\usage{
PM_tutorial()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Introductory tutorial to Pmetrics
}
\details{
This function will create a \emph{Learn} folder in the current working
directory or one you specify. The folder will contain all files and subfolders necessary to
conduct the tutorial found at https://lapkb.github.io/PM_tutorial/.
After the Learn folder is created, open the
\emph{Rscript/Learn.R} script to begin the tutorial.
}
