% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMbuild.R
\name{PMbuild}
\alias{PMbuild}
\title{Build Pmetrics}
\usage{
PMbuild()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Compile Rust source code used by Pmetrics.
}
\author{
Michael Neely and Julian Otalvaro
}
