% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMbuild.R
\name{PMbuild}
\alias{PMbuild}
\title{Build Pmetrics}
\usage{
PMbuild(skipRegistration = F, autoyes = F)
}
\description{
\code{PMBuild} will ensure all dependent packages are installed and compile
Fortran source code for permanent Pmetrics modules
}
\author{
Michael Neely
}
