% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PMcheck}
\alias{PMcheck}
\title{Check Pmetrics Inputs for Errors}
\usage{
PMcheck(data, model, fix = F, quiet = F)
}
\arguments{
\item{data}{The name of a Pmetrics .csv matrix file in the current working directory,
the full path to one not in the current working directory, or a data.frame containing
the output of a previous \code{\link{PMreadMatrix}} command.}

\item{model}{The filename of a Pmetrics model file in the current working directory.  This parameter is optional.
If specified, and the data object has no errors, the model file will be evaluated.}

\item{fix}{Boolean operator; if \code{TRUE}, Pmetrics will attempt to fix errors in the data file.
Default is \code{FALSE}.}

\item{quiet}{Boolean operator to suppress printed output.  Default is false.}
}
\value{
If \code{fix=TRUE}, then \code{PMcheck} returns a PMmatrix data object which has been
cleaned of errors as much as possible, displaying a report on the console.
If \code{fix=FALSE}, then \code{PMcheck} returns a list of objects of class \emph{PMerr}.  Each object is itself a list whose
first object (\code{$msg}) is a character vector with \dQuote{OK} plus a brief description if there is no error, or the error.
The second object (\code{$results}) is a vector of the row numbers that contain that error.
 \item{colorder}{The first 14 columns must be named id, evid, time, dur, dose, addl, ii, input, out, outeq, c0, c1, c2, and c3 in that order.}
 \item{maxchar}{All id and covariate values should be less than or equal to 11 characters.}
 \item{obsT0}{Time=0 events should not be observations.}
 \item{obsMiss}{Missing observations should be -99 (not \dQuote{.}, which is simply a placeholder).}
 \item{doseComp}{Make sure all dose records are complete, i.e. contain id, time, evid=1 or 4, duration (0 for bolus), dose, input number.}
 \item{obsComp}{Make sure all observation records are complete, i.e. contain id, time, output, and outeq number.}
 \item{evid1}{Ensure that each subject's first record is an evid=1.}
 \item{T0}{Make sure each subject's first time=0.}
 \item{covT0}{Make sure that there is an non-missing entry for each covariate at time=0 for each subject.}
 \item{covNumeric}{Ensure that all covariate entries are numeric.}
 \item{timeOrder}{Ensure that all times within a subject ID are monotonically increasing.}
 \item{contigID}{Ensure that all subject IDs are contiguous.}
 \item{missEVID}{Ensure that all rows have an EVID value.}
 \item{missTIME}{Ensure that all rows have a TIME value.}
 \item{nonNum}{Ensure that all columns which must be numeric are numeric.  These are EVID,
 TIME,DUR,DOSE,ADDL,II,INPUT,OUT,OUTEQ,C0,C1,C2,C3.  Covariate columns are checked separately
 (see above) but also must be numeric.}
}
\description{
This function will check a .csv file or a data frame containing a
previously loaded .csv file (the output of \code{\link{PMreadMatrix}} for errors
which would cause the analysis to fail.  If a model file is provided, and the data
file has no errors, it will also check the model file for errors.
}
\details{
Either a filename or a data object in memory are accepted as \code{data}.
The format of the .csv matrix file is fairly rigid.
It must have the following features.  Text is case-sensitive.
\itemize{
 \item A header in row 1 with the appropriate version, currently \dQuote{POPDATA DEC_11}
	\item Column headers in row 2.  These headers are: #ID, EVID, TIME, DUR, DOSE, ADDL, II, INPUT, OUT, OUTEQ,
C0, C1, C2, C3.
 \item No cell should be empty.  It should either contain a value or \dQuote{.} as a placeholder.
	\item Columns after OUTEQ are interpreted as covariates.
	\item All subject records must begin with a dose event (EVID=1).
	\item All subject records must begin with TIME=0.
 \item All dose events (EVID=1) must have entries in ID, EVID, TIME, DUR, DOSE and INPUT.  ADDL and II are optional, but if ADDL is not 0 or
missing, then II is mandatory.
 \item All observation events (EVID=0) must have entries in ID, EVID, TIME, OUT, OUTEQ.
 If an observation is missing, use \emph{-99}; otherwise use a \dQuote{.} as a placeholder
 in cells that are not required (e.g. INPUT for an observation event).
 \item If covariates are present in the data, there must be an entry for every covariate at time 0 fore each subject.
 \item All covariates must be numeric.
 \item All times within a subject ID must be monotonically increasing.
 \item All subject IDs must be contiguous.
 \item All rows must have EVID and TIME values.
 \item EVID, TIME, DUR, DOSE, ADDL, II, INPUT, OUT, OUTEQ,
C0, C1, C2, C3 must all have numeric entries.
 }

To use this function, see the example below.  As another example, assume that
there is a PMdata object called \dQuote{mdata} loaded in memory.
Also assume that on rows 1, 10, 150, there are incomplete dose records.

Use \code{err <- PMcheck(mdata)} to run the check.
To see the rows in mdata with problems, read the report on the console.
In this example, you would use \code{mdata[err$doseComp$results,]}.

You could then try to fix the problem(s) with \code{mdata2 <- PMcheck(mdata,fix=T)}.  Note that we are now returning
a PMmatrix data object (hopefully cleaned of errors) rather than the PMerr object returned when \code{fix=FALSE}.
Pmetrics handles each of the errors in the following ways.
\itemize{
 \item If the columns are simply out of order, they will be reordered.  If some are missing, the fix must
 be done by the user, i.e. manually.
 \item All id and covariate values are truncated to 11 characters.
 \item Time=0 observations are deleted.
 \item Missing observations are set to -99 (not \dQuote{.}).
 \item Incomplete dose records are flagged for the user to fix manually.
 \item Incomplete observation records are flagged for the user to fix manually.
 \item Subjects without an EVID=1 as first event are flagged for the user to fix manually.
 \item Subjects with TIME != 0 as first event have dummy dose=0 events inserted at time 0.
 \item Subjects with a missing covariate at time 0 are flagged for the user to fix manually.
 \item Non-numeric covariates are converted to numeric (via \link{factor}).
 \item Non-ordered times are sorted within a subject if there are no EVID=4 events; otherwise the
 user must fix manually.
 \item Non-contiguous subject ID rows are combined and sorted if there are no EVID=4 events; otherwise the
 user must fix manually.
 \item Rows missing an EVID are assigned a value of 0 if DOSE is  missing, 1 otherwise.
 \item Rows missing a TIME value are flagged for the user to fix manually.
 \item Columns that are non-numeric which must be numeric are flagged for the user to fix manually.
 These are EVID, TIME, DUR, DOSE, ADDL, II, INPUT, OUT, OUTEQ, C0, C1, C2, and C3.
 Covariate columns are fixed separately (see above).
}
}
\examples{
data(PMex3)
err <- PMcheck(badData)
badData[err$obsLast$results,]
badData[err$covT0$results,]
goodData <- PMcheck(badData,fix=T)
PMcheck(goodData)
#you have to fix manually problems which require data entry
}
\author{
Michael Neely
}
\seealso{
\code{\link{PMwriteMatrix}}, \code{\link{PMreadMatrix}}
}

