% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PMcheckMatrix}
\alias{ITload,}
\alias{ITreport,}
\alias{NPload,}
\alias{NPreport,}
\alias{PMcheckMatrix}
\alias{PMcheckMatrix,}
\alias{PMdiag}
\alias{PMfixMatrix,}
\title{Deprecated functions.}
\usage{
PMcheckMatrix(...)
}
\description{
The following functions are deprecated in Pmetrics.
}
\author{
Michael Neely
}

