% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMcompare}
\alias{PMcompare}
\title{Compare NPAG or IT2B runs}
\usage{
PMcompare(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Compare NPAG or IT2B runs, based on objects loaded by \link{PMload}, which
is now deprecated, because it loaded Pmetrics objects from a prior run into
user's global environment, which was
unsafe programming practice and caused warnings in CRAN.
}
\seealso{
\link{PM_load}
}
\author{
Michael Neely
}
\keyword{internal}
