% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMFortranConfig.R
\name{PMFortranConfig}
\alias{PMFortranConfig}
\title{Read or define the Fortran compiler and command line template}
\usage{
PMFortranConfig(reconfig = F)
}
\arguments{
\item{reconfig}{Default is \code{FALSE}.  If \code{TRUE}, will allow user to change
the previously specified compiler and template.}
}
\value{
\code{PMFortranConfig} returns the compile command template specific to the chosen
compiler.
}
\description{
Fortran compiler commands
}
\details{
This function will read or define the installed Fortran compiler and generate
a command line template appropriate to the compiler.

Templates are defined for the following compilers: \strong{gfortran}, \strong{g95},
\strong{Intel Visual}, and \strong{Lahey}.  Additionally, users may specify a custom command line
template for any other compiler.  Within the template \emph{<exec>} is used as a placeholder
for the filename of the executable file, and \emph{<files>} as a placeholder for the files to
compile and link, both of which will be defined at run time by the appropriate Pmetrics functions.
The Pmetrics functions which use a Fortran compiler are \link{NPrun}, \link{ITrun},
\link{ERRrun}, and \link{SIMrun}.
}
\seealso{
\link{NPrun}, \link{ITrun},[ERRrun], and \link{SIMrun}
}
\author{
Michael Neely
}
