% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMFortranConfig}
\alias{PMFortranConfig}
\alias{PMFortranConfig,}
\alias{PMcheckMatrix,}
\alias{PMfixMatrix,}
\alias{NPload,}
\alias{ITload,}
\alias{NPreport,}
\alias{ITreport,}
\alias{PMdiag,}
\alias{PMreadDefaults,}
\alias{PMwriteDefaults,}
\alias{makeNPDE}
\title{Deprecated functions.}
\usage{
PMFortranConfig(...)
}
\description{
The following functions are deprecated in Pmetrics.
}
\author{
Michael Neely
}
