% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMgetCRCL}
\alias{PMgetCRCL}
\title{Add CRCL}
\usage{
PMgetCRCL(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use add_renal instead.
}
\author{
Michael Neely
}
\keyword{internal}
