% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMload.R
\name{PMload}
\alias{PMload}
\title{Load Pmetrics NPAG or IT2B output}
\usage{
PMload(run = 1, ..., remote = F, server_address)
}
\arguments{
\item{run}{The numerical value of the folder number containing the run results.  This
number will also be used to name objects uniquely by appending \dQuote{.\code{run}}, 
e.g. NPdata.1 or ITdata.1 if run=1. This parameter is \code{1} by default.}

\item{\dots}{Additional runs to load if desired.}

\item{remote}{Default is \code{FALSE}.  Set to \code{TRUE} if loading results of an NPAG run on remote server.
See \code{\link{NPrun}}. Currently remote runs are not configured for IT2B or the Simulator.}

\item{server_address}{If missing, will use the default server address returned by getPMoptions(). 
Pmetrics will prompt the user to set this address the first time the \code{remote} argument is set to \code{TRUE}
in \code{\link{NPrun}}.}
}
\value{
The following objects are loaded into R.
\item{NPdata/ITdata }{List with all output from NPAG/IT2B}
\item{pop }{ NPAG only: Population predictions for each output equation}
\item{post }{ NPAG only: Individual posterior predictions for each output equation}
\item{final }{Final cycle population support points and parameter summary statistics}
\item{cycle }{Cycle log-likelihood, AIC, BIC, Gamma/lambda, and normalized parameter means, medians and SDs}
\item{op }{List of observed vs. population and posterior predicted plots for each output equation}
\item{cov }{Data frame of subject ID, covariate values, and Bayesian posterior parameter estimates}
\item{mdata }{The original .csv data file used in the run}
\item{valid }{If \code{\link{makeValid}} has been executed after a run, this object will be added to 
the save data.  It contains the information required to plot visual predictive checks and normalized prediction
error discrepancies via the npde code developed by Comets et al}
}
\description{
Loads all the data from an \emph{NPAG} or \emph{IT2B} run
}
\seealso{
\code{\link{PMreport}}, \code{\link{NPparse}}, \code{\link{ITparse}}, 
\code{\link{makeFinal}}, \code{\link{makeCycle}}, \code{\link{makeOP}}, \code{\link{makeCov}}, 
\code{\link{makePop}}, \code{\link{makePost}}
}
\author{
Michael Neely
}
