% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMload}
\alias{PMload}
\title{Load Pmetrics NPAG or IT2B output}
\usage{
PMload(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Loaded Pmetrics objects from prior run into user's global environment, which was
unsafe programming practice and causes warnings in CRAN.
}
\seealso{
\link{PM_result}
}
\author{
Michael Neely
}
\keyword{internal}
