% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMremote.R
\name{PMlogin}
\alias{PMlogin}
\title{Login a user in the remote server}
\usage{
PMlogin(email, server_address)
}
\arguments{
\item{email}{User's email.}

\item{server_address}{remote server address. If omitted,  Pmetrics will use the the default one. see  \code{setPMoptions}}
}
\value{
A message that specifies if the user was successfuly logged in, or an explanation of the error.
}
\description{
Login a existent user onto the server.
}
\details{
This functions receives an email of a registered user and optionally the server adress.  
If no server_address is provided Pmetrics will use the the default one. see  \code{setPMoptions}.
The function will prompt the user for the password.
The server session will last until the R session finishes or by using \code{PMlogout}.
}
\author{
Michael Neely
}
