% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMnews.R
\name{PMnews}
\alias{PMnews}
\title{Pmetrics changelog}
\usage{
PMnews(PMversion = packageVersion("Pmetrics"))
}
\arguments{
\item{PMversion}{Default is the current version, otherwise a character string with the starting version you wish to see up to
the current, e.g. \dQuote{0.21}.  Use \dQuote{all} for all versions.}
}
\value{
The changelog for the requested version.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

See changelog for Pmetrics
}
\examples{
\dontrun{
PMnews()
}
}
\author{
Michael Neely
}
\keyword{internal}
