% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PMnews}
\alias{PMnews}
\title{Pmetrics changelog}
\usage{
PMnews(version = packageVersion("Pmetrics"))
}
\arguments{
\item{version}{Default is the current version, otherwise a character string with the starting version you wish to see up to
the current, e.g. \dQuote{0.21}.  Use \dQuote{all} for all versions.}
}
\value{
The changelog for the requested version.
}
\description{
See changelog for Pmetrics
}
\examples{
PMnews()
}
\author{
Michael Neely
}

