% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PMpatch}
\alias{PMpatch}
\title{Download and install Pmetrics patches}
\usage{
PMpatch()
}
\value{
A Pmetrics patch which will be installed via \code{source}
}
\description{
Download and install Pmetrics patches from LAPK website
}
\author{
Michael Neely
}

