% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_data.R
\name{PMreadMatrix}
\alias{PMreadMatrix}
\title{Read a Pmetrics data file}
\usage{
PMreadMatrix(
  file,
  sep = getPMoptions("sep"),
  dec = getPMoptions("dec"),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{file}{The name of the file to be loaded, including the full path if not
in the current working directory (check with \link{getwd}).}

\item{sep}{Delimiter between columns, which is a comma by default, but can be changed with
\link{setPMoptions}.}

\item{dec}{Decimal separator, which is a period by default, but can be changed with
\link{setPMoptions}.}

\item{quiet}{Default is \code{FALSE}.  If \code{TRUE}, there will be no report to
the console on the contents of file.}

\item{\dots}{Other parameters to be passed to \code{\link[readr:read_delim]{readr::read_delim()}}}
}
\value{
\code{PMreadMatrix} returns a data frame of class "PMmatrix".
If the file is successfully read and \code{quiet=F},
the column headers will be reported to the console as a validation check.
Note that this function converts the column headers in the
\code{file} from upper to lowercase for convenient referencing in R.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Reads a Pmetrics .csv matrix input file into R.
This function is largely superseded as the function is called automatically
when data are intialized as a \link{PM_data} object with \code{PM_data$new()}. There
is rarely need to call \code{PMreadMatrix()} directly any longer.
}
\details{
As of Pmetrics version 2, the structure of a valid .csv file relaxed.
Minimal required columns are id, time, dose, and out. This function is now included
as part of the \link{PM_data} R6 object to create new \code{PM_data} objects. Users should
rarely have a need to call \code{PMreadMatrix} as a standalone function unless they
continue to use Pmetrics in its legacy mode (versions < 2.0). Note that support
for legacy Pmetrics will eventually wither as the package evolves.

There are a number of other options for columns in the data input.  Details can
be found in the \href{https://lapkb.github.io/Pmetrics/articles/data.html}{documentation}.
}
\seealso{
\link{PMwriteMatrix}, \link{PMcheck}, and \link{plot.PM_data}
}
\author{
Michael Neely
}
