% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMreport}
\alias{PMreport}
\title{Pmetrics run report}
\usage{
PMreport(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use PM_report() instead.
}
\author{
Michael Neely
}
\keyword{internal}
