% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{PMsave}
\alias{PMsave}
\title{Save Pmetrics objects}
\usage{
PMsave(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

No longer applies as Pmetrics R6 objects have their own save methods.
}
\seealso{
\link{PM_result}, \link{PM_sim}, \link{PM_pta}
}
\author{
Michael Neely
}
\keyword{internal}
