% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PMwriteDefaults}
\alias{PMwriteDefaults}
\title{Save Pmetrics Session Defaults}
\usage{
PMwriteDefaults()
}
\value{
None
}
\description{
Write session defaults to disk so that they are persistent.
}
\details{
The default behavior of any Pmetrics function can be set by the user for a session with \code{\link{setDefaults}}.  However, these
custom defaults will not be saved.  Use this function to write them to the Pmetrics library for session-to-session permanence.
}
\author{
Michael Neely
}
\seealso{
\code{\link{setDefaults}}, \code{\link{getDefaults}}, \code{\link{unsetDefaults}}
}

