% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMwrk2csv.R
\name{PMwrk2csv}
\alias{PMwrk2csv}
\title{Convert Old .wrk Files to .csv Matrix File}
\usage{
PMwrk2csv(prefix, ext = NULL, nsub)
}
\arguments{
\item{prefix}{The alphabetic prefix of the working copy files to be converted,
as a character vector.}

\item{ext}{The extension of the working copy files files, if it exists.
Does not have to be specified.}

\item{nsub}{The number of subjects, or working copy files to read.}
}
\value{
A new file will be created with the name equal to \code{prefix} and
an extension of "csv".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Convert old-style, USC*PACK single drug working copy files into
a \link{PM_data} object and write a .csv file to the current working directory.
}
\details{
This function will determine if the working copy files are old and convert them.
New, multi-drug working copy files will be ignored. IDs will be suffixed with
.1 to .9 for <10 subjects, .01 to .99 for <100 subjects and .001 to .999 for <1000 subjects,
as needed to ensure unique ID numbers.
}
\author{
Michael Neely
}
