% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMMopt.R
\name{print.MMopt}
\alias{print.MMopt}
\title{Print Pmetrics Multiple-Model Optimal Sampling Objects}
\usage{
\method{print}{MMopt}(x, ...)
}
\arguments{
\item{x}{The name of an \emph{MMopt} data object generated by \link{MM_opt}}

\item{...}{Not used}
}
\value{
Prints the optimal sampling times and Bayes Risk.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print \emph{MMopt} objects
}
\seealso{
\link{MM_opt}
}
\author{
Michael Neely
}
