% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.MMopt}
\alias{print.MMopt}
\title{Print Pmetrics Multiple-Model Optimal Sampling Objects}
\usage{
\method{print}{MMopt}(x)
}
\arguments{
\item{x}{The name of an \emph{MMopt} data object generated by \code{\link{MMopt}}}
}
\value{
Prints the optimal sampling times and Bayes Risk.
}
\description{
Print \emph{MMopt} objects
}
\details{
Simulated observations are plotted on the y-axis vs. time on the x.axis.
Optimal sampling times are indicated as vertical lines.
}
\author{
Michael Neely
}
\seealso{
\code{\link{MMopt}}
}

