% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPMdopt.R
\name{print.PMdopt}
\alias{print.PMdopt}
\title{Print PMdopt}
\usage{
\method{print}{PMdopt}(x, ...)
}
\arguments{
\item{x}{A PMdopt object made by \link{Dopt}.}

\item{...}{Not used.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print a Pmetrics PMdopt Object made by \link{Dopt}.
}
\seealso{
\link{Dopt}
}
\author{
Michael Neely
}
