% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPMerr.R
\name{print.PMerr}
\alias{print.PMerr}
\title{Print Data Errors}
\usage{
\method{print}{PMerr}(x, ...)
}
\arguments{
\item{x}{A PMerr object made by validation in \link{PM_data}.}

\item{...}{Not used.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print the errors in a Pmetrics data file or \link{PM_data} object.
}
\seealso{
\link{PM_data}
}
\author{
Michael Neely
}
