% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPMnpc.R
\name{print.PMnpc}
\alias{print.PMnpc}
\title{Print NPC}
\usage{
\method{print}{PMnpc}(x, ...)
}
\arguments{
\item{x}{A PMnpc object made by \link{plot.PM_sim}.}

\item{\dots}{Other parameters which are not necessary.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print a Pmetrics NPC (Numerical Predictive Check) object, made by \link{plot.PM_sim}.
}
\seealso{
\link{plot.PM_sim}
}
\author{
Michael Neely
}
