% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSummaryPMfinal.R
\name{print.summary.PMfinal}
\alias{print.summary.PMfinal}
\title{Print Summary of Parameter Values and Credibility Intervals}
\usage{
\method{print}{summary.PMfinal}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A summary.PMfinal object made by \link{summary.PM_final}.}

\item{digits}{Integer, used for number of digits to print.}

\item{...}{Not used.}
}
\value{
A printed object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print a Pmetrics Final Summary Object
}
\details{
Print a summary of parameter medians and MAWD, with point estimates and credibilty intervals
from an object made by \link{summary.PM_final}.
}
\seealso{
\link{summary.PM_final}
}
\author{
Michael Neely
}
