% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSummaryPMmatrix.R
\name{print.summary.PMmatrix}
\alias{print.summary.PMmatrix}
\title{Summarize Covariates and Bayesian Posterior Parameter Values}
\usage{
\method{print}{summary.PMmatrix}(x)
}
\arguments{
\item{x}{A summary.PMmatrix object made by \code{\link{summary.PMmatrix}}}
}
\value{
A formatted printing of a \emph{summary.PMmatrix} object
}
\description{
Print the summary of a Pmetrics PMmatrix object
}
\details{
Summarize the raw data used for a Pmetrics run.
}
\seealso{
\code{\link{summary.PMmatrix}}
}
\author{
Michael Neely
}
