% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSummaryPMop.R
\name{print.summary.PMop}
\alias{print.summary.PMop}
\title{Print Summary of Observations and Predictions}
\usage{
\method{print}{summary.PMop}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{x}{A summary.PMop object made by \code{\link{summary.PMop}}.}

\item{digits}{Integer, used for number of digits to print.}

\item{\dots}{Other parameters which are not necessary.}
}
\value{
A printed object.
}
\description{
Print a Pmetrics Observed vs. Predicted Summary Object
}
\details{
Print a summary of observations, predictions and errors in a summary.PMop object made by \code{\link{summary.PMop}}.
}
\seealso{
\code{\link{summary.PMop}}
}
\author{
Michael Neely
}
