% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_model.R
\name{proportional}
\alias{proportional}
\title{Proportional error model}
\usage{
proportional(prop, constant = FALSE)
}
\arguments{
\item{prop}{Initial value for gamma}

\item{constant}{Estimate if \code{FALSE} (default).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Create an proportional (gamma) error model
}
