% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMoptions.R
\name{setPMoptions}
\alias{setPMoptions}
\title{Set Pmetrics User Options}
\usage{
setPMoptions(launch.app = TRUE)
}
\arguments{
\item{launch.app}{Launch the app to set options. Default \code{TRUE}.}
}
\value{
The user preferences file will be updated.  This will persist from session to session
and if stored in the external location, through Pmetrics versions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Set user options for Pmetrics
}
\details{
When you call this function with the default \code{launch.app = TRUE}, it will start
a Shiny app to set options for the Pmetrics package.
Also, when the Pmetrics package is first loaded with \code{library(Pmetrics)},
this function will be called with \code{launch.app = TRUE} to read saved options from
a \emph{PMoptions.json} file stored in a folder outside
of the Pmetrics package, so that your options will persist when Pmetrics is updated.
}
\author{
Michael Neely
}
