% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPMoptions.R
\name{setPMoptions}
\alias{setPMoptions}
\title{Set Pmetrics User Options}
\usage{
setPMoptions(sep, dec)
}
\arguments{
\item{sep}{The field separator character; \dQuote{,} by default, but could be \dQuote{;}}

\item{dec}{The decimal separator character; \dQuote{.} by default, but could be \dQuote{,}}
}
\value{
The user preferences file will be updated.  This will persist from session to session.
}
\description{
Set user options for Pmetrics
}
\details{
This function will set user options for Pmetrics.
}
\author{
Michael Neely
}
