% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_sim.R
\name{SIMparse}
\alias{SIMparse}
\title{Parse Pmetrics Simulator Output}
\usage{
SIMparse(file, include, exclude, combine = F, quiet = F)
}
\arguments{
\item{file}{An output file or files of the simulator in the current working directory, or the full
pathname to the file.  To load and combine multiple outputs, specify files separated by commas
or using wild cards.  See details.}

\item{include}{A vector of files to include in the parsing.  For example, if you used a wild
card in the \code{file} argument, such as \dQuote{simout?.txt}, which returned four files:
simout1.txt, simout2.txt, simout3.txt and simout4.txt, and you wished to only parse the first
and fourth file, specify \code{include=c(1,4)}.}

\item{exclude}{See the discussion for \code{include}, but this will exclude specified files.}

\item{combine}{Boolean parameter, default \code{False}, which specifies whether you wish to combine
the parsed files into a single PMsim object.  This can be useful for making visual predictive
checks, for example.  If \code{combine=F}, and multiple files are parsed, then the return object
will be a list of PMsim objects, which can be plotted or otherwise accessed using standard list
referencing, e.g. \code{simlist[[1]]}, \code{simlist[[2]]}, etc.}

\item{quiet}{Suppress messages}
}
\value{
If one file is parsed or multiple files are parsed and combined, the return will be a list with five items, of class \emph{PMsim}.
If multiple files are parsed and not combined, then the return will be a \emph{PM_simlist} of \emph{PMsim} objects.
\item{obs }{An data frame of simulated observations with 4 columns: id, time, out, outeq.
\emph{id} is the number of the simulated subject, which will have a unique ending appended
if simulations are combined, such that \emph{id} will become x.y with x being the simulated profile
number and y being the simulation template number.  \emph{time} is the time of the simulated
output, \emph{out} of output equation number \emph{outeq}.}
\item{amt }{An data frame of simulated amounts with 4 columns: id, time, out, comp.
\emph{id} is the number of the simulated subject, which will have a unique ending appended
if simulations are combined, such that \emph{id} will become x.y with x being the simulated profile
number and y being the simulation template number.  \emph{time} is the time of the simulated
amount, \emph{out} in compartment  number \emph{comp}.}
\item{parValues }{A datframe of the simulated parameter values, combined across files as necessary}
\item{totalSets}{The total number of parmeter sets simulated, which may be greater than the number of rows in \code{parValues} if some sets
were discarded for being outside specified limits.  For more than one file parsed, this will the total number in all files.}
\item{totalMeans}{The means of each simulated parameter based on all profiles in a given file (even those discarded for exceeding limits).
For more than one file parsed, this will be the weighted averages for all simulations.}
\item{totalCov}{The covariances of the simulated parameter sets based on all profiles in a given file (even those discarded for exceeding limits).
For more than one file parsed, this will be the weighted averages for all simulations.}
A plot method exists in \code{\link{plot.PMsim}} for \emph{PMsim} objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Parses the output of the Pmetrics simulator.
This is largely superseded as it is automatically called by either of the simulation
methods: \code{PM_result$sim()} or \code{PM_sim$new()}. There is rarely a need any longer to call this
function directly.
}
\details{
For \code{file} specification \dQuote{?} will be matched by just a single numeral or character; \dQuote{*} will be
matched by any number of consecutive alphanumeric characters.  Examples include \code{file='simout1.txt,simout2.txt,simout3.txt'},
\code{file='simout?.txt'} and \code{file='sim*.txt'}.All three will find the files simout1.txt,
simout2.txt, and simout3.txt in the working directory. The second example would also find simout4.txt, etc.  The third
example would also find sim_1.txt if that existed.
Note that to combine simulator output files, the numbers of simulated profiles may differ.
The number of outputs and times of observations also may differ, although combining these may lead to
strange plots since not all profiles have the same observations.
For parallel execution, the function requires packages 'doParallel' and 'foreach'. If not installed, it will try to install, failing that it will run in serial mode.
}
\seealso{
\link{PM_result}, \link{PM_sim}, \link{SIMrun}
}
\author{
Michael Neely
}
