% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simulate_all}
\alias{simulate_all}
\title{Simulates all subjects in the data set using the model DSL source.}
\usage{
simulate_all(data_path, model_source, theta, kind, solver = NULL)
}
\arguments{
\item{data_path}{Path to the data file.}

\item{model_source}{DSL source code for the model.}

\item{theta}{Data frame of support points.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}

\item{solver}{Optional ODE solver name.}
}
\value{
Simulation results.
}
\description{
Simulates all subjects in the data set using the model DSL source.
}
\keyword{internal}
