% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simulate_all}
\alias{simulate_all}
\title{Simulates all subjects in the data set using the model at the given path.}
\usage{
simulate_all(data_path, model_path, spp)
}
\description{
Simulates all subjects in the data set using the model at the given path.
}
