% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simulate_one}
\alias{simulate_one}
\title{Simulates the first subject in the data set using the given model.}
\usage{
simulate_one(data_path, model_source, spp, solver = NULL)
}
\arguments{
\item{data_path}{Path to the data file.}

\item{model_source}{Model definition written in the pharmsol DSL.}

\item{spp}{One support point as a numeric vector.}

\item{solver}{Optional ODE solver name.}
}
\value{
Simulation results.
}
\description{
Simulates the first subject in the data set using the given model.
}
\keyword{internal}
