% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simulate_one}
\alias{simulate_one}
\title{Simulates the first subject in the data set using the model DSL source.}
\usage{
simulate_one(data_path, model_source, spp, kind, solver = NULL)
}
\arguments{
\item{data_path}{Path to the data file.}

\item{model_source}{DSL source code for the model.}

\item{spp}{One support point as a numeric vector with probabiltity.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}

\item{solver}{Optional ODE solver name.}
}
\value{
Simulation results.
}
\description{
Simulates the first subject in the data set using the model DSL source.
}
\keyword{internal}
