% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{simulate_one}
\alias{simulate_one}
\title{Simulates the first subject in the data set using the model at the given path.}
\usage{
simulate_one(data_path, model_path, spp, kind)
}
\description{
Simulates the first subject in the data set using the model at the given path.
}
