% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/rextendr_doc.R
\name{simulate_one}
\alias{simulate_one}
\title{Simulates the first subject in the data set using the model at the given path.}
\usage{
simulate_one(data_path, model_path, spp, kind)

simulate_one(data_path, model_path, spp, kind)
}
\arguments{
\item{data_path}{Path to the data file.}

\item{model_path}{Path to the compiled model file.}

\item{spp}{Which support point to simulate.}

\item{kind}{Kind of model, which can either be "ODE" or "Analytical".}
}
\value{
Simulation results.
}
\description{
Simulates the first subject in the data set using the model at the given path.

Simulates one subject in the data set using the model at the given path.
}
