% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_cov.R
\name{summary.PM_cov}
\alias{summary.PM_cov}
\title{Summarize Covariates and Bayesian Posterior Parameter Values}
\usage{
\method{summary}{PM_cov}(object, icen = "median", ...)
}
\arguments{
\item{object}{A PM_cov object}

\item{icen}{Summary function for covariates with time dependent values and posterior parameters.
Default is "median", but can specify "mean".}

\item{...}{Not used.}
}
\value{
A data frame with the summary of the PM_cov object for each subject's covariates and
Bayesian posterior parameter values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize a Pmetrics Covariate object
}
\details{
This is a function usually called by the \verb{$summary()} method for \link{PM_cov} objects
with a \link{PM_result} to summarize covariates and Bayesian posterior parameter
values for each subject. The function can
be called directly on a \link{PM_cov} object. See examples.Summarize .
}
\examples{
\dontrun{
NPex$cov$summary() # preferred
summary(NPex$cov) # alternative
NPex$cov$summary(icen = "mean") # use mean as summary
}
}
\seealso{
\link{PM_cov}
}
\author{
Michael Neely
}
