% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_data.R
\name{summary.PM_data}
\alias{summary.PM_data}
\title{Summarize PM_data objects}
\usage{
\method{summary}{PM_data}(object, formula, FUN, include, exclude, ...)
}
\arguments{
\item{object}{A \link{PM_data} object.}

\item{formula}{Optional formula for specifying custom summaries.  See \link{aggregate}
and \link{formula} for details on how to specify formulae in R. If, for example, the data contain
a covariate for weight named 'wt', then to summarize the mean dose in mg/kg per subject specify
\code{formula = dose/wt ~ id} and  \code{FUN = mean}.}

\item{FUN}{The summary function to apply to \link{formula}, if specified. This is not
quoted, and usual choices will be \link{mean}, \link{median}, \link{max}, or \link{min}.}

\item{include}{A vector of subject IDs to include in the summary, e.g. \code{c(1:3,5,15)}}

\item{exclude}{A vector of subject IDs to exclude in the summary, e.g. \code{c(4,6:14,16:20)}}

\item{...}{Additional arguments to \code{FUN}, e.g. \code{na.rm = TRUE}}
}
\value{
A list of class \emph{summary.PM_data} with the following items:
\itemize{
\item \strong{nsub} Number of subjects
\item \strong{ndrug} Number of drug inputs
\item \strong{numeqt} Number of outputs
\item \strong{nobsXouteq} Number of observations by outeq
\item \strong{missObsXouteq} Number of missing observations by outeq
\item \strong{loqObsXouteq} Number of observations coded as below the limit of quantification by outeq
\item \strong{ncov} Number of covariates
\item \strong{covnames} Covariate names
\item \strong{ndoseXid} Number of doses per input per subject
\item \strong{nobsXid} Number of observations per outeq per subject
\item \strong{doseXid} Doses per input per subject
\item \strong{obsXid} Observations per outeq per subject
\item \strong{formula} Results of including \link{formula}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize the raw data used for a Pmetrics run.
}
\seealso{
\link{aggregate}
}
\author{
Michael Neely
}
