% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_result.R
\name{summary.PM_final}
\alias{summary.PM_final}
\title{Wrapper function for summmary.PMfinal}
\usage{
\method{summary}{PM_final}(obj, ...)
}
\arguments{
\item{obj}{The \emph{PM_final} object to summarize}

\item{...}{Arguments passed to \link{summary.PMfinal}}
}
\value{
A \link{summary.PMfinal} object
}
\description{
This redirects to summary.PMfinal for PM_final R6 objects
}
\details{
See \link{summary.PMfinal}. Alternative way to summarize is
\code{PM_result$final$summary()}.
}
