% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_pta.R
\name{summary.PM_pta}
\alias{summary.PM_pta}
\title{Summarize PM_pta}
\usage{
\method{summary}{PM_pta}(object, ...)
}
\arguments{
\item{object}{The \emph{PM_pta} object to summarize}

\item{...}{Arguments passed to \link{summary.PMpta}}
}
\value{
A \link{summary.PMpta} object
}
\description{
`r lifecycle::badge("stable")

Wrapper function for summmary.PMpta
}
\details{
This redirects to \link{summary.PMpta} for PM_pta R6 objects
The R6 method to summarize is \code{PM_pta$summary()}.
}
