% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_pta.R
\name{summary.PM_pta}
\alias{summary.PM_pta}
\title{Wrapper function for summmary.PMpta}
\usage{
\method{summary}{PM_pta}(obj, ...)
}
\arguments{
\item{obj}{The \emph{PM_pta} object to summarize}

\item{...}{Arguments passed to \link{summary.PMpta}}
}
\value{
A \link{summary.PMpta} object
}
\description{
This redirects to summary.PMpta for PM_pta R6 objects
}
\details{
See \link{summary.PMpta}. Alternative way to summarize is
\code{PM_pta$summary()}.
}
