% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPMcov.R
\name{summary.PMcov}
\alias{summary.PMcov}
\title{Summarize Covariates and Bayesian Posterior Parameter Values}
\usage{
\method{summary}{PMcov}(object, icen = "median", ...)
}
\arguments{
\item{object}{A PMcov object made by \link{makeCov}.}

\item{icen}{Summary function for covariates and posterior parameters. Default is "median", but can specify "mean".}

\item{...}{Not used.}
}
\value{
A data frame with the summary of the PMcov object for each subject's covariates and
Bayesian posterior parameter values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize a Pmetrics Covariate object
}
\details{
Summarize covariates and Bayesian posterior parameter values for each subject.
}
\seealso{
\link{makeCov}, \link{PM_cov}
}
\author{
Michael Neely
}
