% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPMcov.R
\name{summary.PMcov}
\alias{summary.PMcov}
\title{Summarize Covariates and Bayesian Posterior Parameter Values}
\usage{
\method{summary}{PMcov}(object, ..., icen = "median")
}
\arguments{
\item{object}{A PMcov object made by \code{\link{makeCov}}.}

\item{\dots}{Other parameters which are not necessary.}

\item{icen}{Summary function for covariates and posterior parameters. Default is \dQuote{median}, but can specify \dQuote{mean}.}
}
\value{
A data frame with the summary of the PMcov object for each subject's covariates and 
Bayesian posterior parameter values.
}
\description{
Summarize a Pmetrics Covariate object
}
\details{
Summarize covariates and Bayesian posterior parameter values for each subject.
}
\seealso{
\code{\link{makeCov}}
}
\author{
Michael Neely
}
