% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPMdopt.R
\name{summary.PMdopt}
\alias{summary.PMdopt}
\title{Summarize PMdopt objects}
\usage{
\method{summary}{PMdopt}(x)
}
\arguments{
\item{x}{A PMdopt object made by \code{\link{Dopt}}.}
}
\value{
The weighted mean D-optimal times.
}
\description{
Summarize a Pmetrics D-optimal object
}
\details{
Summarize observations, predictions and errors in a PMdopt object made by \code{\link{Dopt}}.
}
\seealso{
\code{\link{makeOP}}
}
\author{
Michael Neely
}
