% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.PMmatrix}
\alias{summary.PMmatrix}
\title{Summarize PMmatrix objects}
\usage{
\method{summary}{PMmatrix}(x, formula, FUN, ..., include, exclude)
}
\arguments{
\item{x}{A PMmatrix object loaded by \code{\link{PMreadMatrix}} or \code{\link{PMload}}.}

\item{formula}{Optional formula for specifying custom summaries.  See \code{\link{aggregate}}
and \code{\link{formula}} for details on how to specify formulae in R.}

\item{FUN}{The summary function to apply to \code{formula}, if specified.}

\item{include}{A vector of subject IDs to include in the summary, e.g. c(1:3,5,15)}

\item{exclude}{A vector of subject IDs to exclude in the summary, e.g. c(4,6:14,16:20)}

\item{\dots}{Additional arguments to \code{FUN}, e.g. \code{na.rm=T}}
}
\value{
A list of class \emph{summary.PMmatrix} with the summary of the PMmatrix object,
containing the following items:
\item{nsub}{Number of subjects}
\item{ndrug}{Number of drug inputs}
\item{numeqt}{Number of outputs}
\item{nobsXouteq}{Number of observations by outeq}
\item{missObsXouteq}{Number of missing observations by outeq}
\item{ncov}{Number of covariates}
\item{covnames}{Covariate names}
\item{ndoseXid}{Number of doses per input per subject}
\item{nobsXid}{Number of observations per outeq per subject}
\item{doseXid}{Doses per input per subject}
\item{obsXid}{Observations per outeq per subject}
\item{formula}{Results of including \code{formula}}
}
\description{
Summarize a Pmetrics PMmatrix object
}
\details{
Summarize the raw data used for a Pmetrics run.
}
\author{
Michael Neely
}
\seealso{
\code{\link{print.summary.PMmatrix}}, \code{\link{aggregate}}
}

