% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPMop.R
\name{summary.PMop}
\alias{summary.PMop}
\title{Summarize Observations and Predictions}
\usage{
\method{summary}{PMop}(
  object,
  ...,
  digits = max(3, getOption("digits") - 3),
  pred.type = "post",
  icen = "median",
  outeq = 1
)
}
\arguments{
\item{object}{A PMop object made by \code{\link{makeOP}}.}

\item{\dots}{Other parameters which can be passed to \code{summary}.}

\item{digits}{Integer, used for number of digits to print.}

\item{pred.type}{Either 'post' for a posterior object or 'pop' for a population object.  Default is 'post'.}

\item{icen}{Can be either "median" for the predictions based on medians of \code{pred.type} parameter value
distributions, or "mean".  Default is "median".}

\item{outeq}{Output equation number.  Default is 1.}
}
\value{
A list with two xs.  The first component of the list is a
matrix with the minimum, first quartile, median, third quartile, maximum,
mean and standard deviation for times, observations and predictions in \code{x}.
The second contains the mean prediction error,
the mean weighted prediction error (bias), the mean squared prediction error, root mean sqaured error (RMSE),
percent root mean squared error (%RMSE), the mean weighted
squared prediction error, the bias-adjusted mean squared prediction error, and the bias-
adjusted mean weighted squared prediction error (imprecision).
}
\description{
Summarize a Pmetrics Observed vs. Predicted x
}
\details{
Summarize observations, predictions and errors in a PMop x made by \code{\link{makeOP}}.
}
\seealso{
\code{\link{makeOP}}
}
\author{
Michael Neely
}
