% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPMpta.R
\name{summary.PMpta}
\alias{summary.PMpta}
\title{Summarize Percent Target Attainment}
\usage{
\method{summary}{PMpta}(object, at = 1, ci = 0.95, ...)
}
\arguments{
\item{object}{A PMpta object made by \link{makePTA}.}

\item{at}{Which object in the \emph{PM_pta} result list to plot. By default 1.}

\item{ci}{Width of the interval for pharmacodynamic index reporting.  Default is 0.95, i.e. 2.5th to 97.5th percentile.}

\item{...}{Not used.}
}
\value{
A tibble with the following columns (only the first three if \code{at = "intersect"}):
\itemize{
\item \strong{sim_num} is the number of the simulation
\item \strong{target} is the target for the row, if targets are discrete, not used for simulated targets
\item \strong{label} is the simulation label, for reference
\item \strong{prop_success} is the proportion of simulated profiles that met the success definition
\item \strong{median} is the median parmacodynamic index (PDI), i.e. the proportion or ratio depending on the target type
\item \strong{lower} is the lower bound of the interval defined by \code{ci}
\item \strong{upper} is the upper bound of the interval defined by \code{ci}
\item \strong{mean} is the mean of the PDI
\item \strong{sd} is the standard deviation of the PDI
\item \strong{min} is the minimum PDI
\item \strong{max} is the maximum PDI
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize a Pmetrics Percent Target Attainment Object
}
\details{
Summarize Pharmacodynamic Index (PDI) statistics and success proportions in a \emph{PMpta} object made by \link{makePTA}
and usually found in a \link{PM_pta} object. The PDI is the metric calculated by the target type and target, e.g. AUC/Target,
or \%time>target. Since a PDI cannot be calculated for intersections, summarizing the intersection object only
provides the success proportion per simulation/target.
}
\seealso{
\link{makePTA}, \link{PM_pta}
}
\author{
Michael Neely
}
