% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.PMpta}
\alias{summary.PMpta}
\title{Summarize Percent Target Attainment}
\usage{
\method{summary}{PMpta}(object, ...)
}
\arguments{
\item{object}{A PMpta object made by \code{\link{makePTA}}.}

\item{\dots}{Other parameters which can be passed to \code{summary}.}
}
\value{
A data frame with the following columns: simnum, target, success, meanratio, and sdratio.
\emph{simnum} is the number of the simulation; \emph{target} is the specified target;
\emph{success}  has the proportion with a ratio > \code{success}; \emph{meanratio} and \emph{sdratio}
are the mean and standard deviation of the target ratios for each simulation and target.
}
\description{
Summarize a Pmetrics Percent Target Attainment Object
}
\details{
Summarize target statistics and success proportions in a PMpta object made by \code{\link{makePTA}}.
}
\author{
Michael Neely
}
\seealso{
\code{\link{makePTA}}
}

