% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lib.R
\name{two_comp_bolus_cl}
\alias{two_comp_bolus_cl}
\title{Two-compartment bolus clearance model template}
\usage{
two_comp_bolus_cl()
}
\value{
A \code{PM_lib} object.
}
\description{
Create a \code{PM_lib} object for the \code{two_comp_bolus_cl} model template.
}
