% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgrowth.R
\name{zBMI}
\alias{zBMI}
\title{Extract CDC pediatric BMI z-scores}
\usage{
zBMI(agemos, sex, bmi, wt, ht, data = "CDC")
}
\arguments{
\item{agemos}{The age in months.  Should be between 24 and 240.5.}

\item{sex}{A single quoted character: "M" for males, "F" for females.  Default is "M".}

\item{bmi}{The individual's BMI. If specified, \code{wt} and \code{ht} are not necessary
and will be ignored.}

\item{wt}{The individual's weight in kg as an alternative to specifying \code{BMI}.
Will be ignored if \code{BMI} is specified.}

\item{ht}{The individual's height in centimeters. Required if \code{wt} is specified
and \code{BMI} is not. Ignored if \code{BMI} is specified.}

\item{data}{Source data for calculations. Default is "CDC" which uses the \link{cdc_bmi}
dataset. The alternative is "NHANES", which uses the \link{ger_bmi} dataset.}
}
\value{
A list with objects calculated for \code{agemos} and \code{sex}.
\itemize{
\item z Z-score
\item mod_z Modified Z-score for extreme BMI
\item per BMI percentile
\item mod_per Modified BMI percentile
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Will extract BMI z-scores based on a single age in months and sex. Overweight is
a z-score of >1.04, and obese is a z-score > 1.64. Calculations are based on
\href{https://www.cdc.gov/nccdphp/dnpa/growthcharts/resources/biv-cutoffs.pdf}{CDC formulae}

For a z-score > 3, which indicates an extreme BMI, consider using the modified
z-score and percentile.
}
\examples{
\dontrun{
zBMI(agemos = 36, bmi = 15)
}
}
\author{
Michael Neely
}
